\name{adm}
\alias{adm}

\title{Admissions and Test Scores: ADM}

\description{
This file contains information about the undergraduate selection process for entering first-time, degree/certificate-seeking students. This includes information about admission considerations,  applicants, applicants that were admitted, and admitted students who enrolled.
}

\section{SAT and ACT test}{
SAT and ACT test scores are included for institutions, that require test scores for admission. These data are applicable for institutions that do not have an open admissions policy for entering first-time students.
Beginning with 2016, writing scores for both SAT and ACT are no longer collected.
}

\section{Admission Consideration}{
The possible values for the admission consideration variables have changed from data prior to 2016.  "Do not know" is no longer an option and "considered but not required" was added.  See frequencies for complete listing.
}

