\name{al}
\alias{al}

\title{Academic Library: AL}

\description{
The Academic Library survey became part of the Integrated Postsecondary Education Data system in collection year 2014-15. Data include characteristics of the library, collections, expenditures, and services.
}

\section{Data Availability}{
These data are only applicable for degree-granting institutions.  Degree-granting institutions with total expenditures over $100,000 dollars will have expenditure data.
In 2016-17 the types of collections was expanded to collect information on physical and electronic serials.
}

\section{Data Releases}{Preliminary release data have been edited but are subject to further NCES quality control procedures. Imputed data for nonresponding institutions are not included. These data are used for First Look (Preliminary Data) publications. Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions. These data are used for First Look (Provisional Data) publications. Final release data include revisions to the provisional release data that have been made by institutions during the subsequent data collection year. The final release data can be used when the most up to date data are required; however, these data may not match tables from the First Look reports based on preliminary and provisional data.}

\section{Glossary}{
  \describe{
    \item{Total library circulations}{
    The number of times book and media units are "checked out" from the general and reserve collections. Includes intitial transactions and renewals.
    }
    \item{Ongoing commitments to subscriptions}{
    The expenditures for ongoing subscriptions to serials in all formats, including duplicates, for all outlets. These are publications issued in successive parts, usually at regular intervals, and, as a rule, intended to be continued indefinitely.
    Serial subscriptions include periodicals, newspapers, annuals (reports, yearbooks, etc.), memoirs, proceedings, and transactions of societies. Includes the costs of electronic serials bought in aggregations and serial packages.
    }
  }
}
