\name{c_a}
\alias{c_a}

\title{Completions: C_A}

\description{
This file contains the number of awards by type of program, level of award (certificate or degree), first or second major and by race/ethnicity and gender.
}

\section{Reporting}{
Data covers all awards granted between July 1 of previous year and June 30 of current year.
}

\section{Type of Instructional Program}{
Type of program is categorized according to the Classification of Instructional Programs (CIP), a detailed coding system for postsecondary instructional programs.
}

\section{Data Layout}{
This file contains multiple records per institution. Each record is uniquely defined by the variables IPEDS ID (UNITID), classification of instructional program (CIPCODE), first or second major (MAJORNUM) and award level (AWLEVEL).  Each record will contain the total awards,  awards for men and women, and the total awards and awards for men and women for all nine race/ethnicity categories.
}


\section{Glossary}{
  \describe{
  \item{Classification of instrucional Program (CIP) code}{A six-digit code in the form xx.xxxx that identifies instructional  program specialties within educational institutions.}
  }
}
