\name{c_b}
\alias{c_b}

\title{Completions: C_B}

\description{
This file contains the number of students who completed any degree or certificate by race/ethnicity and gender.
}

\section{Reporting}{
Data covers all awards granted between July 1 of previous year and June 30 of current year.
}

\section{Data Layout}{
This file contains one record per institution. Each record is uniquely defined by the variables IPEDS ID (UNITID).  Each record will contain the total number of students,  men and women, and the total number of students for men and women for all nine race/ethnicity categories.
}


