\name{efa}
\alias{efa}

\title{Fall Enrollment: EFA}

\description{
This file contains the number of students enrolled in the fall, by race/ethnicity, gender, attendance (full- or part-time) status and level of student.
}

\section{Reporting}{
Institutions with traditional academic year calendar systems (semester, quarter, trimester or  4-1-4) report their enrollment as of October 15 or the official fall reporting date of the institution.

Institutions with calendar systems that differ by program or allow continuous enrollment report students that are enrolled at any time between August 1 and October 31.
}

\section{Data Layout}{
This file has multiple records per institution. Records are uniquely defined by the variables IPEDS ID (UNITID), and the attendance status and level of student (EFALEVEL). Each record will contain the total enrollment, enrollment for men and women, and the total enrollment and enrollment for men and women for all nine race/ethnicity categories.
}


\section{Glossary}{
\describe{

\item{Attendance status of student [SECTION]}{Full and part-time status of student}

\item{Credit}{Recognition of attendance or performance in an instructional activity (course or program) that can be applied by a recipient toward the requirements for a degree, diploma, certificate, or other formal award.}

\item{Degree/certificate-seeking students}{Students enrolled in courses for credit and recognized by the institution as seeking a degree, certificate, or other formal award. High school students also enrolled in postsecondary courses for credit are not considered degree/certificate-seeking.  }

\item{Doctor's degree}{Professional practice - as defined by the institution.  }

\item{First-time student (undergraduate)}{A student who has no prior postsecondary experience (except as noted below) attending any institution for the first time at the undergraduate level. This includes students enrolled in academic or occupational programs. It also includes students enrolled in the fall term who attended college for the first time in the prior summer term, and students who entered with advanced standing (college credits earned before graduation from high school).  }

\item{Full-time student Graduate}{A student enrolled for 9 or more semester credits, or 9 or more quarter credits, or a student involved in thesis or dissertation preparation that is considered full time by the institution.}

\item{Full-time student Undergraduate}{A student enrolled for 12 or more semester credits , or 12 or more quarter credits, or 24 or more contact hours a week each term.}

\item{Graduate student}{A student who holds a bachelor's degree or above and is taking courses at the postbaccalaureate level. These students may or may not be enrolled in graduate programs.  }

\item{Level of student [EFALEVEL]}{Level,  full- and part-time status, degree-seeking/non degree-seeking status  and year of study of student

This variable identifies the level of enrollment data for the institution.  Enrollment counts are available by level of student (undergraduate or graduate).  Undergraduate enrollments are disaggregated by degree-seeking and non-degree seeking status. Degree-seeking enrollments are further broken out by first-time, first year students, transfer-ins and continuing students.  All of the above categories are available by full- and part-time status.}

\item{Part-time student Graduate}{A student enrolled for less than 9 semester or quarter credits.}

\item{Part-time student Undergraduate}{A student enrolled for either less than 12 semester or quarter credits, or less than 24 contact hours a week each term.}

\item{Transfer-in student (undergraduate) }{A student entering the reporting institution for the first time but known to have previously attended a postsecondary institution at the same level. The student may transfer with or without credit.  }

\item{Undergraduate}{A student enrolled in a 4- or 5-year bachelor's degree program, an associate's degree program, or a vocational or technical program below the baccalaureate.  }
  }
}
