\name{efc}
\alias{efc}

\title{Fall Enrollment: EFC}

\description{
This file contains the number of first-time freshmen by state of residence, along with data on the number who graduated from high school the previous year.
}

\section{Reporting}{
 Institutions with traditional academic year calendar systems (semester, quarter, trimester or  4-1-4) report their enrollment as of October 15 or the official fall reporting date of the institution.

 Institutions with calendar systems that differ by program or allow continuous enrollment report students that are enrolled at any time between August 1 and October 31.
}

\section{Data Layout}{
This file has multiple records per institution.  Records are uniquely defined by the variables IPEDS ID (UNITID), and the state of residence (EFCSTATE).
}

\section{Optional Data: Enrollment by Residency}{
Submission of enrollment of first-time undergraduate students by residency is mandatory in even-numbered years and optional in odd-numbered years. Do not use these data in odd-numbered years to produce national or state estimates as data will not be available from all institutions. Data in optional years are provided to allow users to compare institutional level data among reporting institutions.
}


\section{Glossary}{
\describe{

\item{Degree/Certificate-seeking students}{Students enrolled in courses for credit who are recognized by the institution as seeking a degree or other formal award. At the undergraduate level, this is intended to include students enrolled in vocational or occupational programs.}

\item{First-time student(Undergraduate)}{A student attending any institution for the first time at the undergraduate level. Includes students enrolled in academic or occupational programs. Also includes students enrolled in the fall term who attended college for the first time in the prior summer term, and students who entered with advanced standing (college credits earned before graduation from high school). }

\item{State of residence}{A person’s permanent address as determined by such evidence as a driver’s license or voter registration. For entering freshmen, state of residence may be the legal state of residence of a parent or guardian.}

\item{Undergraduate}{A student enrolled in a 4- or 5-year bachelor's degree program, an associate's degree program, or a vocational or technical program below the baccalaureate. }
  }
}
