\name{efcp}
\alias{efcp}

\title{Fall Enrollment: EFCP}

\description{
This file contains the number of students enrolled in the fall, by race/ethnicity, gender, attendance (full- or part-time) status and level of student for selected major fields of study.}

\section{Major Field of Study}{
Major field of studies included are:
 \itemize{
    \item 13.0000 Education
    \item 14.0000 Engineering
    \item 22.0101 Law (LL.B., J.D.)
    \item 26.0000 Biological Sciences/Life Sciences
    \item 27.0000 Mathematics
    \item 40.0000 Physical Sciences
    \item 51.0401 Dentistry (D.D.S., D.M.D.)
    \item 51.1201 Medicine (M.D.)
    \item 52.0000 Business Management & Administrative Services
  }

These are the only fields of study required from 4-year institutions for the Office for Civil Rights.
}

\section{Reporting}{
 This data file contains enrollment data for 4-year institutions only.

 Institutions with traditional academic year calendar systems (semester, quarter, trimester or  4-1-4) report their enrollment as of October 15 or the official fall reporting date of the institution.

 Institutions with calendar systems that differ by program or allow continuous enrollment report students that are enrolled at any time between August 1 and October 31.
}

\section{Data Layout}{
This file has multiple records per institution.  Records are uniquely defined by the variables IPEDS ID (UNITID), and the major field of study, attendance status and level of student (EFCIPLEV).   Each record will contain the total enrollment,  enrollment for men and women, and the total enrollment and enrollment for men and women for all nine race/ethnicity categories.
}

\section{Data Releases}{
Preliminary release data have been edited but are subject to further NCES quality control procedures. Imputed data for nonresponding institutions are not included. These data are used for First Look (Preliminary Data) publications.

Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions. These data are used for First Look (Provisional Data) publications.

Final release data include revisions to the provisional release data that have been made by institutions during the subsequent data collection year. The final release data can be used when the most up to date data are required; however, these data may not match tables from the First Look reports based on preliminary and provisional data.
}

\section{Glossary}{
\describe{

\item{Attendance status of student [SECTION]}{Defines full and part-time status of student}

\item{Classification of instrucional Program (CIP) code}{A six-digit code in the form xx.xxxx that identifies instructional  program specialties within educational institutions.}

\item{Credit}{Recognition of attendance or performance in an instructional activity (course or program) that can be applied by a recipient toward the requirements for a degree, diploma, certificate, or other formal award.}
  }
}
