\name{efd}
\alias{efd}

\title{Fall Enrollment: EFD}

\description{
This file contains data on the total entering class, first-year retention rates
and the student-to-faculty ratio.
}
\section{Total Entering Class}{
The total entering class is the number of
incoming students (students enrolling for the first time in a postsecondary
institution versus students transferring in from another postsecondary
institution) at an institution.

This measure permits the calculation of the
graduation rate cohort as a proportion of the total entering student body.
}
\section{Retention Rate}{
The
first-year retention rate data measure the percentage of first-year students
who had persisted in or completed their educational program a year later (began
in fall of the previous year and persisted or completed in fall of the current year).
}

\section{Exclusions}{
Institutions can exclude the number of students from the first-time cohort who left the institution for
any of the following reasons:
  \itemize{
    \item died or were totally and permanently disabled;
    \item to serve in the armed forces (including those called to active duty);
    \item to serve with a foreign aid service of the Federal Government (e.g. Peace Corps);
    \item to serve on official church missions.
  }
}


\section{Student-to-Faculty Ratio}{
  Student-to-faculty ratio is defined as total
  FTE students (not in graduate or professional programs) divided by total
  FTE instructional staff (not teaching in graduate or professional programs).

}

\section{Reporting}{
All data on this file is applicable only to institutions with undergraduate
  students.

  Institutions with traditional academic year calendar systems
  (semester, quarter, trimester or 4-1-4) report their enrollment as of October
  15 or the official fall reporting date of the institution (academic year).
  Institutions with calendar systems that differ by program or allow continuous
  enrollment report students that are enrolled at any time between August 1 and
  October 31 (program year). Only academic year institutions report entering class
  data. Institutions with short programs (less-than-one-year) will not have data
  on retention.

  Beginning with the fall 2016 data, instititions
can include first-time bachelor’s-seeking (or equivalent) study abroad students
who were excluded from the first-time cohort but who have re-enrolled at the
institution their second year.
}

\section{Data Releases}{
Preliminary release data have been edited but are subject to further NCES quality control procedures. Imputed data for nonresponding institutions are not included. These data are used for First Look (Preliminary Data) publications.

Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions. These data are used for First Look (Provisional Data) publications.

Final release data include revisions to the provisional release data that have been made by institutions during the subsequent data collection year. The final release data can be used when the most up to date data are required; however, these data may not match tables from the First Look reports based on preliminary and provisional data.
}

\section{Glossary}{
\describe{
  \item{First-time student (undergraduate)}{
  A student who has no prior postsecondary experience (except as noted below) attending any institution for the first time at the undergraduate level. This includes students enrolled in academic or occupational programs. It also includes students enrolled in the fall term who attended college for the first time in the prior summer term, and students who entered with advanced standing (college credits earned before graduation from high school).}
  \item{Total entering students at the undergraduate level}{
  In addition to the students in the Graduation Rate Survey (GRS) cohort, this would include:
    \itemize{
    \item part-time undergraduate students;
    \item non-degree/certificate-seeking undergraduates;
    \item students who initially attended the prior summer term and returned again in the fall
(other than those included in the GRS cohort);
    \item students transferring into the institution at any undergraduate level for the first time.}
  }
  \item{Full-time cohort}{
   Includes full-time first-time degree/certificate seeking students enrolled at the institution in the fall of the prior year (including those enrolled for the first time the preceding summer term and those whose intent was not known upon entry to the institution).  The full-time cohort for 4-year institutions should only include bachelor degree-seeking students.

   The student status (full- or part-time) is based on the prior year's fall status, even if this has changed by the current year (e.g., if a student has gone from part-time to full-time, they will be reported as part of the part-time cohort).}
   \item{Part-time cohort}{
   Includes part-time first-time degree/certificate seeking students enrolled at the institution in the fall of the prior year (including those enrolled for the first time the preceding summer term and those whose intent was not known upon entry to the institution).  The part-time cohort for 4-year institutions should only include bachelor degree-seeking students.

   The student status (full- or part-time) should be based on the prior year's fall status, even if this has changed by the current year (e.g., if a student has gone from part-time to full-time, report them as part of the part-time cohort).}
  \item{Full-time adjusted cohort from prior year}{
  Full-time cohort from prior year minus number of exclusions plus the number of Inclusions.}
  \item{Exclusions}{Students from the cohort or subcohort who left the institution for any of the following reasons: died or were totally and permanently disabled; to serve in the armed forces; to serve with a foreign aid service of the Federal Government, such as the Peace Corps; or to serve on official church missions.}
  \item{Inclusions}{The number of first-time degree seeking  study abroad students who were excluded from the prior year first-time full-time cohort but who have re-enrolled at the institution their second year.}
  \item{Retention rate}{The percent of the (fall cohort from the prior year minus exclusions from the fall cohort) that re-enrolled at the institution as either full- or part-time in the current year}
  \item{Student-to-faculty ratio}{ Total FTE students not in graduate or professional programs divided by total FTE instructional staff not teaching in graduate or professional programs.}
  \item{Total FTE students}{Equal to the number of full-time students plus 1/3 the number of part-time students (Fall enrollment component).}
  \item{Total FTE instructional staff}{Equal to the number of full-time instructional staff to 1/3 the number of part-time instructional staff (Human Resource Component, EAP section).}
  \item{Instructional staff}{Include employees whose primary function/occupational activity is primarily instruction or instruction/research/public service and are not medical school employees.}
  \item{Graduate or first-professional students}{Students enrolled in graduate or professional programs such as medicine, law, veterinary, dentistry, social work, or public health, in which faculty teach virtually only graduate-level students (often referred to as "stand-alone" or "independent" programs) are excluded from both full-time and part-time counts.}
  }
}


\section{More Info}{
\href{https://nces.ed.gov/ipeds/use-the-data/survey-components/8/fall-enrollment}{IPEDS Fall Enrollment Survey Components}
}
