\name{effy}
\alias{effy}

\title{12-Month Enrollment: EFFY}

\description{
This file contains the unduplicated head count of students enrolled over a 12-month period for both undergraduate and graduate levels.

These enrollment data are particularly valuable for institutions that use non-traditional calendar systems and offer short-term programs. Because this enrollment measure encompasses an entire year, it provides a more complete picture of the number of students these schools serve.
}

\section{Data Layout}{
 Each record is uniquely defined by the variables IPEDS ID (UNITID), and the level of enrollment (EFFYLEV). Each record will contain the total head count for men and women, and the total headcount for men and women for all nine race/ethnicity categories.
}

\section{Glossary}{
\describe{
  \item{Level of study}{
  \itemize{
  \item 1 - Total students enrolled for credit
  \item 2 - Undergraduate students - A student enrolled in a 4- or 5-year bachelor's degree program, an associate's degree program, or a vocational or technical program below the baccalaureate. Students who have already earned a bachelor's degree but are taking undergraduate courses FOR CREDIT are included as undergraduates.
  \item 4 - Graduate student -  A student who holds a bachelor's degree or above and is taking courses at the postbaccalaureate level. These students may or may not be enrolled in graduate programs.
  }
  }
  \item{Original level of study on survey form}{
  \itemize{
  \item 1 - Undergraduate students - A student enrolled in a 4- or 5-year bachelor's degree program, an associate's degree program, or a vocational or technical program below the baccalaureate. Students who have already earned a bachelor's degree but are taking undergraduate courses FOR CREDIT are included as undergraduates.
  \item 3 - Graduate student -  A student who holds a bachelor's degree or above and is taking courses at the postbaccalaureate level. These students may or may not be enrolled in graduate programs.
  }
  }
  }
}
