\name{efia}
\alias{efia}

\title{12-Month Enrollment: EFIA}

\description{
This table contains data on instructional activity measured in total credit and/or contact hours delivered by institutions during a 12-month period.
}

\section{Full-time Equivalent (FTE)}{
The credit hour and contact hour activity data are used to derive 12-month full-time equivalent (FTE) enrollments for both undergraduate and graduate levels. The graduate level does not include credit hours for doctoral professional practice students.

Institutions can choose to accept the derived FTE  or report their own FTE.  Both reported and estimated/derived FTE are available in this data table.

In addition, the reported FTE of Doctoral Professional practice students are also included.
}


\section{Glossary}{
\describe{
  \item{Contact Hour Activity}{
  The provision of coursework to students which  can be measured in terms of contact or clock hours. For the purposes of this survey, total contact hour activity was determined by  multiplying the contact hour value of the program by the number of  students enrolled in the program. For example, a 600 hour program with  6 students enrolled would contribute 3600 contact hours to the total  12-month activity as long as all 3600 hours were completed during the  timeframe. However,if 3 students completed their program but the remaining 3 only completed half the program, the institution should report 1,800 hours for the first three students and 900 hours for the  second group; or a total of 2,700 hours toward the 12-month total.}

  \item{Credit Hour Activity}{
  The provision of coursework to students which  can be measured in terms of credit hours. For the purposes of this survey, total credit hour activity was determined by multiplying the  credit hour value of the course by the number of students enrolled in the course for credit. For example, the credit hour activity for a  3-credit course with an enrollment of 30 students is 90 credit hours.

Using this as a guide, institutions were asked to compute the credit  hours attempted for each course and sum the activity for all credit hour courses over the total 12-month period. Enrollment is based on  the number of students enrolled at the close of the official drop/add period or other census date of the institution.}

  \item{Doctor's degree - professional practice}{
  A doctor's degree that is conferred upon completion of a program providing the knowledge and skills for the recognition, credential, or license required for professional practice. The degree is awarded after a period of study such that the total time to the degree, including both pre-professional and professional preparation, equals at least six full-time equivalent academic years. Some of these degrees were formerly classified as first-professional and may include:

  Chiropractic (D.C. or D.C.M.); Dentistry (D.D.S. or D.M.D.); Law (L.L.B. or J.D.);
  Medicine (M.D.); Optometry (O.D.); Osteopathic Medicine (D.O); Pharmacy (Pharm.D.);
  Podiatry (D.P.M., Pod.D., D.P.); or, Veterinary Medicine (D.V.M.),
  and others, as designated by the awarding institution.}

  \item{Estimated full-time equivalent (FTE) graduate enrollment}{
  Generated as follows:
  \itemize{
  \item For institutions with a semester, trimester, or 4-1-4 plan, the number of FTE graduate students is the number of graduate credit hours divided by 24.
  \item For institutions with a quarter plan, graduate FTE is the is the number ofgraduate credit hours divided by 36.}}

  \item{Estimated full-time equivalent (FTE) undergraduate enrollment}{
  Generated as follows:
  \itemize{
  \item For institutions with a semester, trimester, or 4-1-4 plan, the number of FTE undergraduate is the sum of undergraduate credit hours divided by 30 and contact hours divided by 900.
  \item For institutions with a quarter plan, undergraduate undergraduate credit hours divided by 45 and contact hours divided by 900.
  \item For institutions with continuous enrollment over a 12-month period, undergraduate credit hours were divided by 30 and contact hours were divided by 900.}}

  \item{Instructional Activity}{
  The provision of coursework to students which  can be measured in various terms.}

  \item{Reported full-time equivalent (FTE) undergraduate/graduate enrollment}{
  NCES uses estimated FTE undergraduate/graduate enrollment to calculate expenses by function per FTE and core revenues per FTE as reported in the IPEDS Data Feedback Report.

  If the generated estimate was not reasonable, the institution provided their best estimate for undergraduate/graduate FTE.  If the institution did not provide an FTE, then the reported FTE was set to the estimated FTE.}
  }
}
