\name{f_f1a}
\alias{f_f1a}

\title{Finance: F_F1A}

\description{
This data file contains institutional finance data for public institutions that use accounting standards established by the Governmental Accounting Standards Board (GASB).

Finance data includes institutional revenues by source, expenditures by functional categories, expenditures by natural classification categories, endowments, and assets and liabilities.

This information provides context for understanding the cost and resources of providing postsecondary education. It is used to calculate the contribution of postsecondary education to the gross national product.
}

\section{Amendments}{
Beginning with fiscal year 2016 detailed expenditures by function and the natural classifications of fringe benefits, operations and maintenance of plant, depreciation and interest are no longer available.
}

\section{Release 2 - Update}{
Release 2 contains an update to other revenues and additions (F1B23) for nondegree-granting institutions. Total other revenues and additions (F1B24)  include capital appropriations (F1B20), capital grants and contracts (F1B21), additions to permanent endowments (F1B22), and other revenues and additions (F1B23).  Degree-granting institutions report all of the above components and nondegree-granting institutions only report the total (F1B24).  Release 2 copies the value of F1B24 to other revenues and additions (F1B23) for nondegree-granting institutions, so that the detail will sum to the total for all institutions.
}

\section{Data Releases}{
Preliminary release data have been edited but are subject to further NCES quality control procedures. Imputed data for nonresponding institutions are not included. These data are used for First Look (Preliminary Data) publications.

Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions. These data are used for First Look (Provisional Data) publications.

Final release data include revisions to the provisional release data that have been made by institutions during the subsequent data collection year. The final release data can be used when the most up to date data are required; however, these data may not match tables from the First Look reports based on preliminary and provisional data.
}

\section{Glossary}{
\describe{
  \item{Total current assets}{
  Includes cash and equivalents, investments, accounts and notes receivable (net of allowance for uncollectible amounts), inventories, and all other assets classified as current assets as discussed in Chapter 2 of Accounting Research Bulletin 43.

  Beginning with FY2013 (Fiscal Year 2013), GASB statements 63,65 required institutions to display deferred inflows and outflows of resources on their Statement of Net Position. Thus from FY2014 to FY2016, current assets included deferred outflows of resources. Beginning with FY2016, deferred outflows of resources was collected separately from current assets.}
  \item{Depreciable capital assets, net of depreciation}{
  The net amount of all depreciable capital assets after reducing the gross amount for accumulated depreciation.}
  \item{Capital assets}{
  Tangible or intangible assets that are capitalized under an institution's capitalization policy; some of these assets are subject to depreciation and some are not. These assets consist of land and land improvements, buildings, building improvements, machinery, equipment, infrastructure, and all other assets that are used in operations and that have initial useful lives extending beyond one year. Also includes collections of works of art and historical treasure and library collections; however under certain conditions such collections may not be capitalized. Includes property acquired under capital leases and intangible assets such as patents, copyrights, trademarks, goodwill, and software. Excluded are assets that are part of endowment funds or other capital fund investments in real estate period.}
  \item{Other noncurrent assets}{
  The assets that are not reasonably expected to be realized in cash or sold or consumed during the next normal operating cycle (normally  one year) of the institution other than depreciable capital assets included
above.}
  \item{Total noncurrent assets}{
  The assets that are not reasonably expected to be realized in cash or sold or consumed during the next normal operating cycle (normally one year) of the institution.}
  \item{Total assets}{
  The sum of current and noncurrent assets.}
  \item{Deferred outflows of resources }{
  A consumption of net assets by a government that is applicable to future periods. Examples of deferred outflows of resources include changes in fair values in hedging instruments and changes in the net pension liability that are not considered pension expense (as described in GASB Statement 68, Accounting and Financial Reporting for Pensions: an amendment of GASB Statement No. 27).}
  \item{Deferred inflows of resources}{
  Acquisition of net assets acquired by a government that is applicable to future reporting periods. Examples of deferred inflows are the difference in a debt refunding between reacquisition and net carrying amount of the old debt, upfront payments in service concession arrangements, and change in fair values in hedging instruments.  }
  \item{Long-term debt, current portion}{
  The amount of long-term debt that is expected to require current assets to pay or liquidate during the next year.}
  \item{Other current liabilities}{
  Liabilities whose liquidation is reasonably expected to require the use of resources classified as current assets or the  creation of other current liabilities within the next year and are not included in the category of long term debt, current portion.}
  \item{Total current liabilities}{Liabilities whose liquidation is reasonably expected to require the use of resources classified as current assets or the creation of other current liabilities within the next year such as accounts payable, salaries payable, and the current portion of long-term debt.}
  \item{Long-term debt}{
  Debt of the institution in the form of bonds, notes, capital  leases, and other forms of debt that are repayable over a period greater than one year. }
  \item{Other noncurrent liabilities}{
  Liabilities whose liquidation is not reasonably  expected to require the use of resources classified as current assets or the  creation of other current liabilities within the next year, other than long  term debt. This includes long-term accrued liabilities (such as for compensated absences, claims & judgments, and post-employment/post-retirement benefits).}
  \item{Total noncurrent liabilities}{
  Liabilities whose liquidation is not  reasonably expected to require the use of resources classified as current assets or the creation of other current liabilities within the next year. This  includes the noncurrent portion of long-term debt and long-term accrued  liabilities (such as for compensated absences, claims & judgments, and post- employment/post-retirement benefits). }
  \item{Total liabilities}{
  The sum of current and noncurrent liabilities.}
  \item{Net assets invested in capital assets, net of related debt}{
  Represented by  the total of all capital assets, reduced by accumulated depreciation, and reduced by the amount of outstanding bonds, mortgages, notes, or other borrowings that are attributable to the acquisition, construction, or  improvement of those assets.}
  \item{Restricted-expendable net assets}{
  Those that have constraints placed on use that are  either:

  (a) externally imposed by creditors, grantors, contributors, or laws and regulations of other governments or

  (b) imposed by law through constitutional provisions or enabling legislation, and are not required to be retained in perpetuity.}
  \item{Restricted-nonexpendable net assets}{
  Those that have constraints placed on use that are  either:

  (a) externally imposed by creditors, grantors, contributors, or laws and regulations of other governments or

  (b) imposed by law through constitutional provisions or enabling legislation, and are required to be retained  in perpetuity. }
  \item{Unrestricted net assets }{
  Net assets held by the institution upon which no restrictions have been placed by the donor or other party external to the  institution.  }
  \item{Net position}{
  The excess of assets over liabilities or the residual interest in the institution's assets remaining after liabilities are deducted. The change in net position results from revenues, gains, expenses, and losses.

  GASB institutions classify net position into three categories: invested in capital, net of related debt; restricted (with separate displays of restricted-expendable and restricted-nonexpendable net assets); and unrestricted.  Net position beginning with the 2013-14 collection includes deferred inflows and outflows of resources, per GASB 63,65. This term is similar to the "Net assets" term used by FASB institutions.
  }
  \item{Ending balance of land and land improvements}{
  The book value at the end of the fiscal year for capital assets consisting of land and improvements such as athletic fields, golf courses, lakes, etc. Land is nondepreciable; some land improvements are depreciable assets and some are nondepreciable.  }
  \item{Ending balance of infrastructure}{
  The book value at the end of the fiscal year for capital assets consisting of roads, bridges, drainage systems, water and sewer systems, and other similar assets. Infrastructure assets usually have longer useful lives than other capital assets such as buildings.}
  \item{Ending balance of buildings}{
  The book value at the end of the fiscal  year for capital assets built or acquired for occupancy and use by the entity. These are structures such as classrooms, research facilities, administrative offices, storage, etc. Includes built-in fixtures and equipment that are essentially part of the permanent structure. Buildings held for the production of revenue are not included here.}
  \item{Equipment, including art and library collections - Ending balance }{
  Captital assets that include moveable tangible property such as research equipment, vehicles, office equipment, library collections (capitalized amount of books, films, tapes, and other materials maintained in library collections intended for use by patrons), and capitalized art collections.  }
  \item{Ending balance of construction in progress}{
  The book value of assets that were under construction or development that have not yet been placed into service, such as a building or parking lot, as of the end of the fiscal year.  }
  \item{Ending balance for plant, property and equipment }{
   The book value at the end of the fiscal year of total assets for plant, property and equipment.  }
  \item{Plant, property and equipment}{
   Include: land and land improvements, infrastucture, buildings, equipment, art and library collections, property obtained under capital leases and construction in progress.  }
  \item{Ending balance of accumulated depreciation }{
  Total depreciation charged as expenses to date (in the current year and in prior years) on the capital assets of the institution.}
  \item{Intangible assets, net of accumulated amortization }{
  All capital assets consisting of certain nonmaterial rights and benefits of an institution, such as patents, copyrights, trademarks and goodwill. This amount should be reduced by total accumulated amortization.  }
  \item{Other capital assets}{
  All other amounts for capital assets not classified as:
 Land & land improvements,
 Infrastructure, Buildings,
 Equipment, Construction in progress,
 Accumulated depreciation, and Intangible assets, net of accumulated amortization.  }
  \item{Total revenues }{
  Other additions represents the sum of all revenues and
other additions to net assets, including operating and nonoperating revenues
and other revenues and additions.  }
  \item{Total expenses }{
  Other deductions represents the sum of operating and non-
operating expenses and deductions.  }
  \item{ Change in net position during year}{
  The net difference between total revenues
and other additions and total expenses and other deductions.  }
  \item{ Net position beginning of year}{
  The value of net position as of the beginning of  the fiscal year.}
  \item{Adjustments to beginning net posiiton }{
  Unusual and infrequent adjustments that
are not recorded as current year revenues, expenses, gains, or losses. This
includes adjustments for retroactive applications of changes in accounting
principles and prior period adjustments. This line on the IPEDS survey is an
automatic balancing line.  }
  \item{ Net position end of year}{
  The sum of net assets invested in capital assts,
net of related debt, restricted-expendable net assets, restricted-nonexpendable
net assets, and unrestricted net assets.  It can be calculated as the
difference between total assets and total liabilities.}
  \item{Tuition and fees }{
  Revenues from all tuition and fees assessed against
students (net of refunds and discounts & allowances) for educational purposes.
If tuition or fees are remitted to the state as an offset to the state
appropriation, the total of such tuition or fees should be deducted from the
total state appropriation and added to the total for tuition and fees.

If an all-inclusive charge is made for tuition, board, room, and other services, a
reasonable distribution is made between revenues for tuition and fees and revenues for auxiliary enterprises. Tuition and fees excludes charges for room,board, and other services rendered by auxiliary enterprises. }
  \item{Federal/State operating grants and contracts }{
  Revenues from federal/state government
agencies that are for specific research projects or other types of programs
and that are classified as operating revenues.  Examples are research projects
and similar activities for which amounts are received or expenditures are
reimbursable under the terms of a grant or contract.  }
  \item{Local/private operating grants and contracts }{
   Revenues from local
governmental and nongovernmental agencies and organizations that are for
specific research projects or other types of programs and that are classified
as operating revenues.  Examples are research projects and similar activities
for which amounts are received or expenditures are reimbursable under the terms
of a grant or contract.
  }
  \item{Sales and services of auxiliary enterprises, after deducting discounts and allowances}{
  Revenues (net of discounts and allowances such as scholarships)
generated by auxiliary enterprises that exist to furnish a service to students,
faculty, or staff, and that charge a fee that is directly related to the
cost of the service.  Examples are residence halls, food services, student
health services, intercollegiate athletics, college unions, college stores,
and movie theaters.
  }
  \item{Sales and services of hospitals, after deducting patient contractual
allowances }{
  Include operating revenues (net of patient contractual allowances)
for a hospital operated by the institution and clinics associated with
training.  It excludes clinics that are part of a student health services
program that should be reported elsewhere.}
  \item{Sales and services of educational activities (revenues) }{
  Revenues from the sales of goods or services that are incidental to the conduct of instruction, research or public service. Examples include film rentals, sales of scientific and literary publications, testing services, university presses, dairy products, machine shop products, data processing services, cosmetology services, and sales of handcrafts prepared in classes.  }
  \item{Independent operations }{
  Include all operating revenues associated with
operations independent of the primary missions of the institution.  This
category generally includes only those revenues associated with major
federally funded research and development centers.  It does not include
net profit (or loss) from operations owned and managed as investments of the
institution's endowment funds.}
  \item{Other sources - operating }{
  Includes any other operating revenues that are not
included in previously listed sources, namely, tuition and fees, grants and
contracts, sales and services of auxiliary enterprises and hospitals, and
independent operations.}
  \item{ Total operating revenues}{
   The sum of all operating revenues.  They result
from providing services and producing and delivering goods. }
  \item{ Federal appropriations}{
  Amounts received by the institution through acts of
a federal legislative body, except grants and contracts.  Funds reported in
this category are for meeting current operating expenses, not for specific
projects or programs.  An example is federal land-grant appropriations.
However land grant appropriations may be included in operating revenues if
accounted for by the institution as operating revenue.}
  \item{ State appropriations}{
  Amounts received by the institution through acts of a
state legislative body, except grants and contracts and capital appropriations.
Funds reported in this category are for meeting current operating expenses,
not for specific projects or programs.}
  \item{ Local appropriations, education district taxes and similar support}{
  Amounts received from property or other taxes assessed directly by or for an
institution below the state level.  It includes any other similar general
support provided to the institution from governments below the state level,
including local government appropriations.}
  \item{Federal non-operating grants }{
  Amounts reported as nonoperating revenues from federal government agencies that are provided on a non-exchange basis.Pell grants are included here, however, revenues from the Federal Direct Student Loan Program are not. Capital grants and gifts are also not included here.}
  \item{State non-operating grants}{
  Amounts reported as non-operating revenues from state governmental agencies that are provided on a non-exchange basis. This excludes capital grants and gifts.}
  \item{Local non-operating grants }{
  Amounts reported as non-operating revenues from local governmental agencies that are provided on a non-exchange basis. This excludes capital grants and gifts.}
  \item{Gifts, including contributions from affiliated organizations}{
  Revenues from private donors for which no legal consideration is provided. It includes all gifts or contributions to the institution except those classified as additions to permanent endowments or capital grants and gifts.}
  \item{Investment income }{
  Revenues derived from the institution's investments, including investments of endowment funds. Such income may take the form of interest income, dividend income, rental income or royalty income and includes both realized and unrealized gains and losses.}
  \item{Other non-operating revenues }{
  Revenues generated from non-exchange  transactions not reported as appropriations, grants, gifts, or investment   income.}
  \item{ Total non-operating revenues}{
  Represents the sum of all revenues generated from non-exchange transactions. }
  \item{Capital appropriations }{
  Amounts provided by government appropriations intended primarily for acquisition or construction of capital assets for the  institution.}
  \item{Capital grants and gifts}{
  Amounts received from gifts or grants primarily  intended for the acquisition or construction of capital assets for the institution.}
  \item{Additions to permanent endowments }{
  Gifts or other additions to endowments that are permanently nonexpendable.}
  \item{Other revenues and additions}{
  Increases to net assets that are not included  in other categories.}
  \item{Total other revenues and additions}{
  The sum of capital grants, gifts, and appropriations, additions to permanent endowments and other revenues   not otherwise classified. }
  \item{Total all revenues and other additions}{
  The sum of all revenues   and other additions to net assets.}
  \item{ }{ }
  \item{ }{ }
  \item{ }{ }
  \item{ }{ }
  \item{ }{ }
  \item{ }{ }
  \item{ }{ }
  \item{ }{ }.
  }
}
