\name{f_f1a}
\alias{f_f1a}

\title{Finance: F_F1A}

\description{
This data file contains the assets, liabilities, and net assets for component units of an institution.  These component units use the FASB accounting standard.

This file includes the name or names of the entities whose data is reported and the primary nature (general nature) of each component unit reported such as "fund raising/investment", "research", etc.
}

\section{Component Unit}{
 The term component units applies to GASB institutions only. A component unit is a legally separate organization for which the governing board and/or management of the primary institution is financially accountable. It can be another organization for which the nature and significance of its relationship with a primary institution is such that exclusion would cause the primary institution's financial statements to be misleading or incomplete.
}

\section{Data Availabilty}{
Financial data for component units are available for fiscal years 2002 through 2007.
}

\section{Data Releases}{
Preliminary release data have been edited but are subject to further NCES quality control procedures. Imputed data for nonresponding institutions are not included. These data are used for First Look (Preliminary Data) publications.

Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions. These data are used for First Look (Provisional Data) publications.

Final release data include revisions to the provisional release data that have been made by institutions during the subsequent data collection year. The final release data can be used when the most up to date data are required; however, these data may not match tables from the First Look reports based on preliminary and provisional data.
}

\section{Glossary}{
\describe{
  \item{Form ID }{
  A numeric indicator of which Form F component unit is being reported for this record. }
  \item{Entities included }{
  Entities indicates what component units are included in this record.  This could be a single entity or a combination of entities with similar purpose. It may also be a combination of smaller component units.}
  \item{Primary nature of unit(s) (PNATURE) }{
   Indicates the primary purpose or nature of the component unit(s) reported in this record, such as fundraising or athletic program support.}
  \item{ Long term investments}{
  Money or capital of the component unit that is invested for purposes of receiving a profitable return over a period of time of more than one year.}
  \item{Other assets}{
  Physical items (tangible) or rights (intangible) other than long term investments included in F1AF01 that have value and that are owned by the component unit.}
  \item{Total assets }{
  The sum value of all physical items (tangible) or rights (intangible) that have value and that are owned by the component unit.  }
  \item{ Total liabilities}{
  The sum of Debts and obligations of the component unit owed to outsiders or claims or rights, expressed in monetary terms, of a component unit's editors. }
  \item{Net assets - Temporarily restricted}{
  Net assets of FASB organizations whose use by the organization has been limited by donor specification as to use or the time when use may occur (such as a later period of time or after specified events have occurred).}
  \item{Net assets - Permanently restricted}{
  Net assets of FASB organizations that must be maintained in perpetuity. Permanently restricted net assets increase when organizations receive contributions for which donor-imposed restrictions limiting the instituiton's use of an asset or its economic benefits neither expire with the passage of time nor can be removed by the organization's meeting certain requirements.}
  \item{Net assets - Unrestricted}{
  Net assets of FASB organizations upon which no restrictions have been placed by the donor or other party external to the organization.}
  \item{Total net assets }{
  The sum of temporarily restricted, permanently restricted, and unrestricted net assets of the component unit organization.  It may be calculated as the difference between the assets and liabilities of the component unit.}
  \item{Investment return }{
  Income from assets including dividends, interest earnings, royalties, rent, gains (losses) etc. }
  \item{Other revenues, gains, and other support}{
  The inflow of resources or other enhancement of net assets of an organization or settlements of its liabilities (or a combination of both) from delivering or producing goods, rendering services, or other activities that constitute the organization's ongoing major or central operations, other than investment return. Also included are increases in the organization's net assets from peripheral or incidental transactions.
  }
  \item{Total revenues, gains, and other support }{
  The inflow of resources or other enhancement of net assets of an organization or settlements of its liabilities (or a combination of both) from delivering or producing goods, rendering services, or other activities that constitute the organization's ongoing major or central operations. Also included are creases in the organization's net assets from peripheral or incidental transactions. }
  \item{Total expenses }{
   The outflow or other using up of assets or incurrence of liabilities (or a combination of both) from delivering or producing goods,  rendering services, or carrying out other activities that constitute the insitution's ongoing major or central operations or in generating revenues.  This amount includes disbursements to the institution associated with the component unit. }
  \item{Expenses paid to institution}{
  The amount of total expenses that were disbursed from the component unit to the institution.}
  \item{Total losses }{
  Decreases in net assets from an organization's peripheral or incidental transactions and other events affecting the organization other than those that result from expenses.}
  \item{Change in net assets }{
  The term used to describe the net amount of revenues, expenses, gains, and losses for the reporting period.}
  \item{Net assets - Beginning of year}{
  The value of net assets as reported at the end of the previous fiscal year. }
  \item{ Adjustments to beginning net assets}{
  Unusual and infrequent adjustments that are not recorded as current year revenues, expenses, gains, or losses. This includes adjustments for retroactive applications of changes in accounting principles and prior period adjustments.}
  \item{Net assets -  End of year}{
  The sum of temporarily restricted, permanently restricted, and unrestricted net assets of the component unit organization.  It may be calculated as the difference between the assets and liabilities of the component unit.}
  \item{Total expenses and losses }{
  The sum of the outflow or other using up of assets or incurrence of liabilities (or a combination of both) from delivering or producing goods, rendering services, or carrying out other activities that constitute the insitution's ongoing major or central operations or in generating revenues and decreases in net assets from an organization's peripheral or incidental transactions and other events affecting the organization.}
  }
}
