\name{f_f1a_g}
\alias{f_f1a_g}

\title{Finance: F_F1A_G}

\description{
This data file contains the assets, liabilities, and net assets for component units of an institution.  These component units use the GASB accounting standard.

This file includes the name or names of the entities whose data is reported and the primary nature (general nature) of each component unit reported such as "fund raising/investment", "research", etc.
}

\section{Component Units}{
The term component units applies to GASB institutions only. A component unit is a legally separate organization for which the governing board and/or management of the primary institution is financially accountable. It can be another organization for which the nature and significance of its relationship with a primary institution is such that exclusion would cause the primary institution's financial statements to be misleading or incomplete.
}

\section{Data Availabilty}{
Financial data for component units are available for fiscal years 2002 through 2007.
}

\section{Glossary}{
\describe{
  \item{Form ID }{
  A numeric indicator of which Form F component unit is being reported for this record. }
  \item{Entities included }{
  Entities indicates what component units are included in this record.  This could be a single entity or a combination of entities with similar purpose. It may also be a combination of smaller component units.}
  \item{Primary nature of unit(s) (PNATURE) }{
   Indicates the primary purpose or nature of the component unit(s) reported in this record, such as fundraising or athletic program support.}
  \item{Total current assets}{
  Assets that are reasonably expected to be realized in  cash or sold or consumed during the next normal operating cycle (normally one   year) of the organization.}
  \item{Total non-current assets}{
  Assets that are not reasonably expected to be  realized in cash or sold or consumed during the next normal operating cycle (normally one year) of the organization.  }
  \item{Total assets }{
  The sum of current and noncurrent assets.  Assets are physical items (tangible) or rights (intangible) that have value and that are owned by the component unit.  An asset is useful to the organization because it is a source of future services or because it can be used to secure future benefits. }
  \item{ Total current liabilities}{
  Liabilities whose liquidation is reasonably expected to require the use of resources classified as current assets or the creation of other current liabilities within the next year.}
  \item{Total noncurrent liabilities}{
  Liabilities whose liquidation is not reasonably expected to require the use of resources classified as current assets or the creation of other current liabilities within the next year. This  includes the noncurrent portion of long-term debt and long-term accrued liabilities (such as for compensated absences, claims & judgments, and post-employment/post-retirement benefits).}
  \item{Total liabilities }{
  The sum of current and noncurrent liabilities.  }
  \item{Liabilities }{
  Debts and obligations of the component unit owed to outsiders  or claims or rights, expressed in monetary terms, of an component unit's creditors. }
  \item{Net assets - Invested in capital assets, net of related debt}{
  The total of all capital assets, reduced by accumulated depreciation, and reduced by the amount of outstanding bonds, mortgages, notes, or other borrowings that are attributable to the acquisition, construction, or improvement of those assets. }
  \item{Net assets - Restricted-expendable}{
  These assets have constraints placed on use that are either:

  (a) externally imposed by creditors, grantors, contributors, or laws and regulations of other governments or

  (b) imposed by law through constitutional provisions or enabling legislation, and are not required to be retained in perpetuity.}
  \item{Net assets - Restricted-nonexpendable }{
  These assets have constraints placed on use that are either:

  (a) externally imposed by creditors, grantors, contributors, or laws and regulations of other governments or

  (b) imposed by law through constitutional provisions or enabling legislation, and are required to be retained in perpetuity.}
  \item{Net assets - Unrestricted }{
  Net assets held by the component unit upon which no restrictions have been placed by the donor or other party external to the component unit.}
  \item{Total net assets }{
   The sum of net assets invested in capital assts, net of related debt, restricted-expendable net assets, restricted-nonexpendable net assets, and unrestricted net assets.  It can be calculated as the difference between total assets and total liabilities.}
  \item{Total operating revenues}{
  The sum of all operating revenues.  They result from providing services and producing and delivering goods.}
  \item{Total operating expenses }{
  The sum of all operating expenses that result from providing goods and services. Operating transactions are incurred in the course of the operating activities of the component unit. }
  \item{Expenses paid to institution }{
  The amount of total expenses that were disbursed from the component unit to the institution.  }
  \item{Net operating revenues (Expenses) }{
  The difference between total operating revenues and total operating expenses. }
  \item{Total nonoperating revenues}{
  The sum of all revenues generated from non-exchange transactions. }
  \item{Total nonoperating expenses }{
  The sum of all expenses incurred that are not classified as operating expenses. }
  \item{Net income before other revenues, expenses, gains, or losses }{
  The net value of  operating and nonoperating revenues and expenses.  }
  \item{Total other additions and deductions}{
  The total of all revenues, expenses, gains, and losses included in the component unit's general purpose financial statements below the line on the Statement of Revenues, Expenses, and Changes in Net Assets for income before other revenues, expenses, gains, and losses. }
  \item{ Change in net assets}{
  The net difference between total revenues and other additions and total expenses and other deductions.}
  \item{Net assets - Beginning of year }{
  The value of net assets as of the beginning of the fiscal year. }
  \item{Adjustments to beginning net assets}{
  Unusual and infrequent adjustments that are not recorded as current year revenues, expenses, gains, or losses. This includes adjustments for retroactive applications of changes in accounting principles and prior period adjustments. This line on the IPEDS survey is an automatic balancing line. }
  \item{Net assets - End of year }{
   The sum of net assets invested in capital assets, net of related debt, restricted-expendable net assets, restricted-nonexpendable net assets, and unrestricted net assets.  It can be calculated as the difference between total assets and total liabilities.}
  }
}
