\name{gr200}
\alias{gr200}

\title{Graduation Rates: GR200}

\description{
This file contains the graduation rate status as of August 31 of the current year for the cohort of full-time, first-time degree/certificate-seeking undergraduates.
}

\section{Four-Year Institutions}{
Data for four year institutions include:
\itemize{
  \item the number of bachelor degree-seeking students who were enrolled 8 years prior (the cohort)
  \item the number of bachelor degree seeking students who completed a bachelor's degree within 100, 150  or 200 percent of normal time.
  }
}

\section{Less than 4-year institutions}{
Data for less than 4-year institutions include:

\itemize{
  \item the number of full-time, first-time students who were enrolled 4 years prior (the cohort).
  \item the number of students who completed any degree/certificate within 100, 150, or 200 percent of normal time.
  }
}


\section{Glossary}{
\describe{
\item{4-year Graduation rate - bachelor's degree within 100 percent of normal time [BAGR100]}{The number of students from the adjusted bachelor's degree-seeking cohort, who completed a bachelor's degree within 100 percent of normal time (4-years) divided by the adjusted cohort.}

\item{6-year Graduation rate - bachelor's degree within 150 percent of normal time [BAGR150]}{The number of students from the adjusted bachelor's degree-seeking cohort, who completed a bachelor's degree within 150 percent of normal time (6-years) divided by the adjusted cohort.}

\item{8-year Graduation rate - bachelor's degree within 200 percent of normal time [BAGR200]}{The number of students from the adjusted bachelor's degree-seeking cohort, who completed a bachelor's degree within 200 percent of normal time (8-years) divided by the adjusted cohort.}

\item{Additional exclusions from bachelor's degree-seeking cohort [BAAEXCL]}{Additional exclusions who left institution between 150 percent of normal time and 200 percent of normal time (and have neither graduated nor transferred to another institution) because of one of the reasons described in "Exclusions."

NOTE: Students who leave the institution for one of the he reasons described in "Exclusions", but return prior to the 200 percent status date of August 31, may still be reported as exclusions.}

\item{Additional exclusions from degree/certificate-seeking cohort [L4AEXCL]}{Additional exclusions from degree/certificate-seeking cohort (two-year and less-than-two year institutions)

Additional exclusions who left institution between 150 percent of normal time and 200 percent of normal time (and have neither graduated nor transferred to another institution) because of the reasons described in "Exclusions."

NOTE: Students who leave the institution for the reasons described in "Exclusions", but return prior to the 200 percent status date of August 31, may still be reported as exclusions.}

\item{Adjusted bachelor's degree-seeking cohort within 150 percent of normal time [BAAC150]}{Revised cohort minus exclusions}

\item{Adjusted bachelor's degree-seeking cohort within 200 percent of normal time [BAAC200]}{Adjusted cohort minus additional exclusion who left institution between 150 percent of normal time and 200 percent of normal time.}

\item{Adjusted degree/certificate-seeking cohort within 150 percent of normal time [L4AC150]}{Revised cohort minus exclusions  (two-year and less-than-two year institutions)}

\item{Adjusted degree/certificate-seeking cohort within 200 percent of normal time [L4AC200]}{Adjusted degree/certificate-seeking cohort at 150 percent of normal time minus additional exclusions.(two-year and less-than-two year institutions)

NOTE: Normal time is 2 years for an associate's degree in a standard term-based institution; and the various scheduled times for certificate programs.}

\item{Exclusions}{Students from the cohort or subcohort who left the institution for any of the following reasons: died or were totally and permanently disabled; to serve in the armed forces; to serve with a foreign aid service of the Federal Government, such as the Peace Corps; or to serve on official church missions.}

\item{Exclusions from bachelor's degree-seeking cohort within 150 percent percent of normal time [BAEXCLU]}{Exclusions who left institution within 150 percent of normal time( and have neither graduated nor transferred to another institution) because of one of the reasons described in "Exclusions".

NOTE: Students who leave the institution for one of those reasons described in "Exclusions", but return prior to the 150 percent status date of August 31, may still be reported as exclusions.}

\item{Exclusions from degree/certificate-seeking cohort within 150 percent percent of normal time [L4EXCLU]}{Exclusions who left institution (two year and less-than-two year institutions) within 150 percent of normal time  and have neither graduated nor transferred to another institution) because of one of the reasons described in "Exclusions."

NOTE: Students who leave the institution for one of the reasons described in "Exclusions", but return prior to the 150 percent status date of August 31, may still be reported as exclusions.}

\item{Graduation rate - degree/certificate within 100 percent of normal time [L4GR100]}{The number of students (two-year and less-than-two year institutions) from the adjusted degree/certificate-seeking cohort, who completed a program within 100 percent of normal time  divided by the adjusted cohort.

NOTE: Normal time is 2 years for an associate's degree in a standard term-based institution; and the various scheduled times for certificate programs.}

\item{Graduation rate - degree/certificate within 150 percent of normal time [L4GR150]}{The number of students from the adjusted degree/certificate-seeking cohort,  who completed a program within 150 percent of normal time divided by the adjusted degree/certificate-seeking cohort.(two-year and less-than-two year institutions)

NOTE: Normal time is 2 years for an associate's degree in a standard term-based institution; and the various scheduled times for certificate programs.}

\item{Graduation rate - degree/certificate within 200 percent of normal time [L4GR200]}{The number of students from the adjusted degree/certificate-seeking cohort,  who completed a program within 200 percent of normal time divided by the adjusted cohort. (two-year and less-than-two year institutions)

NOTE: Normal time is 2 years for an associate's degree in a standard term-based institution; and the various scheduled times for certificate programs.}

\item{Number completed a  degree/certificate between 150 percent and 200 percent of normal time [L4NC200A]}{The number of students from the adjusted degree/certificate-seeking cohort,  who completed a program  between 150 percent and  200 percent of normal time. (two-year and less-than-two year institutions).

NOTE: Normal time is 2 years for an associate's degree in a standard term-based institution; and the various scheduled times for certificate programs.}

\item{Number completed a bachelor's degree between 150 percent and 200 percent of normal time [BANC200A]}{The number of students from the adjusted bachelor's degree-seeking cohort,  who completed a bachelor's degree between 150 percent and  200 percent (7 or 8 years) of normal time .}

\item{Number completed a bachelor's degree within 100 percent of normal time (4-years) [BANC100]}{The number of students from the adjusted bachelor's degree-seeking cohort,  who completed a bachelor's degree within 100 percent of normal time (4 years).}

\item{Number completed a bachelor's degree within 150 percent of normal time (6-years) [BANC150]}{The number of students from the adjusted bachelor's degree-seeking cohort,  who completed a bachelor's degree within 150 percent of normal time (6 years).}

\item{Number completed a bachelor's degree within 200 percent of normal time (8-years) [BANC200]}{The number of students from the adjusted bachelor's degree-seeking cohort, (8 years prior current), who completed a bachelor's degree within 200 percent of normal time  (8 years).}

\item{Number completed a degree/certificate  within 150 percent of normal time [L4NC150]}{The number of students  from the adjusted degree/certificate-seeking cohort,  who completed a program within 150 percent of normal time. (two-year and less-than-two year institutions)

NOTE: Normal time is 2 years for an associate's degree in a standard term-based institution; and the various scheduled times for certificate programs.}

\item{Number completed a degree/certificate within 100 percent of normal time [L4NC100]}{The number of students from the adjusted degree/certificate-seeking cohort, who completed a degree/certificate within 100 percent of normal time. (two-year and less-than-two year institutions).

NOTE: Normal time is 2 years for an associate's degree in a standard term-based institution; and the various scheduled times for certificate programs.}

\item{Number completed a degree/certificate within 200 percent of normal time [L4NC200]}{The number of students from the adjusted degree/certificate-seeking cohort,  who completed a program within 200 percent of normal time. (two-year and less-than-two year institutions).

NOTE: Normal time is 2 years for an associate's degree in a standard term-based institution; and the various scheduled times for certificate programs.}

\item{Revised degree/certificate-seeking cohort, (cohort 4 years prior current) [L4REVCT]}{Revised degree/certificate-seeking cohort (cohort 4 years prior current) in two-year and less-than-two year institutions}
\item{Still enrolled [BASTEND]}{The number of students who are still enrolled as of the fall census date}

\item{Still enrolled [L4STEND]}{The number of students who are still enrolled as of the fall census date (two-year and less-than-two year institutions).}



  }
}
