\name{gr_l2}
\alias{gr_l2}

\title{Graduation Rates: GR_L2}

\description{
This file contains the graduation rate status as of August 31 of current year for the cohort of full-time, first-time degree/certificate-seeking undergraduates enrolled in 3 years prior to the current year for less-than-2-year institutions. Included are the number of students enrolled in 2015 and the number of those students that completed their program in 100 or 150 percent of normal time, or had transferred out.
}

\section{Subcohorts}{
Beginning with 2016, graduation rate data are now available for the subcohorts of:
 \itemize{
  \item Pell grant recipients,
  \item Subsidized Stafford Loan recipients not receiving Pell Grants,
  \item students not receiving Pell grants or subsidized Stafford loans.
  }
The number who completed in 150 percent of normal time are included for the subcohorts.
}

\section{Glossary}{
\describe{
  \item{Adjusted cohort [LINE_50]}{
  Revised cohort minus exclusions.}

  \item{Did not receive Pell grant or subsidized Stafford loan - adjusted cohort [NRLIN50]}{
  The number cohort of students who did not receive a Pell grant of subsidized Stafford loan minus the number of exclusions.}

  \item{Did not receive Pell grant or subsidized Stafford loan - exclusions [NRLIN45]}{
  Exclusions from the revised cohort of students who did not receive a Pell grant or a subsidized Stafford loan within 150 percent percent of normal time.

  NOTE: Students who leave the institution for one of the above reasons, but return prior to the 150 percent status date of August 31, may still be reported as exclusions.}

  \item{Did not receive Pell grant or subsidized Stafford loan - revised cohort [NRLIN10]}{
  The total number of students of the revised cohort who did not receive a Pell grant or a subsidized stafford loan.}

  \item{Did not receive Pell grant or subsidized Stafford loan -  total number completed an award within 150 percent of normal time to completion [NRLIN11] }{
  This is the number of students in the adjusted cohort of students not receiving a Pell grant or subsidized Stafford loan, that completed an award in 150 percent of normal time to complete. }

  \item{Exclusions [LINE_45] }{
Those students who may be removed (deleted) from a cohort (or subcohort). For the Graduation Rates data collection students maybe removed from a cohort if they left the institution for one of the following reasons:
 \itemize{
  \item died or were totally and permanently disabled;
  \item to serve in the armed forces;
  \item to serve with a foreign aid service of the federal government, such as the Peace Corps;
  \item to serve on official church missions.}
  }

  \item{No longer enrolled [LINE_52] }{
   Adjusted cohort minus the sum of number of completers in 150 percent of normal time, number of transfer-outs and the number still enrolled.  }

  \item{Normal time to completion }{
  The amount of time necessary for a student to complete all requirements for a degree or certificate according to the institution's catalog. This is typically 4 years (8 semesters or trimesters, or 12 quarters, excluding summer terms) for a bachelor's degree in a standard term-based institution; 2 years (4 semesters or trimesters, or 6 quarters, excluding summer terms) for an associate's degree in a standard term-based institution; and the various scheduled times for certificate programs. }

  \item{Pell Grant recipients - adjusted  cohort [PGLIN50]}{
  The number of Pell grant recipients in the revised cohort minus the number of exclusions.}

  \item{Pell Grant recipients - exclusions [PGLIN45]}{
  Exclusions from the Pell grant cohort within 150 percent of normal time.

  NOTE: Students who leave the institution for one of the reasons described in "Exclusions [LINE_45]", but return prior to the 150 percent status date of August 31, may still be reported as exclusions. }

  \item{Pell Grant recipients - revised cohort [PGLIN10] }{
  Number of students in the cohort/subcohort who were recipients of a Pell Grant in the cohort year.

  Recipients are defined as those students receiving and using their Pell Grant.  These students must receive the award upon entry into the institution.  Students that are recipients of an award at a later point after entry, are not included.}

  \item{Pell Grant  recipients - total number that completed an award  within 150 percent of normal time to completion [PGLIN11]}{
  The number of students in the adjusted Pell grant recipient cohort that completed an award in 150 percent of normal time to complete.  }

  \item{Revised Cohort [LINE_10]}{
  Initial cohort after revisions are made. This is the number from which graduation and transfer-out rates are calculated. Cohorts may be revised if an institution discovers that incorrect data were reported in an earlier year.}

  \item{Still enrolled [LINE_51] }{
  The number of students who are still enrolled at your institution as of the fall census date for the fall term following 150 percent of normal time to completion of their program.}

  \item{Subsidized Stafford Loan recipients not receiving Pell Grants - adjusted cohort [SSLIN50] }{
  The number of susidized Stafford loan recipients who did not receive a Pell Grant in the revised cohort minus the number of exclusions. }

  \item{Subsidized Stafford Loan recipients not receiving Pell Grants - exclusions [SSLIN45]}{
  Exclusions from the subsidized Stafford loan who did not receive a Pell grant cohort within 150 percent of normal time.

NOTE: Students who leave the institution for one of the reasons described in "Exclusions [LINE_45]", but return prior to the 150 percent status date of August 31, may still be reported as exclusions. }

  \item{Subsidized Stafford Loan recipients not receiving Pell Grants - revised cohort [SSLIN10]}{
  Number of students in the cohort/subcohort who were recipients of a subsidized Stafford loan and who did not receive a Pell grant in the cohort year.

  Recipients are defined as those students receiving and using their subsidized Stafford loan.  These students must receive the award upon entry into the institution.  Students that are recipients of an award at a later point after entry, are not included.  Students receiving a Pell grant are not included.}

  \item{Subsidized Stafford Loan recipients not receiving Pell Grants - total number completed an award within 150 percent of normal time to completion [SSLIN11]}{
  The number of students in the adjusted cohort of subsidized Stafford loan recipients who did not receive a Pell grant, that completed an award in 150 percent of normal time to complete.}

  \item{Transfer-out students [LINE_30] }{
  A student that leaves the reporting institution and enrolls at another institution. }
  }
}
