\name{hd}
\alias{hd}

\title{Institutional Characteristics: HD}

\description{
This file contains directory information for every institution in the IPEDS universe of current year.
}

\section{Characteristics}{
The file includes name, address, city, state, zip code and various URL links to the institution's home page, admissions, financial aid offices and  the net price calculator.  Identifies institutions as currently active, institutions that participate in Title IV federal financial aid programs for which IPEDS is mandatory.

It also includes variables derived from the 2019 - 20 Institutional Characteristics survey, such as control and level of institution, highest level and highest degree offered and Carnegie classifications.
}


\section{Glossary}{
\describe{
\item{Academic year}{The period of time generally extending from September to June; usually equated to 2 semesters or trimesters, 3 quarters, or the period covered by a 4 - 1 - 4 calendar system.
}

\item{Bureau of Economic Analysis (BEA) regions [OBEREG]}{A set of Geographic Areas that are aggregations of the states. The regional classifications, which were developed in the mid - 1950s, are based on the homogeneity of the states in terms of economic characteristics, such as the industrial composition of the labor force, and in terms of demographic, social, and cultural characteristics. BEA groups all 50 states and the District of Columbia into eight distinct regions for purposes of data collecting and analyses.  IPEDS/HEGIS assigns U.S. service schools as their own region.

\itemize{
  \item 0  -  US Service schools
  \item 1  -  New England CT ME MA NH RI VT
  \item 2  -  Mid East DE DC MD NJ NY PA
  \item 3  -  Great Lakes IL IN MI OH WI
  \item 4  -  Plains IA KS MN MO NE ND SD
  \item 5  -  Southeast AL AR FL GA KY LA MS NC SC TN VA WV
  \item 6  -  Southwest AZ NM OK TX
  \item 7  -  Rocky Mountains CO ID MT UT WY
  \item 8  -  Far West AK CA HI NV OR WA
  \item 9  -  Outlying areas AS FM GU MH MP PR PW VI
  \item  - 3  -  Not available
}
}

\item{Carnegie Classification 2000 [CARNEGIE]}{The 2000 Carnegie Classification includes all colleges and universities in the United States that are degree - granting and accredited by an agency recognized by the U.S. Secretary of Education.

The 2000 edition classifies institutions based on their degree - granting activities from 1995 - 96 through 1997 - 98.

\itemize{
  \item 15 Doctoral/Research Universities -  - Extensive: These institutions typically offer a wide range of baccalaureate programs, and they are committed to graduate education through the doctorate.
They award 50 or more doctoral degrees per year across at least 15 disciplines.
  \item16 Doctoral/Research Universities -  - Intensive:
These institutions typically offer a wide range of baccalaureate programs, and they are committed to
graduate education through the doctorate. They award at least ten doctoral degrees per year across three or more disciplines,2 or at least 20 doctoral degrees per year overall.
  \item 21 Master's Colleges and Universities I: These institutions typically offer a wide range of baccalaureate programs, and they are committed to graduate education through the master's degree. They award 40 or more master's degrees per year across three or more disciplines.
  \item 22 Master's (Comprehensive) Colleges and Universities II: These institutions typically offer a wide range of baccalaureate programs, and they are committed to graduate education through the master's degree.  They award 20 or more master's degrees per year.
  \item 31 Baccalaureate Colleges -  - Liberal Arts: These institutions are primarily undergraduate colleges with major emphasis on baccalaureate programs. They award at least half of their baccalaureate degrees in liberal arts fields.
  \item 32 Baccalaureate Colleges -  - General: These institutions are primarily undergraduate colleges with major emphasis on baccalaureate programs. They award less than half of their baccalaureate degrees in liberal arts fields.
  \item 33 Baccalaureate/Associate's Colleges:
These institutions are undergraduate colleges where the majority of conferrals are at the subbaccalaureate level (associate's degrees and certificates), but bachelor's degrees account for at least ten percent of undergraduate awards.
  \item 40 Associate's Colleges: These institutions offer associate's degree and certificate programs but, with few exceptions, award no baccalaureate degrees.
}

Specialized Institutions  -  These institutions offer degrees ranging from the bachelor's to the doctorate, and typically award a majority of degrees in a single field. The list includes only institutions that are listed as separate campuses in the Higher Education Directory.
Specialized institutions include:

\itemize{
  \item 51 Theological seminaries and other specialized faith - related institutions:
These institutions primarily offer religious instruction or train members of the clergy.
  \item 52 Medical schools and medical centers:
These institutions award most of their professional degrees in medicine. In some instances, they include other health professions programs, such as dentistry, pharmacy, or nursing.
  \item 53 Other separate health profession schools: These institutions award most of their degrees in such fields as chiropractic, nursing, pharmacy, or podiatry.
  \item 54 Schools of engineering and technology: These institutions award most of their bachelor's or graduate degrees in technical fields of study.
  \item 55 Schools of business and management: These institutions award most of their bachelor's or graduate degrees in business or business - related programs.
  \item 56 Schools of art, music, and design: These institutions award most of their bachelor's or graduate degrees in art, music, design, architecture, or some combination of such fields.
  \item 57 Schools of law: These institutions award most of their degrees in law.
  \item 58 Teachers colleges: These institutions award most of their bachelor's or graduate degrees in education or education - related fields.
  \item 59 Other specialized institutions: Institutions in this category include graduate centers, maritime academies, military institutes, and institutions that do not fit any other classification category.
  \item 60 Tribal Colleges and Universities: These colleges are, with few exceptions, tribally controlled and located on reservations. They are all members of the American Indian Higher Education Consortium.
}
}


\item{Carnegie Classification 2005/2010: Basic [CCBASIC]}{
To ensure continuity of the classification framework and to allow comparison across years, the 2010 Classification update retains the same structure of six parallel classifications, initially adopted in 2005. They are as follows: Basic Classification (the traditional Carnegie Classification Framework), Undergraduate and Graduate Instructional Program classifications, Enrollment Profile and Undergraduate Profile classifications, and Size & Setting classification.

These classifications are time - specific snapshots of institutional attributes and behavior based on data from 2008 to 2010. Collectively they depict the most current landscape of U.S. colleges and universities. Institutions might be classified differently using a different timeframe. Individual classifications are not updated with more recent data.

Associate's Colleges. Includes institutions where all degrees are at the associate's level, or where bachelor's degrees account for less than 10 percent of all undergraduate degrees. Excludes institutions eligible for classification as Tribal Colleges or Special Focus Institutions.
\itemize{
  \item Assoc/Pub - R - S: Associate's  -  Public Rural - serving Small
  \item Assoc/Pub - R - M: Associate's  -  Public Rural - serving Medium
  \item Assoc/Pub - R - L: Associate's  -  Public Rural - serving Large
  \item Assoc/Pub - S - SC: Associate's  -  Public Suburban - serving Single Campus
  \item Assoc/Pub - S - MC: Associate's  -  Public Suburban - serving Multicampus
  \item Assoc/Pub - U - SC: Associate's  -  Public Urban - serving Single Campus
  \item Assoc/Pub - U - MC: Associate's  -  Public Urban - serving Multicampus
  \item Assoc/Pub - Spec: Associate's  -  Public Special Use
  \item Assoc/PrivNFP: Associate's  -  Private Not - for - profit
  \item Assoc/PrivFP: Associate's  -  Private For - profit
  \item Assoc/Pub2in4: Associate's  -  Public 2 - year Colleges under Universities
  \item Assoc/Pub4: Associate's  -  Public 4 - year, Primarily Associate's
  \item Assoc/PrivNFP4: Associate's  -  Private Not - for - profit 4 - year, Primarily Associate's
  \item Assoc/PrivFP4: Associate's  -  Private For - profit 4 - year, Primarily Associate's
}

Doctorate - granting Universities. Includes institutions that award at least 20 doctoral degrees per year (excluding doctoral - level degrees that qualify recipients for entry into professional practice, such as the JD, MD, PharmD, DPT, etc.) Excludes Special Focus Institutions and Tribal Colleges.
\itemize{
  \item RU/VH: Research Universities (very high research activity)
  \item RU/H: Research Universities (high research activity)
  \item DRU: Doctoral/Research Universities
}

Master's Colleges and Universities. Includes institutions that award at least 50 master's degrees per year. Excludes Special Focus Institutions and Tribal Colleges.
\itemize{
  \item Master's/L: Master's Colleges and Universities (larger programs)
  \item Master's/M: Master's Colleges and Universities (medium programs)
  \item Master's/S: Master's Colleges and Universities (smaller programs)
}

Baccalaureate Colleges. Includes institutions where baccalaureate degrees represent at least 10 percent of all undergraduate degrees and that award fewer than 50 master's degrees or fewer than 20 doctoral degrees per year. Excludes Special Focus Institutions and Tribal Colleges.
\itemize{
  \item Bac/A&S: Baccalaureate Colleges  -  Arts & Sciences
  \item Bac/Diverse: Baccalaureate Colleges  -  Diverse Fields
  \item Bac/Assoc: Baccalaureate/Associate's Colleges
}

Special Focus Institutions. Institutions awarding baccalaureate or higher - level degrees where a high concentration of degrees is in a single field or set of related fields. Excludes Tribal Colleges.
\itemize{
  \item Spec/Faith: Theological seminaries, Bible colleges, and other faith - related institutions
  \item Spec/Medical: Medical schools and medical centers
  \item Spec/Health: Other health profession schools
  \item Spec/Engg: Schools of engineering
  \item Spec/Tech: Other technology - related schools
  \item Spec/Bus: Schools of business and management
  \item Spec/Arts: Schools of art, music, and design
  \item Spec/Law: Schools of law
  \item Spec/Other: Other special - focus institutions
}

Tribal Colleges. Colleges and universities that are members of the American Indian Higher Education Consortium, as identified in IPEDS Institutional Characteristics.
\itemize{
  \item Tribal: Tribal Colleges
}

For a complete description and technical details visit the Carnegie Foundation Website at www.carnegiefoundation.org/classifications/}

\item{Carnegie Classification 2015: Basic [C15BASIC]}{The Basic Classification is an update of the traditional classification framework developed by the Carnegie Commission on Higher Education in 1970 to support its research program. The Basic Classification was published for use in 1973, and subsequently updated in 1976, 1987, 1994, 2000, 2005 and 2010 and 2015. The 2005 edition involved some significant changes from the previous releases by dividing Associate's colleges into subcategories and using a multi - measure research index to classify doctorate - granting institutions. The 2010 update retained the same classification structure as the 2005 edition. In the 2015 update, the Associate's Categories were substantially redefined and the categories of the Research Doctoral Universities changed (but not the calculation methodology). Please see the Basic Classification Methodology for details regarding how this classification is calculated.

Note: The "shorthand" labels for the Doctoral Universities and Master's Colleges and Universities were restored in the 2015 update to numeric sequences (R1, R2, R3, and M1, M2, M3) to denote that each one is based on differences in quantitative levels. For doctoral universities, the levels are based on a research activity index and for master's colleges and universities it is based on number of degrees conferred.

Doctoral Universities  -  Includes institutions that awarded at least 20 research/scholarship doctoral degrees during the update year (this does not include professional practice doctoral - level degrees, such as the JD, MD, PharmD, DPT, etc.). Excludes Special Focus Institutions and Tribal Colleges.

Master's Colleges and Universities  -  Generally includes institutions that awarded at least 50 master's degrees and fewer than 20 doctoral degrees during the update year (with occasional exceptions   -   see Methodology). Excludes Special Focus Institutions and Tribal Colleges.

Baccalaureate Colleges  -  Includes institutions where baccalaureate or higher degrees represent at least 50 percent of all degrees but where fewer than 50 master's degrees or 20 doctoral degrees were awarded during the update year. (Some institutions above the master's degree threshold are also included; see Methodology.) Excludes Special Focus Institutions and Tribal Colleges.

Baccalaureate/Associate's Colleges  -  Includes four - year colleges (by virtue of having at least one baccalaureate degree program) with fewer than 50 percent of degrees awarded at the associate's level. Excludes Special Focus Institutions and Tribal Colleges.

Associate's Colleges  -  Institutions at which the highest level degree awarded is an associate's degree. The institutions are sorted into nine categories based on the intersection of two factors: disciplinary focus (transfer, career & technical or mixed) and dominant student type (traditional, nontraditional or mixed). Excludes Special Focus Institutions and Tribal Colleges.

Special Focus Institutions -  Institutions where a high concentration of degrees is in a single field or set of related fields. Excludes Tribal Colleges.

Tribal Colleges  -  Colleges and universities that are members of the American Indian Higher Education Consortium, as identified in IPEDS Institutional Characteristics.

Classifications are time - specific snapshots of institutional attributes and behavior based on 2013 - 14 data.

The methodology can be found at  http://carnegieclassifications.iu.edu/methodology/basic.php.}

\item{Carnegie Classification 2018: Basic [C18BASIC]}{The Basic Classification is an update of the traditional classification framework developed by the Carnegie Commission on Higher Education in 1970 to support its research program. The Basic Classification was originally published for public use in 1973, and subsequently updated in 1976, 1987, 1994, 2000, 2005, 2010, 2015 and 2018. In the 2018 update, the Doctoral Universities have been reshaped to better accommodate “Doctor's degree   -   professional practice” within our methodology. Please see the Basic Classification Methodology for details regarding how this classification is calculated.

Doctoral universities  -  Includes institutions that awarded at least 20 research/scholarship doctoral degrees during the update year and also institutions with below 20 research/scholarship doctoral degrees that awarded at least 30 professional practice doctoral degrees in at least 2 programs. Excludes Special Focus Institutions and Tribal Colleges.

The first two categories include only institutions that awarded at least 20 research/scholarship doctoral degrees and had at least $5 million in total research expenditures (as reported through the National Science Foundation (NSF) Higher Education Research & Development Survey (HERD)).
\itemize{
  \item R1: Doctoral Universities   -   Very high research activity
  \item R2: Doctoral Universities   -   High research activity
  \item D/PU: Doctoral/Professional Universities
  }

Master's Colleges and Universities  -  Generally includes institutions that awarded at least 50 master's degrees and fewer than 20 doctoral degrees during the update year (with occasional exceptions   -   see Methodology). Excludes Special Focus Institutions and Tribal Colleges.

Baccalaureate Colleges  -  Includes institutions where baccalaureate or higher degrees represent at least 50 percent of all degrees but where fewer than 50 master's degrees or 20 doctoral degrees were awarded during the update year. (Some institutions above the master's degree threshold are also included; see Methodology.) Excludes Special Focus Institutions and Tribal Colleges.

Baccalaureate/Associate's Colleges  -  Includes four - year colleges (by virtue of having at least one baccalaureate degree program)  that conferred more than 50 percent of degrees at the associate's level. Excludes Special Focus Institutions and Tribal Colleges.

Associate's Colleges  -  Institutions at which the highest level degree awarded is an associate's degree. The institutions are sorted into nine categories based on the intersection of two factors: disciplinary focus (transfer, career & technical or mixed) and dominant student type (traditional, nontraditional or mixed). Excludes Special Focus Institutions and Tribal Colleges.

Special Focus Institutions -  Institutions where a high concentration of degrees is in a single field or set of related fields. Excludes Tribal Colleges.

Tribal Colleges  -  Colleges and universities that are members of the American Indian Higher Education Consortium, as identified in IPEDS Institutional Characteristics.

Classifications are time - specific snapshots of institutional attributes and behavior based on 2016 - 17 data.

The methodology can be found at  http://carnegieclassifications.iu.edu/definitions.php.}

\item{Carnegie Classification 2018: Enrollment Profile [C18ENPRF]}{Enrollment Profile Classification provides a bird's eye view of the student population by grouping institutions according to the mix of students enrolled at the undergraduate and graduate levels. Exclusively undergraduate institutions are further disaggregated by level (two - year and four - year). For institutions with both undergraduate and graduate students, institutions are grouped according to the distribution of full - time equivalent (FTE*) students across the two levels, giving an approximate measure of the student population's "center of gravity." As a result, it reflects important differences with respect to educational mission as well as institutional climate and culture  -  differences that can have implications for infrastructure, services, and resource allocation. For more information regarding how this classification is calculated, go to http://carnegieclassifications.iu.edu/definitions.php. Classifications are time - specific snapshots of institutional attributes and behavior based on 2016 - 17 data.

\itemize{
  \item Exclusively undergraduate two - year  -  Only undergraduates enrolled at these associate's degree granting institutions.
  \item Exclusively undergraduate four - year  -  Only undergraduates enrolled at these bachelor's or higher degree granting institutions.
  \item Very high undergraduate  -  Undergraduate and graduate students enrolled, with the latter group accounting for less than 10 percent of FTE* enrollment.
  \item High undergraduate  -  Undergraduate and graduate students enrolled, with the latter group accounting for 10  -  24 percent of FTE* enrollment.
  \item Majority undergraduate  -  Undergraduate and graduate students enrolled, with the latter group accounting for 25  -  49 percent of FTE* enrollment.
  \item Majority graduate  -  Undergraduate and graduate students enrolled, with the latter group accounting for at least half of FTE* enrollment.
  \item Exclusively graduate  -  Only graduate students enrolled.
}

*FTE: Full - time equivalent enrollment was calculated as full - time headcount plus one - third part - time headcount.}

\item{Carnegie Classification 2018: Graduate Instructional Program [C18IPGRD]}{Graduate and Instructional Program classification examines the nature of graduate education, with a special focus on the mix of graduate programs across fields of study. In this classification, a single graduate - level degree
qualifies an institution for inclusion.

The classification is based on the level of graduate degrees awarded (master's degrees, and doctoral degrees categorized as either research/scholarship, professional practice, or other doctorate),
the number of fields represented by the degrees awarded, and the mix or concentration of degrees by broad disciplinary domain. The classification has two parts: one for institutions that award at
least one research/scholarship doctoral degrees (hereinafter referred to as research doctoral degrees), and one postbaccalaureate degree - granting institutions that either offer only master's
degrees or that also offer professional practice or other doctoral degrees (based on the record of degree conferrals, not program offerings). Within each group, institutions are classified respect
to the breadth of graduate offerings and the concentration of degrees in certain fields or combinations of fields.

For two categories of institutions offering research doctorates, institutions offering medical education (defined as human or veterinary medical education, including allopathic medicine,
osteopathic medicine, dentistry, and veterinary medicine) were distinguished. Institutions in other categories may also offer medical education, but the numbers were not large enough to justify
subcategories, and judged it preferable to differentiate with respect to the other graduate fields, rather than with respect to the presence or absence of medical education.

NOTE: Because a single research doctoral degree (as defined in the IPEDS data collection of the National Center for Education Statistics) qualifies an institution for inclusion in the doctoral categories, institutions with large master's or professional programs and modest doctoral - level programs are currently classified according to their doctoral programs.
The categories are listed below. The term "comprehensive" is used here to denote comprehensiveness of offerings across a range of fields.
Classifications are time - specific snapshots of institutional attributes and behavior based on 2016 - 17 data. For the methodology, go to http://carnegieclassifications.iu.edu/definitions.php
\itemize{
  \item Postbaccalaureate: Single program  -  Education  -  Awarded master's or professional practice/other doctoral degrees in education as their only postbaccalaureate program.
  \item Postbaccalaureate: Single program  -  Business  -   Awarded master's or professional practice/other doctoral degrees in business as their only postbaccalaureate program.
  \item Postbaccalaureate: Single program  -  Other  -  Awarded master's or professional practice/other doctoral degrees in a single field other than education or business as their only postbaccalaureate program.
  \item Postbaccalaureate: Comprehensive programs  -  Awarded at least one master's degree or professional practice/other doctoral degrees in each of the humanities, social sciences, and STEM* fields, as well as such graduate degrees in one or more professional fields.
  \item Postbaccalaureate: Arts & sciences - dominant  -  Awarded master's or professional practice/other doctoral degrees in some arts and sciences fields. They may also award master's or non - research doctoral degrees in other fields, but in lesser numbers.
  \item Postbaccalaureate: Education - dominant, with arts & sciences  -  Awarded master's or professional practice/other doctoral degrees in both arts and sciences and professional fields, and the field with the largest number of such graduate degrees was education.
  \item Postbaccalaureate: Business - dominant, with arts & sciences  -  Awarded master's or professional practice/other doctoral degrees in both arts and sciences and professional fields, and the field with the largest number of such graduate degrees was business.
  \item Postbaccalaureate: Other - dominant, with arts & sciences  -  Awarded master's or professional practice/other doctoral degrees in both arts and sciences and professional fields, and the field with the largest number of such graduate degrees was a professional field other than business or education.
  \item Postbaccalaureate: Education - dominant, with other professional programs  -  Awarded master's or professional practice/other doctoral degrees in professional fields only, and the field with the largest number of such graduate degrees was education.
  \item Postbaccalaureate: Business - dominant, with other professional programs  -  Awarded master's or professional practice/other doctoral degrees in professional fields only, and the field with the largest number of such graduate degrees was business.
  \item Postbaccalaureate: Other - dominant, with other professional programs  -  Awarded master's or professional practice/other doctoral degrees in professional fields only, and the field with the largest number of such graduate degrees was a field other than business or education.
  \item Research Doctoral: Single program  -  Education  -  Awarded research doctoral degrees in education but not in other fields (they may have more extensive offerings at the master's or professional practice/other doctoral level).
  \item Research Doctoral: Single program  -  Other  -   Awarded research doctoral degrees in a single field other than education (they may have more extensive offerings at the master's or professional practice/other doctoral level).
  \item Research Doctoral: Comprehensive programs, with medical/veterinary school  -  Awarded research doctoral degrees in the humanities, social sciences, and STEM* fields, as well as in medicine, dentistry, and/or veterinary medicine. They also offer may also offer master's and professional practice/other doctoral degrees in other fields.
  \item Research Doctoral: Comprehensive programs, no medical/veterinary school  -  Awarded research doctoral degrees in the humanities, social sciences, and STEM* fields. They may also offer master's or professional practice/other degrees in fields other than medicine, dentistry, or veterinary medicine.
  \item Research Doctoral: Humanities/social sciences - dominant  -  Awarded research doctoral degrees in a range of fields, with the largest number of research doctorates in the humanities or social sciences.
  \item Research Doctoral: STEM - dominant  -  Awarded research doctoral degrees in a range of fields, with the largest number of research doctorates in the STEM* fields.
  \item Research Doctoral: Professional - dominant  -  Awarded research doctoral degrees in a range of fields, and the largest number of research doctorates were in professions other than engineering (such as education, health professions, law, public policy, or social work).
}

* STEM: Science, technology, engineering, and mathematics.}

\item{Carnegie Classification 2018: Size and Setting [C18SZSET]}{The Size and Setting Classification describes institutions' size (student population) and residential character. Because residential character applies to the undergraduate student body, exclusively graduate institutions are not included. For the full definition and methodology go to http://carnegieclassifications.iu.edu/definitions.php. Classifications are time - specific snapshots of institutional attributes and behavior based on 2016 - 17 data.

Two year
\itemize{
  \item Very small  -   FTE* enrollment of fewer than 500 students at these associate's degree granting institutions.
  \item Small  -   FTE* enrollment of 500  -  1,999 students at these associate's degree granting institutions.
  \item Medium  -   FTE* enrollment of 2,000  -  4,999 students at these associate's degree granting institutions.
  \item Large  -  FTE* enrollment of 5,000  -  9,999 students at these associate's degree granting institutions.
  \item Very large  -  FTE* enrollment of at least 10,000 students at these associate's degree granting institutions.
}

Four - year
\itemize{
  \item Very small Primarily nonresidential  -  FTE* enrollment of fewer than 1,000 degree - seeking students at these bachelor's or higher degree granting institutions. Fewer than 25 percent of degree - seeking undergraduates live on campus** and/or fewer than 50 percent attend full time (includes exclusively distance education institutions).
  \item Very small Primarily residential  -  FTE* enrollment of fewer than 1,000 degree - seeking students at these bachelor's or higher degree granting institutions. 25 - 49 percent of degree - seeking undergraduates live on campus** and at least 50 percent attend full time.
  \item Very small Highly residential  -  FTE* enrollment of fewer than 1,000 degree - seeking students at these bachelor's or higher degree granting institutions. At least half of degree - seeking undergraduates live on campus** and at least 80 percent attend full time.
  \item Small Primarily nonresidential  -  FTE* enrollment of 1,000  -  2,999 degree - seeking students at these bachelor's or higher degree granting institutions. Fewer than 25 percent of degree - seeking undergraduates live on campus** and/or fewer than 50 percent attend full time (includes exclusively distance education institutions).
  \item Small Primarily residential  -   FTE* enrollment of 1,000  -  2,999 degree - seeking students at these bachelor's or higher degree granting institutions. 25 - 49 percent of degree - seeking undergraduates live on campus** and at least 50 percent attend full time.
  \item Small Highly residential  -  FTE* enrollment of 1,000  -  2,999 degree - seeking students at these bachelor's or higher degree granting institutions. At least half of degree - seeking undergraduates live on campus** and at least 80 percent attend full time.
  \item Medium Primarily nonresidential  -  FTE* enrollment of 3,000  -  9,999 degree - seeking students at these bachelor's or higher degree granting institutions. Fewer than 25 percent of degree - seeking undergraduates live on campus** and/or fewer than 50 percent attend full time (includes exclusively distance education institutions).
  \item Medium Primarily residential  -  FTE* enrollment of 3,000  -  9,999 degree - seeking students at these bachelor's or higher degree granting institutions. 25 - 49 percent of degree - seeking undergraduates live on campus** and at least 50 percent attend full time.
  \item Medium Highly residential  -  FTE* enrollment of 3,000  -  9,999 degree - seeking students at these bachelor's or higher degree granting institutions. At least half of degree - seeking undergraduates live on campus** and at least 80 percent attend full time.
  \item Large Primarily nonresidential  -  FTE* enrollment of at least 10,000 degree - seeking students at these bachelor's or higher degree granting institutions. Fewer than 25 percent of degree - seeking undergraduates live on campus** and/or fewer than 50 percent attend full time (includes exclusively distance education institutions).
  \item Large Primarily residential  -  FTE* enrollment of at least 10,000 degree - seeking students at these bachelor's or higher degree granting institutions. 25 - 49 percent of degree - seeking undergraduates live on campus** and at least 50 percent attend full time.
  \item Large Highly residential  -  FTE* enrollment of at least 10,000 degree - seeking students at these bachelor's or higher degree granting institutions. At least half of degree - seeking undergraduates live on campus** and at least 80 percent attend full time.
}

Exclusively graduate/professional  -  Fall enrollment data indicate that there are no undergraduates enrolled at these institutions. All enrolled students are in graduate - level programs.

*FTE: Full - time equivalent enrollment was calculated as full - time plus one - third part - time.

**On campus is defined as institutionally - owned,  - controlled, or  - affiliated housing.}

\item{Carnegie Classification 2018: Undergraduate Instructional Program [C18IPUG]}{The Undergraduate Instructional Program Classification focuses attention on undergraduate education, regardless of the presence or extent of graduate education. Undergraduate education is an essential component of what most colleges and universities do, as the vast majority of U.S. institutions of higher education teach undergraduates. Indeed, even at institutions with strong commitments to graduate education and the production of new knowledge through research and scholarship, the undergraduate program usually accounts for the majority of student enrollment. For the full definition and methodology, go to http://carnegieclassifications.iu.edu/definitions.php

Classifications are time - specific snapshots of institutional attributes and behavior based on 2016 - 17 data.

\itemize{
  \item Associate's Colleges, High Transfer  -  Awarded associate's degrees but no bachelor's degrees with fewer than 30percent of awards (degrees and certificates) in career & technical programs.
  \item Associate's Colleges, Mixed Transfer/Career & Technical  -  Awarded associate's degrees but no bachelor's degrees with 30 - 49percent of awards (degrees and certificates) in career & technical programs.
  \item Associate's Colleges, High Career & Technical  -  Awarded associate's degrees but no bachelor's degrees with more than 50percent of awards (degrees and certificates) in career & technical programs.
  \item Special Focus Two - Year Institutions  -  Awarded associate's degrees but no bachelor's degrees with typically more than 75percent of awards in a single career & technical program.
  \item Baccalaureate/Associate's Colleges  -  Awarded both associate's and bachelor's degrees, but the majority of degrees awarded were at the associate's level.
  \item Arts & sciences focus, no graduate coexistence  -  At least 80 percent of bachelor's degree majors were in the arts and sciences, and no graduate degrees were awarded in fields corresponding to undergraduate majors.
  \item Arts & sciences focus, some graduate coexistence  -  At least 80 percent of bachelor's degree majors were in the arts and sciences, and graduate degrees were observed in up to half of the fields corresponding to undergraduate majors.
  \item Arts & sciences focus, high graduate coexistence  -  At least 80 percent of bachelor's degree majors were in the arts and sciences, and graduate degrees were observed in at least half of the fields corresponding to undergraduate majors.
  \item Arts & sciences plus professions, no graduate coexistence  -  60  -  79 percent of bachelor's degree majors were in the arts and sciences, and no graduate degrees were awarded in fields corresponding to undergraduate majors.
  \item Arts & sciences plus professions, some graduate coexistence  -  60  -  79 percent of bachelor's degree majors were in the arts and sciences, and graduate degrees were observed in up to half of the fields corresponding to undergraduate majors.
  \item Arts & sciences plus professions, high graduate coexistence  -  60  -  79 percent of bachelor's degree majors were in the arts and sciences, and graduate degrees were observed in at least half of the fields corresponding to undergraduate majors.
  \item Balanced arts & sciences/professions, no graduate coexistence  -  Bachelor's degrees awarded were relatively balanced between arts and sciences and professional fields (41  -  59 percent in each), and no graduate degrees were awarded in fields corresponding to undergraduate majors.
  \item Balanced arts & sciences/professions, some graduate coexistence  - Bachelor's degree majors were relatively balanced between arts and sciences and professional fields (41  -  59 percent in each), and graduate degrees were observed in up to half of the fields corresponding to undergraduate majors.
  \item Balanced arts & sciences/professions, high graduate coexistence  -  Bachelor's degree majors were relatively balanced between arts and sciences and professional fields (41  -  59 percent in each), and graduate degrees were observed in at least half of the fields corresponding to undergraduate majors.
  \item Professions plus arts & sciences, no graduate coexistence  -  60  -  79 percent of bachelor's degree majors were in professional fields (such as business, education, engineering, health, and social work), and no graduate degrees were awarded in fields corresponding to undergraduate majors.
  \item Professions plus arts & sciences, some graduate coexistence  -  60  -  79 percent of bachelor's degree majors were in professional fields, and graduate degrees were observed in up to half of the fields corresponding to undergraduate majors.
  \item Professions plus arts & sciences, high graduate coexistence  -  60  -  79 percent of bachelor's degree majors were in professional fields, and graduate degrees were observed in at least half of the fields corresponding to undergraduate majors.
  \item Professions focus, no graduate coexistence  -  At least 80 percent of bachelor's degree majors were in professional fields (such as business, education, engineering, health, and social work), and no graduate degrees were awarded in fields corresponding to undergraduate majors.
  \item Professions focus, some graduate coexistence  -  At least 80 percent of bachelor's degree majors were in professional fields, and graduate degrees were observed in up to half of the fields corresponding to undergraduate majors.
  \item Professions focus, high graduate coexistence  -  At least 80 percent of bachelor's degree majors were in professional fields, and graduate degrees were observed in at least half of the fields corresponding to undergraduate majors.
}
}

\item{Carnegie Classification 2018: Undergraduate Profile [C18UGPRF]}{The Undergraduate Profile Classification describes the undergraduate population with respect to three characteristics: the proportion of undergraduate students who attend part -  or full - time; background academic achievement characteristics of first - year, first - time students; and the proportion of entering students who transfer in from another institution. Each of these captures important differences in the nature of the undergraduate population. They do not imply differences in the quality of undergraduate education, but they have implications for how an institution serves its students. Please see the Undergraduate Profile Methodology for more detail regarding how this classification was calculated at http://carnegieclassifications.iu.edu/definitions.php.
 Classifications are time - specific snapshots of institutional attributes and behavior based on 2016 - 17 data.

Two - year
\itemize{
  \item Higher part - time  -   60 percent of undergraduates are enrolled part - time at these associate's degree granting institutions.
  \item Mixed part/full - time  -  40  -  59 percent of undergraduates are enrolled part - time at these associate's degree granting institutions.
  \item Medium full - time  -   10  -  39 percent of undergraduates are enrolled part - time at these associate's degree granting institutions.
  \item Higher full - time  -   Less than 10 percent of undergraduates enrolled part - time at these associate's degree granting institutions.
}

Four - year
\itemize{
  \item Higher part - time  -  At least 40 percent of undergraduates are enrolled part - time at these bachelor's or higher degree granting institutions.
  \item Medium full - time, inclusive, lower transfer - in  -  60  -  79 percent of undergraduates are enrolled full - time at these bachelor's or higher degree granting institutions. These institutions either did not report test score data or the scores indicate that they extend educational opportunity to a wide range of students with respect to academic preparation and achievement. Fewer than 20 percent of entering undergraduates are transfer students.
  \item Medium full - time, inclusive, higher transfer - in  -  60  -  79 percent of undergraduates are enrolled full - time at these bachelor's or higher degree granting institutions. These institutions either did not report test score data or the scores indicate that they extend educational opportunity to a wide range of students with respect to academic preparation and achievement. At least 20 percent of entering undergraduates are transfer students.
  \item Medium full - time, selective or more selective, lower transfer - in  -  60  -  79 percent of undergraduates are enrolled full - time at these bachelor's or higher degree granting institutions. Test score data for first - year students indicate that these institutions are selective or more selective in admissions. Fewer than 20 percent of entering undergraduates are transfer students.
  \item Medium full - time, selective or more selective, higher transfer - in  -  60  -  79 percent of undergraduates are enrolled full - time at these bachelor's or higher degree granting institutions. Test score data for first - year students indicate that these institutions are selective or more selective in admissions. At least 20 percent of entering undergraduates are transfer students.
  \item Full - time, inclusive, lower transfer - in  -  At least 80 percent of undergraduates are enrolled full - time at these bachelor's or higher degree granting institutions. These institutions either did not report test score data or the scores indicate that they extend educational opportunity to a wide range of students with respect to academic preparation and achievement. Fewer than 20 percent of entering undergraduates are transfer students.
  \item Full - time, inclusive, higher transfer - in  -  At least 80 percent of undergraduates are enrolled full - time at these bachelor's or higher degree granting institutions. These institutions either did not report test score data or the scores indicate that they extend educational opportunity to a wide range of students with respect to academic preparation and achievement. At least 20 percent of entering undergraduates are transfer students.
  \item Full - time, selective, lower transfer - in  -  At least 80 percent of undergraduates are enrolled full - time at these bachelor's or higher degree granting institutions. Test score data for first - year students indicate that these institutions are selective in admissions (40th to 80th percentile of selectivity among all baccalaureate institutions). Fewer than 20 percent of entering undergraduates are transfer students.
  \item Full - time, selective, higher transfer - in  -  At least 80 percent of undergraduates are enrolled full - time at these bachelor's or higher degree granting institutions. Test score data for first - year students indicate that these institutions are selective in admissions (40th to 80th percentile of selectivity among all baccalaureate institutions). At least 20 percent of entering undergraduates are transfer students.
  \item Full - time, more selective, lower transfer - in  -  At least 80 percent of undergraduates are enrolled full - time at these bachelor's or higher degree granting institutions. Test score data for first - year students indicate that these institutions are more selective in admissions (80th to 100th percentile of selectivity among all baccalaureate institutions). Fewer than 20 percent of entering undergraduates are transfer students.
  \item Full - time, more selective, higher transfer - in  -  At least 80 percent of undergraduates are enrolled full - time at these bachelor's or higher degree granting institutions. Test score data for first - year students indicate that these institutions are more selective in admissions (80th to 100th percentile of selectivity among all baccalaureate institutions). At least 20 percent of entering undergraduates are transfer students.
}
}

\item{CBSA Type Metropolitan or Micropolitan [CBSATYPE]}{Indicates whether the CBSA is a metropolitan or micropolitan statistical area.

Metropolitan Statistical Area  -  A Core Based Statistical Area associated with at least one urbanized area that has a population of at least 50,000.

Micropolitan Statistical Area  -  A Core Based Statistical Area associated with at least one urban cluster that has a population of at least 10,000, but less than 50,000.

For more information see http://www.whitehouse.gov/omb/inforeg/statpolicy.html#ms.}

\item{Combined Statistical Area (CSA) [CSA]}{3 - digit code that identifies the Combined Statistical Area, that an institutions's address is associated with.  If an institution is not associated with a CSA the value of this variable is not applicable ( - 2).

A CSA is a geographic entity consisting of two or more adjacent Core Based Statistical Areas (CBSAs).  A Combined Statistical Area may comprise two or more Metropolitan Statistical Areas, a Metropolitan Statistical Area and a Micropolitan Statistical Area, two or more Micropolitan Statistical Areas, or multiple Metropolitan and Micropolitan Statistical Areas that have social and economic ties as measured by commuting, but at lower levels than are found among counties within Metropolitan and Micropolitan Statistical Areas.

For more information see http://www.whitehouse.gov/omb/inforeg/statpolicy.html#ms.}

\item{Control of institution [CONTROL]}{A classification of whether an institution is operated by publicly elected or appointed officials or by privately elected or appointed officials and derives its major source of funds from private sources.
\itemize{
  \item Public institution  -  An educational institution whose programs and activities are operated by publicly elected or appointed school officials and which is supported primarily by public funds.
  \item Private not - for - profit institution  -  A private institution in which the individual(s) or agency in control receives no compensation, other than wages, rent, or other expenses for the assumption of risk. These include both independent not - for - profit schools and those affiliated with a religious organization.
  \item Private for - profit institution  -  A private institution in which the individual(s) or agency in control receives compensation other than wages, rent, or other expenses for the assumption of risk.
  }}

\item{Core Based Statistical Area (CBSA) [CBSA]}{5 - digit code that identifies the Core Based Statistical Area,that an institution's address is associated with.  If an institution is not associated with a CBSA the value of this variable is not applicable ( - 2).

A CBSA is a geographic entity consisting of the county or counties associated with at least one core (urbanized area or urban cluster) of at least 10,000 population, plus adjacent counties having a high degree of social and economic integration with the core as measured through commuting ties with the counties containing the core.  }

\item{County name [COUNTYNM]}{County name  -   based on the location of the school.}

\item{Date institution closed [CLOSEDAT]}{Date institution closed  -  format is mmddyyyy.}

\item{Degree of urbanization (Urban - centric locale) [LOCALE]}{Locale codes identify the geographic status of a school on an urban continuum ranging from “large city” to “rural.”  They are based on a school's physical address. The urban - centric locale codes introduced in this file are assigned through a methodology developed by the U.S. Census Bureau's Population Division in 2005.  The urban - centric locale codes apply current geographic concepts to the original NCES locale codes used on IPEDS files through 2004.
\itemize{
  \item 11 = City: Large: Territory inside an urbanized area and inside a principal city with population of 250,000 or more.
  \item 12 = City: Midsize: Territory inside an urbanized area and inside a principal city with population less than 250,000 and greater than or equal to 100,000.
  \item 13 = City: Small: Territory inside an urbanized area and inside a principal city with population less than 100,000.
  \item 21 = Suburb: Large: Territory outside a principal city and inside an urbanized area with population of 250,000 or more.
  \item 22 = Suburb: Midsize: Territory outside a principal city and inside an urbanized area with population less than 250,000 and greater than or equal to 100,000.
  \item 23 = Suburb: Small: Territory outside a principal city and inside an urbanized area with population less than 100,000.
  \item 31 = Town: Fringe: Territory inside an urban cluster that is less than or equal to 10 miles from an urbanized area.
  \item 32 = Town: Distant: Territory inside an urban cluster that is more than 10 miles and less than or equal to 35 miles from an urbanized area.
  \item 33 = Town: Remote: Territory inside an urban cluster that is more than 35 miles of an urbanized area.
  \item 41  -  Rural: Fringe: Census - defined rural territory that is less than or equal to 5 miles from an urbanized area, as well as rural territory that is less than or equal to 2.5 miles from an urban cluster.
  \item 42 = Rural: Distant: Census - defined rural territory that is more than 5 miles but less than or equal to 25 miles from an urbanized area, as well as rural territory that is more than 2.5 miles but less than or equal to 10 miles from an urban cluster.
  \item 43 = Rural: Remote: Census - defined rural territory that is more than 25 miles from an urbanized area and is also more than 10 miles from an urban cluster.
}

American Samoa, the Commonwealth of the Northern Mariana Islands, Guam, and the Virgin Islands,  were not assigned a locale code because the geographic and governmental structures of these entities do not fit the definitional scheme used to derive the code.}

\item{Degree - granting status [DEGGRANT]}{A code indicating the degree - granting status of the institution.

Degree - granting institutions offer an associate's, bachelor's, master's, doctor's or a first - professional degree.

Nondegree - granting offers certificates or other formal awards.}

\item{Dun and Bradstreet numbers [DUNS]}{The Data Universal Numbering System (DUNS), is a proprietary system developed and regulated by Dun & Bradstreet (D&B) that assigns a unique numeric identifier, referred to as a "DUNS number" to a single business entity. It was introduced in 1963 to support D&B's credit reporting practice. The DUNS number is a 9 - digit numeric. Some institutions have multiple DUNS numbers.  Multiple DUNS numbers are separated by the vertical bar.}

\item{Employer Identification Number [EIN]}{The number assigned to an institution by the Internal Revenue Service for tax purposes.}

\item{Fall cohort}{The group of students entering in the fall term established for tracking purposes. For the Graduation Rates component, this includes all students who enter an institution as full - time, first - time degree or certificate - seeking undergraduate students during the fall term of a given year.
}

\item{Fips County code [COUNTYCD]}{FIPS county number (two digit FIPS state numeric code + three digits FIPS county code) based on the location of the school.}

\item{Full - year cohort }{The group of students entering at any time during the 12 - month period September 1 through August 31 that is established for tracking and reporting Graduation Rate (GRS) data for institutions that primarily offer occupational programs of varying lengths. Students must be full time and first time to be considered in the cohort.}

\item{General information telephone number [GENTELE]}{General information telephone number.
\itemize{
  \item  - 1  -  Not reported
  \item  - 2  -  Not applicable
  \item  - 9  -  not active
  }}

\item{Graduate offering [GROFFER]}{A code indicating whether the institution offers graduate degrees or certificates.
Graduate degrees or certificates include master's and doctor's degrees and postbaccalaureate and post - master's certificates}

\item{Highest degree offered [HDEGOFR1]}{This variable was revised in 2009 because the first professional category is no longer an award level and the Doctoral award level category has expanded to 3 categories, Doctor's degree  -  Research/scholarships, Doctor's degree  -  professional practice and Doctor's degree - other.

This variable is derived from the level of offering variables (LEVEL1 - LEVEL8 LEVEL12 and LEVEL17 - LEVEL19) from the Institutional Characteristics component as follows:
\itemize{
  \item IF LEVEL17=1 and LEVEL18=1 then HDEGOFR1=11
  \item else if LEVEL17=1 and LEVEL18 ne 1 then HDEGOFR1=12
  \item else if LEVEL17 ne 1 and LEVEL18=1 then HDEGOFR1=13
  \item else if LEVEL17 ne 1 and LEVEL18 ne 1 and LEVEL19=1 then HDEGOFR1=14
  \item else if LEVEL7=1  HDEGOFR1=20
  \item else if LEVEL5=1  HDEGOFR1=30
  \item else if LEVEL3=1  HDEGOFR1=40
  \item else if LEVEL1=1 or LEVEL2=1 or LEVEL4=1 or LEVEL6=1 or LEVEL8=1
  \item or LEVEL12=1 then HDEGOFR1=0
  \item else HDEGOFR1= - 3;
}

For more information see descriptions of LEVEL1 - LEVEL12}

\item{Highest level of offering [HLOFFER]}{Highest level of offering (generated, based on response to IC survey)

\itemize{
  \item 0  -  Other
  \item 1  -  Postsecondary award, certificate or diploma of less than one academic year
  \item 2  -  Postsecondary award, certificate or diploma of at least one but less than two academic years
  \item 3  -  Associate's degree
  \item 4  -  Postsecondary award, certificate or diploma of at least two but less than four academic years
  \item 5  -  Bachelor's degree
  \item 6  -  Postbaccalaureate certificate
  \item 7  -  Master's degree
  \item 8  -  Post - master's certificate
  \item 9  -  Doctor's degree
  \item b  -  None of the above or no answer
  \item - 2  -  Not applicable, first - professional only
  \item - 3  -  Not Available
}
}

\item{Historically Black College or University [HBCU]}{A code to indicate whether the institution is one of the Historically Black College or University (HBCU) institutions.

Historically Black Colleges and Universities (HBCU)  -  The Higher Education Act of 1965, as amended, defines an HBCU as: "...any historically black college or university that was established prior to 1964, whose principal mission was, and is, the education of black Americans, and that is accredited by a nationally recognized accrediting agency or association determined by the Secretary [of Education] to be a reliable authority as to the quality of training offered or is, according to such an agency or association, making reasonable progress toward accreditation." Federal regulations (20 USC 1061 (2)) allow for certain exceptions to the founding date.}

\item{Identification number of multi - institution or multi - campus organization [F1SYSCOD]}{Identification number of multi - institution or multi - campus organization. (See "Multi - institution or multi - campus organization")}

\item{Institution grants a medical degree [MEDICAL]}{A code to indicate whether the institution grants a medical degree.
This variable is derived from the 2012 - 13 completions component of the IPEDS survey.

Medical degrees include:
\itemize{
  \item Medicine (M.D.) (51.1201),
  \item Dentistry (D.D.S., D.M.D.) (51.0401)
  \item Osteopathic Medicine (D.O.) (51.1901)
  \item Veterinary Medicine (D.V.M.) (51.2401)
}
}

\item{Institution has hospital [HOSPITAL]}{A code to indicate whether the institution has hospital.

This variable was derived using Finance data.
If institution reported either revenues from Sales and services of hospitals GASB, hospital revenues (FASB), or expenses for hospital services (GASB or FASB) an institution was determined to have a hospital.  See glossary terms below.


Sales and services of hospitals (revenues) F1B06 - GASB.  Revenues (net of discounts, allowances, and provisions for uncollectible accounts receivable) generated by hospitals from daily patient, special and other services. Revenues of health clinics that are part of a hospital should be included in this category, unless such clinics are part of the student health services program.

Hospitals (revenues) F2D13 - FASB.  Revenues generated by a hospital operated by the postsecondary institution. Includes gifts, grants, appropriations, research revenues, endowment income, and revenues of health clinics that are part of the hospital unless such clinics are part of the student health services program. Sales and service revenues are included net of patient contractual allowances. Revenues associated with the medical school are included elsewhere. Also includes all amounts appropriated by governments (federal, state, local) for the operation of hospitals.

Hospital services (F1C121 - GASB, F2E091 - FASB).  Expenses associated with a hospital operated by the postsecondary institution (but not as a component unit) and reported as a part of the institution. This classification includes nursing expenses, other professional services, general services, administrative services, and fiscal services. Also included are information technology expenses, actual or allocated costs for operation and maintenance of plant, and depreciation related to hospital capital assets. FASB institutions also charge or allocate interest expense to hospitals.}

\item{Institution is active in current year [CYACTIVE]}{A code to indicate whether the institution is active in current year.  This code is derived from the more detailed ACT (status of institution) variable.

Active school CYACTIVE=1 includes:
\itemize{
  \item ACT=A an open (active) institution
  \item ACT=N an institution that opened in the current year
  \item ACT=M an institution that closed in the current year, but will have data from the previous academic year for example Finance or Completions data for academic year 2016 - 17
  \item ACT=R an institution that was previously not active and was restored in the current year
  \item CYACTIVE=2 includes potential adds and restores (ACT=P,R)
  \item CYACTIVE=3 includes all other inactive institutions
  \item ACT=C combined with another active institution
  \item ACT=D delete institution is out of business
  \item ACT=O institution is out - of - scope
  \item ACT=W potential add that was not added
  \item ACT=X potential restore that was not restored
}
}

\item{Institution name alias [IALIAS]}{This is a character string field that contains aliases that an institution can be referenced as. The aliases were submitted by the institution.}

\item{Institution open to the general public [OPENPUBL]}{A code to indicate whether the institution is open for admission to the general public.}

\item{Institution size category [INSTSIZE]}{Institution size category based on total students enrolled for credit, Fall of the current year}

\item{Institutional category [INSTCAT]}{Institutional category was derived using the level of offerings reported on the Institutional Characteristics (IC) component and the number and level of awards that were reported on the Completions (C) component.

Category descriptions:

1) Degree - granting, graduate with no undergraduate degrees  -  These institutions offer a Master's degree, Doctor's degree  or a First - professional degree and do not offer a Bachelor's degree or an Associate's degree.

2) Degree - granting, primarily baccalaureate or above  -  These institutions offer a Bachelor's degree, Master's degree,Doctor's degree or a First - professional degree.  Also, the total number of degrees/certificates at or above the bachelor's level awarded divided by the total number of degrees/certificates awarded is greater than 50 percent.

3) Degree - granting, not primarily baccalaureate or above  -  These institutions offer a Bachelor's degree, Master's degree, Doctor's degree,or a First - professional degree.  Also, the total number of degrees/certificates at or above the bachelor's level awarded divided by the total number of degrees/certificates awarded must be less than or equal to 50 percent.

4) Degree - granting, Associate's and certificates  -  Institutions offer an Associate's degree and may offer other postsecondary certificates, awards or diplomas of less than one academic year; at least one but less - than two academic years; at least two but less - than four academic years. This category also includes institutions that offer a postbaccalaureate certificate, Post - master's certificate or a First - professional certificate and the highest degree offered is an Associate's degree.

5) Nondegree - granting, above the baccalaureate  -  Institutions do not offer Associate's, Bachelor's, Master's, Doctor's or First - professional degrees, but offer either Postbaccaulaureate, Post - master's or First - professional certificates.

6) Nondegree - granting, sub - baccalaureate  -  Institutions do not offer Associate's, Bachelor's , Master's, Doctor's, or First - professional degrees, or certificates above the baccalaureate level. They do offer postsecondary certificates, awards or diplomas of less than one academic year; at least one but less than two academic years; or at least two but less than four academic years.

Technical details for Institutional Category (INSTCAT):

- Total degrees and certificates and total bachelor's degrees and all other degrees/certificates above the bachelor's degree awarded are derived using data from the Completions component.

Total bachelor's degrees and all other degrees/certificates above the bachelor's degree is the sum of:
\itemize{
  \item Bachelor's degrees (CTOTALT,AWLEVEL=5,MAJORNUM=1)
  \item Postbaccalaureate certificates (CTOTALT,AWLEVEL=6,MAJORNUM=1)
  \item Master's degrees (CTOTALT,AWLEVEL=7,MAJORNUM=1)
  \item Post - master's certificates (CTOTALT,AWLEVEL=8,MAJORNUM=1)
  \item Doctor's degrees (CTOTALT,AWLEVEL=9,MAJORNUM=1)
  \item First - professional degrees (CTOTALT,AWLEVEL=10,MAJORNUM=1)
  \item First - professional certificates (CTOTALT,AWLEVEL=11,MAJORNUM=1)
}

Total degrees and certificates is the sum of Bachelor degrees and all other degrees/certificates above the bachelors degree as defined above added to the following:
\itemize{
  \item Associate's degrees (CTOTALT,AWLEVEL=3,MAJORNUM=1)
  \item Postsecondary certificates, awards or diplomas of at least two but less - than four academic years
   (CTOTALT,AWLEVEL=4, MAJORNUM=1).
  \item Postsecondary certificates, awards or diplomas of at least one but less - than two academic years
   (CTOTALT,AWLEVEL=2,MAJORNUM=1);
  \item Postsecondary certificates, awards or diplomas of less than one academic year
    (CTOTALT,AWLEVEL=1,MAJORNUM=1)
}

Institutions are classified as Degree - granting (DEGGRANT=1) or Nondegree - granting(DEGGRANT=2) using level of offerings data reported on the Institutional Characteristics component. Degree - granting institutions offer an Associate's (LEVEL3=1),Bachelor's (LEVEL5=1, Master's (LEVEL7=1), Doctoral (LEVEL9=1), or a First - Professional (LEVEL10=1) degree. Any institution that offers only certificates are Nondegree - granting.

Nondegree - granting institutions that offer a postbaccalaureate certificate (LEVEL6=1) or a post - master's certificate (LEVEL8=1) or a First - professional certificate (LEVEL11=1) are classified as Nondegree - granting, above the baccalaureate (INSTCAT=5).

Nondegree - granting institutions that only offer certificates of less - than four academic years are classified as Nondegree - granting, sub - baccalaureate (INSTCAT=6).

Degree - granting institutions whose highest degree granted are Associate's (HDEGOFFR=40) are classified as Degree - granting, Associate's and certificates (INSTCAT=4). (There are a few 4 - year institutions that grant Postbaccalaureate or Post - master's or First - professional certificates in this category).

Degree - granting institutions that do not grant a Bachelor's degree (LEVEL5=0) and do not grant an Associate's degree (LEVEL3=0) are classified as Degree - granting, graduate with no undergraduate degrees (INSTCAT=1).

The remaining degree - granting institutions offer a bachelor's degree or an associate's degree, or both.
For these institutions a percent of bachelor's degrees and all other degrees/certificates above the bachelor's degree of  total degrees and certificates is generated. If the percent is greater than 50, institutions are classified as Degree - granting, primarily baccalaureate or above (INSTCAT=2), If the percent is 50 or less, institutions are classified as Degree - granting, not primarily baccalaureate or above (INSTCAT=3).

Inactive institutions (CYACTIVE in (2,3)) and administrative units (Sector=0) were coded as not applicable. All Non - Title IV institutions that did not respond to the  IC or Completions components were coded as not reported.
Degree - granting institutions whose completions data are reported with their parent institution were assigned the same code as the parent institution.
New degree - granting institutions that report offering a bachelor's degree on the current institutional characteristics file for the upcoming academic year, and have not yet reported bachelor's degrees on the current completions that covers the previous academic year  June 1 - July 30 were assigned as follows:

If an institution reported a zero in any program (CIP code) for bachelor's degrees or above (indicating the level was offered) and did not report a zero for any program (CIP code) at any level below the bachelor's, the institution was classified as Degree - granting, primarily baccalaureate or above. If an institution reported a zero in any program (CIP code) at any level below the bachelor's degree and did not report a zero for any program (CIP code) for bachelor's degrees or above, the institution was classified as Degree - granting, not primarily baccalaureate or above.

For institutions that reported zeros for bachelor's degrees of above and for levels below the bachelor's, the maximum number of programs by level was used to determine the primary classification.}


\item{Land Grant Institution [LANDGRNT]}{A land - grant college or university is an institution that has been designated by its state legislature or Congress to receive the benefits of the Morrill Acts of 1862 and 1890. The original mission of these institutions, as set forth in the first Morrill Act, was to teach agriculture, military tactics, and the mechanic arts as well as classical studies so that members of the working classes could obtain a liberal, practical education.

Over the years, land - grant status has implied several types of federal support. The first Morrill Act provided grants in the form of federal lands to each state for the establishment of a public institution to fulfill the act's provisions. At different times money was appropriated through legislation such as the second Morrill Act and the Bankhead - Jones Act, although the funding provisions of these acts are no longer in effect. Today, the Nelson Amendment to the Morrill Act provides a permanent annual appropriation of $50,000 per state and territory.

A key component of the land - grant system is the agricultural experiment station program created by the Hatch Act of 1887. The Hatch Act authorized direct payment of federal grant funds to each state to establish an agricultural experiment station in connection with the land - grant institution there. The amount of this appropriation varies from year to year and is determined for each state through a formula based on the number of small farmers there. A major portion of the federal funds must be matched by the state.

To disseminate information gleaned from the experiment stations' research, the Smith - Lever Act of 1914 created a Cooperative Extension Service associated with each U.S. land - grant institution. This act authorized ongoing federal support for extension services, using a formula similar to the Hatch Act's to determine the amount of the appropriation. This act also requires that the states provide matching funds in order to receive the federal monies.


Passage of the First Morrill Act (1862) reflected a growing demand for agricultural and technical education in the United States. While a number of institutions had begun to expand upon the traditional classical curriculum, higher education was still widely unavailable to many agricultural and industrial workers. The Morrill Act was intended to provide a broad segment of the population with a practical education that had direct relevance to their daily lives.

The Second Morrill Act (1890) sought to extend access to higher education by providing additional endowments for all land - grants, but prohibiting distribution of money to states that made distinctions of race in admissions. However, states that provided a separate land - grant institution for blacks were eligible to receive the funds. The institutions that, as a result of this act, were founded or designated the land - grant for blacks in each of the then - segregated Southern states came to be known as "the 1890 land - grants."  Native American tribal colleges are sometimes called the "1994 land - grants."

For more information on Land Grant Institutions see www.nasulgc.org/publications/Land_Grant/land.htm}

\item{Latitude location of institution [LATITUDE]}{Based on the location of the school, the value of LATITUDE ranges from  - 14 to 71. It contains an explicit decimal point. The digits to the left of the decimal represent the number of degrees from the equator; the digits to the right of the decimal represent the fraction of the next degree carried out to six decimal places.}

\item{Level of institution [ICLEVEL]}{A classification of whether an institution's programs are 4 - year or higher (4 year), 2 - but - less - than 4 - year (2 year), or less than 2 - year.}

\item{Longitude location of institution [LONGITUD]}{Based on the location of the school, the value of LONGITUD ranges from  - 170 to 171. The minus sign ( - ) indicates west of the prime meridian. It contains an explicit decimal point. The digits to the left of the decimal point represent the number of degrees from the prime meridian; the digits to the right of the decimal point represent the fraction of the next degree carried out to six decimal places.}

\item{Multi - institution or multi - campus organization [F1SYSTYP]}{Is the institution part of a multi - institution or multi - campus organization that owns, governs, or controls the institution? (See "Multi - institution or multi - campus organization")
}

\item{Multi - institution or multi - campus organization}{A multi - institution or multi - campus organization includes organizations with two or more institutions or campuses.

Non - postsecondary education agencies that govern or control institutions include, but are not limited to, public school districts, art organizations, hospitals and other medical/health organizations.

Not included are:coordinating systems, single institution owner, single institution corporate name, single institution governing board, consortia, associations, religious affiliation (requested in control question).}

\item{Name of chief administrator [CHFNM]}{Name of chief administrator of institution.
\itemize{
  \item  - 1  -  Not reported
  \item  - 2  -  Not applicable
  \item  - 9  -  not active
  }}

\item{Name of multi - institution or multi - campus organization [F1SYSNAM]}{Is the institution part of a multi - institution or multi - campus organization that owns, governs, or controls the institution? (See "Multi - institution or multi - campus organization")}

\item{Net price calculator web address [NPRICURL]}{Net Price Calculator Web Address  for institutions that enroll full - time, first - time undergraduate students}

\item{New England City and Town Area (NECTA) [NECTA]}{5 - digit code that identifies the Core Based Statistical Area in New England, that an institution's address is associated with.  If an institution is not associated with a NECTA the value of this variable is not applicable ( - 2).

A NECTA is a statistical geographic entity that is defined using cities and towns as building blocks and that is conceptually similar to the Core Based Statistical Areas in New England (which are defined using counties as building blocks).

For more information see http://www.whitehouse.gov/omb/inforeg/statpolicy.html#ms.}

\item{Office of Postsecondary Education (OPE) ID Number [OPEID]}{Identification number used by the U.S. Department of Education's Office of Postsecondary Education (OPE) to identify schools that have Program Participation Agreements (PPA) so that its students are eligible to participate in Federal Student Financial Assistance programs under Title IV regulations. This is a 6 - digit number followed by a 2 - digit suffix used to identify branches, additional locations, and other entities that are part of the eligible institution.}

\item{OPE Title IV eligibility indicator code [OPEFLAG]}{Code indicating the institution's degree of eligibility for Title IV aid.}

\item{Postsecondary and Title IV institution indicator [PSET4FLG]}{A derived indicator to easily select institutions who are:
\itemize{
  \item (1) currently in the IPEDS universe,
  \item (2) open to the public,
  \item (3) participant in Title IV program,
  \item (4) primarily postsecondary institution.
}

This variable is used to produce tables that are published in IPEDS First look reports(previously known as E.D. tabs) and tables in the Digest of Education Statistics.
The  "First Look" universe of institutitons includes Title IV primarily postsecondary and Title IV not primarily postsecondary institutions ( PSET4FLG in (1,3)).  Prior to IPEDS 2010 data collection the "First Look/E.D. Tabs" universe included Title IV  primarily postsecondary institutions ( PSET4FLG =1).

It is derived as follows:
\itemize{
  \item If opeflag in (1 2 7 ) and pseflag=1 and openpubl=1 then pset4flg=1
  \item If opeflag in (3 4 5 6) and pseflag=1 and openpubl=1 then pset4flg=2
  \item If opeflag in (1 2 7) and  pseflag=2 and openpubl=1 then pset4flg=3
  \item If opeflag in (3 4 5 6) and pseflag=2 and openpubl=1 then pset4flg=4
  \item If opeflag in (3 4 5 6) and pseflag=1 and openpubl=2 then pset4flg=6
  \item If cyactive=2 then pset4flg=9
  \item
}

The  United States  Air Force, Naval and  Military Academies  are also included as PSET4FLG=1, even though they are opeflag=3, deferment only. The United States Coast Guard, a nonTitle iv school (OPEFLAG=5)  is also included .}

\item{Postsecondary education}{The provision of a formal instructional program whose curriculum is designed primarily for students who are beyond the compulsory age for high school. This includes programs whose purpose is academic, vocational, and continuing professional education, and excludes avocational and adult basic education programs.}

\item{Postsecondary institution indicator [PSEFLAG]}{Identifies an institution whose primary purpose is to provide postsecondary education,is open to the general public and is currently an active (operating) institution.}

\item{Primarily postsecondary indicator [POSTSEC]}{Identifies an institution which has as its sole purpose or one of its primary missions, the provision of postsecondary education.   (See "Postsecondary education")
}

\item{Reporting method for student charges, graduation rates, retention rates and student financial aid [RPTMTH]}{Reporting method identifies how institutions report student charges (full academic year or by program)  and the type of cohort (fall or full - year) used to report graduation rate, retention rate and student financial aid data.

Institutions with standard academic terms (semester, quarter, trimester or 4 - 1 - 4  -  CALSYS in (1,2,3,4)) report student charges for the full academic year and report graduation rate, retention rate and student financial aid data for a FALL cohort.

Institutions with other academic calendar systems (CALYSYS=5) report student charges for the full academic year  and report graduation rate, retention rate and student financial aid for a FULL - YEAR cohort. Other academic year is a category used to describe "non - traditional" calendar systems at 4 - year and 2 - year degree - granting institutions. These can include schools that offer primarily on - line courses or "one course at a time."

Institutions with calendar systems that differ by program or operate on a continuous basis (e.g., every 2 weeks, monthly or other period  -  CALSYS in (6,7)) report student charges by program for the entire length of program and report graduation rate, retention rate and student financial aid for a FULL - YEAR cohort.}

\item{Sector of institution [SECTOR]}{One of nine institutional categories resulting from dividing the universe according to control and level. Control categories are public, private not - for - profit, and private for - profit. Level categories are 4 - year and higher (4 year), 2 - but - less - than 4 - year (2 year), and less than 2 - year. For example: public, 4 - year institutions. (See "Control of institution", "Level of institution")
}

\item{State and 114TH Congressional District ID [CNGDSTCD]}{Congressional district code based on the location of the school. FIPS numeric code for the congressional districts that are legislatively defined subdivision of the state for the purpose of electing representatives to the House of Representative of the United States Congress. The first two digits are the FIPS state numeric code, which makes the congressional district code unique across states The Congressional Districts Codes are two - digit numeric codes used to represent the congressional districts of each multi - district state of the United States. For example, the First Congressional District is identified as “01,” the Second Congressional District as “02,” etc. The Congressional District in a state with only a single representative elected “at large” is designated as “00.” In the 114h Congress, this applies to Alaska, Delaware, Montana, North Dakota, South Dakota, Vermont, and Wyoming. For an entity with a nonvoting delegate—the District of Columbia, American Samoa, Guam, Puerto Rico (whose delegate is referred to as a “resident commissioner”), and the U.S. Virgin Islands—the representational area is designated as “98.” Those entities with no representation in the Congress—the Northern Mariana Islands, Palau, and the several U.S. minor outlying islands are designated as “ - 2.”}

\item{Status of institution [ACT]}{A code to indicate the status of institution's participation in IPEDS.}

\item{Title of chief administrator [CHFTITLE]}{Title of chief administrator of institution.
\itemize{
  \item  - 1  -  Not reported
  \item  - 2  -  Not applicable
  \item  - 9  -  not active
  }}

\item{Tribal college [TRIBAL]}{A code to indicate whether the institution is one of the Tribal Colleges and Universities. These institutions, with few exceptions, are tribally controlled and located on reservations. They are all members of the American Indian Higher Education Consortium.}

\item{Undergraduate offering [UGOFFER]}{A code indicating whether the institution offers undergraduate degrees or certificates.
Undergraduate degrees or certificates include associate's and bachelor's degrees, and certificates that require less - than 4 academic years of study.}

\item{UNITID for merged schools [NEWID]}{UNITID where data is found for merged schools.}

\item{Veterans and Military Servicemembers tuition policies web address [VETURL]}{Does an institution post tuition policies specifically related to Veterans and Military Servicemembers on a website?  Provide URL}

\item{ZIP code [ZIP]}{If includes ZIP+4, does not include dash ( - ), e.g., 060102301. If ZIP+4 not reported, the last four positions will be blank}

  }
}

