\name{ic}
\alias{ic}

\title{Institutional Characteristics: IC}

\description{
This file contains data on program and award level offerings, control and affiliation of institution. It also contains information on special learning opportunities, student services, disability services, tuition plans and athletic conference associations. Services and programs for service members and veterans are also included.
}

\section{Admissions and test score}{
Admissions and test score data, which has been part of the Institutional Characteristics survey component since 2001-02, was moved to the new ADM survey component in 2014 and will be released with other survey components of the IPEDS winter collection module.
}


\section{Glossary}{
\describe{
  \item{Academic [PEO2ISTR]}{Academic, leading to a certificate, degree, or diploma is offered by the institution.

Academic Program - Instructional program leading toward an associate's, bachelor's, master's, doctor's, or first-professional degree or resulting in credits that can be applied to one of these degrees. }

\item{Academic Library}{An entity in a postsecondary institution that provides an organized collection of printed or other materials, or a combination thereof; a staff trained to provide and interpret such materials as required to meet the informational, cultural, recreational, or educational needs of the clientele; an established schedule in which services of the staff are available to the clientele; an established schedule in which services of the staff are available to the clientele; and the physical facilities necessary to support such a collection, staff, and schedule. This definition includes libraries that are part of learning resource centers.}

\item{Academic/career counseling service [STUSRV2]}{Activities designed to assist students in making plans and decisions related to their education, career, or personal development.}

\item{Adult basic remedial or high school equivalent [PEO5ISTR]}{Adult basic or remedial instruction or high school equivalency  is offered by the institution.

Adult Basic Education - Courses designed primarily for students 16 years of age and older to improve basic skills in reading, writing, and arithmetic. These courses are not intended to be part of a program leading to a high school degree, nor are they part of any academic, occupational, or vocational program at the postsecondary level. }

\item{Advanced placement (AP) credits [CREDITS3]}{Advanced placement courses are college-level courses taught in high school. Students may take an examination at the completion of the course; acceptable scores allow students to earn college credit.}

\item{Any alternative tuition plans offered by institution [TUITPL]}{Indicates whether or not any of the following alternative tuition plans are offered by an institution.
\itemize{
  \item Tuition guarantee
  \item Prepaid tuition plan
  \item Tuition payment plan
  \item Other
}
}

\item{Application fee}{That amount of money that an institution charges for processing a student's application for admittance to the institution. This amount is not creditable toward tuition or required fees, nor is it refundable if the student is not admitted to the institution.
}

\item{Associate's degree [LEVEL3]}{Associate's degree is offered by an institution.

Associate's degree - An award that normally requires at least 2 but less than 4 years of full-time equivalent college work.}

\item{Bachelor's degree [LEVEL5]}{Bachelor's degree or equivalent is offered by an institution

Bachelor's degree - An award (baccalaureate or equivalent degree, as determined by the Secretary, U.S. Department of Education) that normally requires at least 4 but not more than 5 years of full-time equivalent college-level work. This includes all bachelor's degrees conferred in a 5-year cooperative (work-study plan) program. A cooperative plan provides for alternate class attendance and employment in business, industry, or government; thus, it allows students to combine actual work experience with their college studies. Also, includes bachelor's degrees in which the normal 4 years of work are completed in 3 years.}

\item{Board charges}{The charge for an academic year for meals, for a specified number of meals per week.}

\item{Calendar system [CALSYS]}{Standard academic terms. Checking one of these systems determines that an institution will provide graduation rates data based on a FALL COHORT and student charges based on a FULL ACADEMIC YEAR
\itemize{
  \item Semester (See "Semester calendar system")
  \item Quarter (See "Quarter calendar system")
  \item Trimester  (See "Trimester calendar system")
  \item 4-1-4 or similar plan (See "Four-one-four plan")
  \item Other academic calendar (See "Other academic calendar system")
}

Other calendar systems (for primarily occupational schools). Checking one of the following determines that an institution will provide graduation rates data based on a FULL YEAR COHORT and student charges data will be requested by PROGRAM.
\itemize{
  \item Differs by program (See "Differs by program (calendar system)")
  \item Continuous basis (every 2 weeks, monthly, or other period)  (See "Continuous basis (for program enrollment)")
}}

\item{Combined charge for room and board [RMBRDAMT]}{What are the typical room and board charges for a student for the full academic year?
 (applicable only if institution CANNOT separate room and board charges.)}

\item{Continuing professional [PEO3ISTR]}{Continuing professional (postbaccalaureate only) is offered by the institution.

Continuing Professional Education - Programs and courses designed specifically for individuals who have completed a professional degree (such as law, medicine, dentistry, or social work) to obtain additional training in their particular field of study }

\item{Continuous basis (for program enrollment)}{A calendar system classification that is used by institutions that enroll students at any time during the academic year. For example, a cosmetology school or a word processing school might allow students to enroll and begin studies at various times, with no requirement that classes begin on a certain date.}

\item{Credit for life experiences [CREDITS2]}{Credit earned by students for what they have learned through independent study, noncredit adult courses, work experience, portfolio demonstration, previous licensure or certification, or completion of other learning opportunities (military, government, or professional). Credit may also be awarded through a credit by examination program}

\item{Credit for military training [VET2]}{Postsecondary credit granted by institutions to military servicemen or veterans for experiences and training gained while in the service.}

\item{Defree/Certificate-seeking students}{Students enrolled in courses for credit who are recognized by the institution as seeking a degree or other formal award. At the undergraduate level, this is intended to include students enrolled in vocational or occupational programs.}

\item{Differs by program (calendar system)}{A calendar system classification that is used by institutions that have occupational/vocational programs of varying lengths. These schools may enroll students at specific times depending on the program desired. For example, a school might offer a 2-month program in January, March, May, September, and November; and a 3-month program in January, April, and October. }

\item{Distance education}{Education that uses one or more technologies to deliver instruction to students who are separated from the instructor and to support regular and substantive interaction between the students and the instructor synchronously or asynchronously.

Technologies used for instruction may include the following: Internet; one-way and two-way transmissions through open broadcasts, closed circuit, cable, microwave, broadband lines, fiber optics, satellite or wireless communication devices; audio conferencing; and video cassette, DVDs, and CD-ROMs, if the cassette, DVDs, and CD-ROMs are used in a course in conjunction with the technologies listed above.}

\item{Distance education course}{A course in which the instructional content is delivered exclusively via distance education.  Requirements for coming to campus for orientation, testing, or academic support services do not exclude a course from being classified as distance education.}

\item{Distance education programs offered [DISTPGS]}{Distant education progams offered

Distance education program	 - A program for which all the required coursework for program completion is able to be completed via distance education courses.}

\item{Doctor's degree - other [LEVEL19]}{Doctor's degree - other is offered by an institution.

Doctor's degree - other - A doctor's degree that does not meet the definition of a doctor's degree - research/scholarship or a doctor's degree - professional practice.}

\item{Doctor's degree - professional practice [LEVEL18]}{Doctor's degree - professional practice is offered by an institution.

Doctor's degree - professional practice - A doctor's degree that is conferred upon completion of a program providing the knowledge and skills for the recognition, credential, or license required for professional practice. The degree is awarded after a period of study such that the total time to the degree, including both pre-professional and professional preparation, equals at least six full-time equivalent academic years. Some of these degrees were formerly classified as “first-professional” and may include: Chiropractic (D.C. or D.C.M.); Dentistry (D.D.S. or D.M.D.); Law (L.L.B. or J.D.); Medicine (M.D.); Optometry (O.D.); Osteopathic Medicine (D.O); Pharmacy (Pharm.D.); Podiatry (D.P.M., Pod.D., D.P.); or, Veterinary Medicine (D.V.M.), and others, as designated by the awarding institution.}

\item{Doctor's degree - research/scholarship [LEVEL17]}{Doctor's degree - research/scholarship is offered by an institution.

Doctor's degree - research/scholarship - A Ph.D. or other doctor's degree that requires advanced work beyond the master's level, including the preparation and defense of a dissertation based on original research, or the planning and execution of an original project demonstrating substantial artistic or scholarly achievement. Some examples of this type of degree may include Ed.D., D.M.A., D.B.A., D.Sc., D.A., or D.M, and others, as designated by the awarding institution.}

\item{Dual credit [CREDITS1]}{A program through which high school students are enrolled in advanced placement (AP) courses, taught at their high school, that fulfill high school graduation requirements and may earn the student college credits.}

\item{Employment services for students [STUSRV3]}{Activities intended to assist students in obtaining part-time employment as a means of defraying part of the cost of their education.}

\item{First-time student (Undergraduate)}{A student attending any institution for the first time at the undergraduate level. Includes students enrolled in academic or occupational programs. Also includes students enrolled in the fall term who attended college for the first time in the prior summer term, and students who entered with advanced standing (college credits earned before graduation from high school). }

\item{Four-one-four plan}{The 4-1-4 calendar consists of 4 courses taken for 4 months, 1 course taken for 1 month, and 4 courses taken for 4 months. There may be an additional summer session. }

\item{Full-time graduate}{A student enrolled for 9 or more semester credits, or 9 or more quarter credits, or students involved in thesis or dissertation preparation that are considered full time by the institution.}

\item{Full-time student Undergraduate}{A student enrolled for 12 or more semester credits, or 12 or more quarter credits, or 24 or more contact hours a week each term.}

\item{Graduate student}{A student who holds a bachelor's or first-professional degree, or equivalent, and is taking courses at the post-baccalaureate level. These students may or may not be enrolled in graduate programs }

\item{In-district student}{A student who is a legal resident of the locality in which he/she attends school and thus is entitled to reduced tuition charges if offered by the institution .}

\item{In-state student}{A student who is a legal resident of the state in which he/she attends school. }

\item{Institutional control or affiliation [CNTLAFFI]}{
\itemize{
  \item  Public (See "Public institution")
  \item Private for-profit  (See "Private for-profit (profit-making) insitution")
  \item Private not-for-profit independent (no religious affiliation)  (See "Private not-for-profit institution ")
  \item Private not-for-profit religious affiliation - Specify  (See "Private not-for-profit institution ")
}}

\item{Less than one year certificate [LEVEL1]}{Postsecondary award, certificate, or diploma of less than one academic year (less than 900 contact or clock hours) are offered by an institution.

Postsecondary award, certificate, or diploma (less than 1 academic year) - Requires completion of an organized program of study at the postsecondary level (below the baccalaureate degree) in less than 1 academic year (2 semesters or 3 quarters) or in less than 900 contact hours by a student enrolled full time.}

\item{Master's degree [LEVEL7]}{Master's degree is offered by an institution

Master's degree - An award that requires the successful completion of a program of study of at least the full-time equivalent of 1 but not more than 2 academic years of work beyond the bachelor's degree.}

\item{Member of Servicemembers Opportunity Colleges [VET5]}{A membership group of over 1,700 institutions that functions to expand and improve voluntary postsecondary opportunities for servicemembers worldwide. It is funded by the Department of Defense through a contract with the American Association of State Colleges and Universities (AASCU).}

\item{Number of meals per week in board charge [MEALSWK]}{If board or meal plans are provided, the number of meals per week in the maximum meal plan offered is entered.
If the number of meals per week can vary (for example, student receives a meal card and charges meals against the card) this field is not applicable.}

\item{Occupational [PEO1ISTR]}{Occupational, may lead to a certificate, degree, or other formal award is offered by the institution.

Occupational Program - A program of study consisting of one or more courses, designed to provide the student with sufficient knowledge and skills to perform in a specific occupation.}

\item{On-campus day care for students' children [STUSRV8]}{A student service designed to provide appropriate care and protection of infants, preschool, and school-age children so their parents can participate in postsecondary education programs.}

\item{One but less than two years certificate [LEVEL2]}{Postsecondary award, certificate, or diploma of at least one but less than two academic years (at least 900 but less than 1800 contact or clock hours) are offered by an institution.

Postsecondary award, certificate, or diploma  (at least 1 but less than 2 academic years) -Requires completion of an organized program of study at the postsecondary level (below the baccalaureate degree) in at least 1 but less than 2 full-time equivalent academic years, or designed for completion in at least 30 but less than 60 credit hours, or in at least 900 but less than 1,800 contact hours.}

\item{Open admission policy [OPENADMP]}{Does an institution have an open admission policy for all or most entering first-time undergraduate-level students?

Open admission - Admission policy whereby the school will accept any student who applies.
}

\item{Other academic calendar system}{Category used to describe "non-traditional" calendar systems at 4-year and 2-year degree-granting institutions. These can include schools that offer primarily on-line courses or "one course at a time". }

\item{Other degree [LEVEL12]}{An award that it cannot be classified in one of the standard levels (1-11) offered by an institution is specified .}

\item{Out-of-state student}{A student who is not a legal resident of the state in which he/she attends school.}

\item{Part-time Graduate}{A student enrolled for either 8 semester credits or less, or 8 quarter credits or less.}

\item{Part-time Student Undergraduate}{A student enrolled for either 11 semester credits or less, or 11 quarter credits or less, or less than 24 contact hours a week each term.}

\item{Placement services for completers [STUSRV4]}{Assistance for students in evaluating their career alternatives as well as in obtaining full-time employment upon leaving the institution.}

\item{Postbaccalaureate certificate [LEVEL6]}{Postbaccalaureate certificate is offered by an institution.

Postbaccalaureate certificate  - An award that requires completion of an organized program of study requiring 18 credit hours beyond the bachelor's; designed for persons who have completed a baccalaureate degree, but do not meet the requirements of academic degrees carrying the title of master.}

\item{Post-master's certificate [LEVEL8]}{Post-master's certificate is offered by an institution.

Post-master's certificate - An award that requires completion of an organized program of study of 24 credit hours beyond the master's degree, but does not meet the requirements of academic degrees at the doctor's level.}

\item{Prepaid tuition plan [TUITPL2]}{A program that allows students or their families to purchase college tuition or tuition credits for future years, at current prices.}

\item{Primary public control [PUBPRIME]}{Indicates primary control for public institutions
Public institutions must designate a primary control; identifying a secondary control is optional.
See "Public institution"}

\item{Private for-profit (profit-making) insitution}{A private institution in which the individual(s) or agency in control receives compensation other than wages, rent, or other expenses for the assumption of risk. }

\item{Private not-for-profit institution}{A private institution in which the individual(s) or agency in control receives no compensation, other than wages, rent, or other expenses for the assumption of risk. These include both independent nonprofit schools and those affiliated with a religious organization.}

\item{Public institution}{An educational institution whose programs and activities are operated by publicly elected or appointed school officials and which is supported primarily by public funds.  }

\item{Quarter calendar system}{A calendar system in which the academic year consists of 3 sessions called quarters of about 12 weeks each. The range may be from 10 to 15 weeks. There may be an additional quarter in the summer.}

\item{Recreational or avocational [PEO4ISTR]}{Recreational or avocational are offered by the institution.

Avocational Programs - Instructional programs in personal interest and leisure categories whose expressed intent is not to produce postsecondary credits, nor to lead to a formal award or an academic degree, nor result in occupationally specific skills.}

\item{Religious affiliation [RELAFFIL]}{Indicates religious affiliation (denomination) for private not-for-profit institutions that are religiously affiliated.}

\item{Remedial services [STUSRV1]}{Instructional activities designed for students deficient in the general competencies necessary for a regular postsecondary curriculum and educational setting.}

\item{Room charges}{The charges for an academic year for rooming accommodations for a typical student sharing a room with one other student. }

\item{ROTC [SLO5]}{What types of special learning opportunities are offered by an institution? }

\item{ROTC - Air Force [SLO53]}{What types of special learning opportunities are offered by an institution?
IF ROTC, Indicate if Air Force}

\item{ROTC - Army [SLO51]}{What types of special learning opportunities are offered by an institution?
IF ROTC, Indicate if Army}

\item{ROTC - Navy [SLO52]}{What types of special learning opportunities are offered by an institution? IF ROTC, Indicate if Navy}

\item{Secondary (high school) [PEO6ISTR]}{Secondary (high school) is offered by the institution.}

\item{Secondary public control [PUBSECON]}{Indicates secondary control for public institutions
Public institutions must designate a primary control; identifying a secondary control is optional.
See "Public institution"}

\item{Semester calendar system}{A calendar system that consists of two semesters during the academic year with about 16 weeks for each semester of instruction. There may be an additional summer session. }

\item{Study abroad [SLO6]}{Study abroad offered by an institution

Study abroad - Arrangement by which a student completes part of the college program studying in another country. Can be at a campus abroad or through a cooperative agreement with some other U.S. college or an institution of another country.}

\item{Teacher certification (below the postsecondary level) [SLO8]}{Teacher certification (below the postsecondary level) is offered by an institution.

Teacher certification - Program designed to prepare students to meet the requirements for certification as teachers in elementary, middle/junior high, and secondary schools.}

\item{Total dormitory capacity [ROOMCAP]}{The maximum number of students that the institution can provide residential facilities for, whether on or off campus. (off-campus dormitory space that is reserved by the institution).}

\item{Trimester calendar system}{An academic year consisting of 3 terms of about 15 weeks each. }

\item{Tuition and required fees (Tuition)}{Amount of money charged to students for instructional services. Tuition may be charged per term, per course, or per credit. (Required fees) - Fixed sum charged to students for items not covered by tuition and required of such a large proportion of all students that the student who does NOT pay the charge is an exception. }

\item{Tuition guaranteed plan [TUITPL1]}{A program where the institution guarantees, to entering first-time students, that tuition will not increase for the years they are enrolled. These guarantees are generally time-bound for four or five years.}

\item{Tuition payment plan [TUITPL3]}{A program that allows tuition to be paid in installments spread out over an agreed upon period of time, sometimes without interest or finance charges.}

\item{Two but less than 4 years certificate [LEVEL4]}{Postsecondary award, certificate, or diploma of at least two but less than four academic years (at least 1800 contact or clock hours) is offered by an institution.

Postsecondary award, certificate, or diploma (at least 2 but less tham 4 academic years) requires completion of an organized program of study at the postsecondary level (below the baccalaureate degree) in at least 2 but less than 4 full-time equivalent academic years, or designed for completion in at least 60 but less than 120 credit hours, or in at least 1,800 but less than 3,600 contact hours.}

\item{Undergraduate (student)}{A student enrolled in a 4- or 5-year bachelor's degree program, an associate's degree program, or a vocational or technical program below the baccalaureate. }

\item{Weekend/evening  college [SLO7]}{Weekend college  is offered by an institution.

Weekend/evening College - A program that allows students to take a complete course of study and attend classes only on weekends or only in evenings.}

\item{Years of college-level work required [YRSCOLL]}{If an institution grants a bachelor's degree or higher but does not offer a full 4-year program of study at the undergraduate level, how many years of completed college-level work are required for entrance? }

\item{Yellow Ribbon Program (officially known as Post-9/11 GI Bill, Yellow Ribbon Program) [VET1]}{A voluntary program through which participating public and private institutions can provide veterans and eligible beneficiaries additional institutional aid to cover the costs of tuitions and fees at their institutions. The Yellow Ribbon Program is a supplementary program to the Post 9/11 GI Bill coverage of in-state tuition and fees. The Department of Veterans Affairs matches the institutional aid provided beyond the in-state tuition and fees, but to certain limit each year}
  }
}
