\name{ic_ay}
\alias{ic_ay}

\title{Institutional Characteristics: IC_AY}

\description{
This file contains data on student charges for a full academic year.
}

\section{Calendar system}{
Institutions that offer primarily academic programs measured in credit hours have a predominant calendar system of semester, quarter, trimester, or 4-1-4 or other academic calendar system report student charges for the full academic year.
}

\section{Characteristics}{
Student charges on this data file include: average tuition and required fees for both full-time undergraduate and graduate students; per credit hour charges for both part-time undergraduate and graduate students;  Typical tuition and required fees for a full-time Doctor's - professional practice student in Chiropractic (D.C. or D.C.M.), Dentistry (D.D.S. or D.M.D.), Law (L.L.B. or J.D.), Medicine (M.D.), Optometry (O.D.), Osteopathic Medicine (D.O), Pharmacy (Pharm.D.), Podiatry (D.P.M., Pod.D., D.P.), or, Veterinary Medicine (D.V.M.), and Law (J.D.).

The price of attendance of full-time first-time undergraduates that are made available to the public on College Navigator is included on this file.  Price of attendance includes amounts for published tuition and required fees, books and supplies, room and board and other expenses.  Estimates for books and supplies, room and board and other expenses are those from the Cost of Attendance report used by the financial aid office in determining financial need.
}


\section{Glossary}{
\describe{

\item{Academic year}{The period of time generally extending from September to June; usually equated to 2 semesters or trimesters, 3 quarters, or the period covered by a 4-1-4 plan.}

\item{Books and supplies}{The average cost of books and supplies for a typical student for an entire academic year (or program). Do not include unusual costs for special groups of students (e.g., engineering or art majors), unless they constitute the majority of students at your institution.}

\item{Credit Hour}{A unit of measure representing an hour (50 minutes) of instruction over a 15-week period in a semester or trimester system or a 10-week period in a quarter system. It is applied toward the total number of hours needed for completing the requirements of a degree, diploma, certificate, or other formal award. }

\item{Doctor's professional practice}{A doctor's degree that is conferred upon completion of a program providing the knowledge and skills for the recognition, credential, or license required for professional practice. The degree is awarded after a period of study such that the total time to the degree, including both pre-professional and professional preparation, equals at least six full-time equivalent academic years. Some of these degrees were formerly classified as first-professional and may include: Chiropractic (D.C. or D.C.M.); Dentistry (D.D.S. or D.M.D.); Law (L.L.B. or J.D.); Medicine (M.D.); Optometry (O.D.); Osteopathic Medicine (D.O); Pharmacy (Pharm.D.); Podiatry (D.P.M., Pod.D., D.P.); or, Veterinary Medicine (D.V.M.), and others, as designated by the awarding institution.}

\item{First-time first-year student}{A student attending any institution for the first time at the undergraduate level. Includes students enrolled in the fall term who attended college for the first time in the prior summer term. Also includes students who entered with advanced standing (college credits earned before graduation from high school).}

\item{First-year student}{A student who has completed less than the equivalent of 1 full year of undergraduate work; that is, less than 30 semester hours (in a 120-hour degree program) or less than 900 contact hours. }

\item{Full-time student Graduate}{A student enrolled for 9 or more semester credits, or 9 or more quarter credits, or students involved in thesis or dissertation preparation that are considered full time by the institution.}

\item{Full-time student Undergraduate}{A student enrolled for 12 or more semester credits, or 12 or more quarter credits, or 24 or more contact hours a week each term.}

\item{Graduate student}{A student who holds a bachelor's or first-professional degree, or equivalent, and is taking courses at the post-baccalaureate level. These students may or may not be enrolled in graduate programs.}

\item{In-district fees}{The fees charged by the institution to those students residing in the locality in which they attend school. This may be a lower rate than in-state fees if offered by the institution. }

\item{In-district per credit hour charge}{The amount charged by the institution to those students residing in the locality in which they attend school. This may be a lower rate than in-state amount if offered by the institution. }

\item{In-district tuition}{The tuition charged by the institution to those students residing in the locality in which they attend school. This may be a lower rate than in-state tuition if offered by the institution.
}

\item{In-state fees}{The fees charged by institutions to those students who meet the state's or institution's residency requirements.}

\item{In-state per credit hour charge}{The amount charged by institutions to those students who meet the state's or institution's residency requirements.}

\item{In-state tuition}{The tuition charged by institutions to those students who meet the state's or institution's residency requirements. }

\item{Other expenses}{The amount of money (estimated by the financial aid office) needed by a student to cover expenses such as laundry, transportation, entertainment, and furnishings. (For the purpose of this survey room and board and tuition and fees are not included.) }

\item{Out-of-state fees}{The fees charged by institutions to those students who do not meet the institution's or state's residency requirements.}

\item{Out-of-state per credit hour charge}{The amount charged by institutions to those students who do not meet the state's or institution's residency requirements.}

\item{Out-of-state tuition}{The tuition charged by institutions to those students who do not meet the institution's or state's residency requirements.}

\item{Part-time student Graduate}{A student enrolled for either 8 semester credits or less, or 8 quarter credits or less.}

\item{Part-time student Undergraduate}{A student enrolled for either 11 semester credits or less, or 11 quarter credits or less, or less than 24 contact hours a week each term.}

\item{Required fees}{Fixed sum charged to students for items not covered by tuition and required of such a large proportion of all students that the student who does NOT pay the charge is an exception.}

\item{Tuition}{Amount of money charged to students for instructional services. Tuition may be charged per term, per course, or per credit.}

\item{Undergraduate (student)}{A student enrolled in a 4- or 5-year bachelor's degree program, an associate's degree program, or a vocational or technical program below the baccalaureate. }
  }
}

