\name{ic_py}
\alias{ic_py}

\title{Institutional Characteristics: IC_PY}

\description{
This file contains data on student charges by program.

Institutions that measure programs in clock hours or have primarily occupational programs measured in credit hours report student charges data by the full length PROGRAM.
}

\section{Characteristics}{
This file contains the number of programs offered and  the price of attendance for entering students that are made available to the public on College Navigator; Price of attendance includes amounts for published tuition and required fees, books and supplies, room and board and other expenses for the largest program.  Estimates for books and supplies, room and board and other expenses are those from the Cost of Attendance report used by the financial aid office in determining financial need. In addition to the price of attendance for the largest program, tuition and fees and books and supplies are also included for up to an additional five programs.
}


\section{Glossary}{
\describe{
\item{Board charges}{The charge for an academic year for meals, for a specified number of meals per week.}

\item{Books and supplies}{The average cost of books and supplies for a typical student for an entire academic year (or program). Do not include unusual costs for special groups of students (e.g., engineering or art majors), unless they constitute the majority of students at your institution.}

\item{CIP code of largest program [CIPCODE1]}{A six-digit code in the form xx.xxxx that identifies instructional program specialties within educational institutions}

\item{Contact hour}{A unit of measure that represents an hour of scheduled instruction given to students. Also referred to as clock hour}

\item{Credit hour}{A unit of measure representing the equivalent of an hour (50 minutes) of instruction per week over the entire term. It is applied toward the total number of credit hours needed for completing the requirements of a degree, diploma, certificate, or other formal award.}

\item{First-time first-yeat student}{A student attending any institution for the first time at the undergraduate level. Includes students enrolled in the fall term who attended college for the first time in the prior summer term. Also includes students who entered with advanced standing (college credits earned before graduation from high school).}

\item{First-year student}{A student who has completed less than the equivalent of 1 full year of undergraduate work; that is, less than 30 semester hours (in a 120-hour degree program) or less than 900 contact hours.}

\item{Length of program}{Total length of program in contact hours }

\item{Number of programs offered [PRGMOFR]}{How many programs (measured in contact or clock hours) are offered at an institution? }

\item{Other expenses}{The amount of money (estimated by the financial aid office) needed by a student to cover expenses such as laundry, transportation, entertainment, and furnishings. (For the purpose of this survey room and board and tuition and fees are not included.)}

\item{Program}{A combination of courses and related activities organized for the attainment of broad educational objectives as described by the institution.}

\item{Required fees}{The fixed sum charged to students for items not covered by tuition and required of such a large proportion of all students that the student who does NOT pay the charge is an exception.}

\item{Room charges}{The charges for an academic year for rooming accommodations for a typical student sharing a room with one other student. }

\item{Tuition}{The amount of money charged to students for instructional services. Tuition may be charged per term, per course, or per credit.}

  }
}

