\name{s_cn}
\alias{s_cn}

\title{Fall Staff: S_CN}

\description{
This file contains the number of employees on the payroll of the institution as of November 1,  by primary function/occupation, race/ethnicity and gender.
}

\section{Data Layout}{
This file has multiple records per institution.  Each record is uniquely defined by the variables IPEDS ID (UNITID), and primary function/occupation  (SCNLEVEL).
}

\section{Institutions}{
These data are applicable to degree-granting institutions with less than 15  full-time employees and all nondegree-granting institutions and related administrative offices.
}


\section{Glossary}{
\describe{
\item{Faculty}{Persons identified by the institution as such and typically those whose initial assignments are made for the purpose of conducting instruction, research or public service as a principal activity (or activities). They may hold academic rank titles of professor, associate professor, assistant professor, instructor, lecturer or the equivalent of any of those academic ranks. Faculty may also include the chancellor/president, provost, vice provosts, deans, directors or the equivalent, as well as associate deans, assistant deans and executive officers of academic departments (chairpersons, heads or the equivalent) if their principal activity is instruction combined with research and/or public service. The designation as "faculty" is separate from the activities to which they may be currently assigned. For example, a newly appointed president of an institution may also be appointed as a faculty member. Graduate, instruction, and research assistants are not included in this category.}

\item{Full-time staff (employees) }{As defined by the institution. The type of appointment at the snapshot date determines whether an employee is full time or part time. The employee's term of contract is not considered in making the determination of full or part time. }

\item{Part-time staff (employees)}{As determined by the institution. The type of appointment at the snapshot date determines whether an employee is full time or part time. The employee's term of contract is not considered in making the determination of full or part time. Casual employees (hired on an ad-hoc basis or occasional basis to meet short-term needs) and students in the College Work-Study Program are not considered part-time staff. }

\item{Primary function/occupational activity [SCNLEVEL]}{The principal activity of a staff member as determined by the institution. If an individual participates in two or more activities, the primary activity is normally determined by the amount of time spent in each activity.}
  }
}

