\name{s_f}
\alias{s_f}

\title{Fall Staff: S_F}

\description{
This file contains the number of full-time instruction/research/public service staff on the payroll of the institution as of November 1,  by faculty status, academic rank, race/ethnicity and gender.
}

\section{Data Layout}{
This file has multiple records per institution.  Each record is uniquely defined by the variables IPEDS ID (UNITID), and the variable ARANK which is the combination of faculty status (tenured, on tenure track, and not on-tenure track/no tenure system) and academic rank (professors, associate professors, etc.) .
}

\section{Institutions}{
 These data are applicable to degree-granting institutions with 15 or more full-time employees and related administrative offices.
}


\section{Glossary}{
\describe{

\item{Faculty}{Persons identified by the institution as such and typically those whose initial assignments are made for the purpose of conducting instruction, research or public service as a principal activity (or activities). They may hold academic rank titles of professor, associate professor, assistant professor, instructor, lecturer or the equivalent of any of those academic ranks. Faculty may also include the chancellor/president, provost, vice provosts, deans, directors or the equivalent, as well as associate deans, assistant deans and executive officers of academic departments (chairpersons, heads or the equivalent) if their principal activity is instruction combined with research and/or public service. The designation as "faculty" is separate from the activities to which they may be currently assigned. For example, a newly appointed president of an institution may also be appointed as a faculty member. Graduate, instruction, and research assistants are not included in this category}

\item{Full-time staff (employees) }{As defined by the institution. The type of appointment at the snapshot date determines whether an employee is full time or part time. The employee's term of contract is not considered in making the determination of full or part time. }

\item{Tenure}{Status of a personnel position, or a person occupying a position or occupation, with respect to permanence of the position. }
  }
}

