\name{s_is}
\alias{s_is}

\title{Fall Staff: S_IS}

\description{
This file contains the number of full-time instructional staff on the payroll of the institution as of November 1,  by faculty and tenure status, academic rank, race/ethnicity and gender.
}

\section{Data Layout}{
This file has multiple records per institution.  Each record is uniquely defined by the variables IPEDS ID (UNITID), and the variable SISCAT which is the combination of faculty and tenure status FACSTAT (tenured, on tenure track, and not on-tenure track/no tenure system) and academic rank ARANK  (professors, associate professors, etc.) .
}

\section{Institutions}{
These data are applicable to degree-granting institutions with 15 or more full-time employees and related administrative offices.
}

\section{Amendments}{
Beginning with 2016, staff that are not on tenure track are further disaggregated by contract length (multi-year, indefinite, annual and less-than-annual).  From 2012 through 2015 the multi-year and indefinite contract lengths were combined as one category and often referred to as multi-year in past documentation.
}

\section{Note}{
Prior to the IPEDS Web-based data collection in 2001, human resource data reported by race/ethnicity and gender was only available in odd-numbered years.  Beginning with 2001, human resource data reported by race/ethnicity and gender was mandatory in odd-numbered years and optional in even-numbered years.  Beginning with 2016 reporting human resource data by race/ethnicity and gender is mandatory annually.
}


\section{Glossary}{
\describe{

\item{Academic rank [ARANK]}{Assigned by the institution and includes Professors, Associate professors, Assistant professors, Instructors, Lecturers, and No academic rank.}

\item{Annual contract}{An annually-renewable contract or employment agreement that is in effect for a stated annual period within one year of execution, and may be equal to a period of 365 days, or a standard academic year, or the equivalent. Does not include contracts for partial year periods such as a single semester, quarter, term, block, or course.}

\item{Faculty}{Persons identified by the institution as such and typically those whose initial assignments are made for the purpose of conducting instruction, research or public service as a principal activity (or activities). They may hold academic rank titles of professor, associate professor, assistant professor, instructor, lecturer or the equivalent of any of those academic ranks. Faculty may also include the chancellor/president, provost, vice provosts, deans, directors or the equivalent, as well as associate deans, assistant deans and executive officers of academic departments (chairpersons, heads or the equivalent) if their principal activity is instruction combined with research and/or public service. The designation as "faculty" is separate from the activities to which they may be currently assigned. For example, a newly appointed president of an institution may also be appointed as a faculty member. Graduate, instruction, and research assistants are not included in this category .}

\item{Instructional staff category [SISCAT]}{Defines instructional staff by faculty and tenure status by academic rank}

\item{Less-than-annual-contract}{A contract or employment agreement that is in effect for a partial year period of less than 365 days, or less than a standard academic year or the equivalent. Includes contracts for partial year periods such as a single semester, quarter, term, block, or course. }

\item{Multi-year contract }{A contract or employment agreement that is in effect for more than one year (e.g., more than 365 days) where subsequent years of the contract may be contingent upon other factors such as the appropriation of funds. The renewal period of a multi-year contract is not on an annual basis (e.g., a 5-year multi-year contract is renewed every 5 years NOT annually). }

\item{Not on tenure track}{Personnel positions that are considered non-tenure earning positions }

\item{Tenure}{Status of a personnel position, or a person occupying a position or occupation, with respect to permanence of the position.}

\item{Tenure track}{Personnel positions that lead to consideration for tenure.  }
  }
}

