\name{s_nh}
\alias{s_nh}

\title{Fall Staff: S_NH}

\description{
This file contains the number of full-time new hires by primary occupation, race/ethnicity and gender.
}

\section{New Hires}{
New hires include persons who were hired for full-time permanent employment for the first time, or after a break in service, between November 1 of last year - October 31 of current year). All new hire data prior to the  Last year Fall data file, used the four month period of July 1st - October 31st to define a new hire.
}

\section{Instructional Staff}{
New hires for instructional staff occupation are also reported by faculty/tenure status. Beginning with 2016, instructional staff that are not on tenure track are further disaggregated by contract length (multi-year, indefinite, annual and less-than-annual).  From 2012 through 2015 the multi-year and indefinite contract lengths were combined as one category and often referred to as multi-year in past documentation.
}

\section{Data Layout}{
 This file has multiple records per institution.  Each record is uniquely defined by the variables IPEDS ID (UNITID), and the variable SNHCAT which defines the primary occupation and faculty/tenure status for instructional staff.
}

\section{Institutions}{
These data are applicable to degree-granting institutions with 15 or more full-time employees and related administrative offices.
}

\section{Note: Amendments}{
Prior to the IPEDS Web-based data collection in 2001, human resource data reported by race/ethnicity and gender was only available in odd-numbered years.  Beginning with 2001, human resource data reported by race/ethnicity and gender was mandatory in odd-numbered years and optional in even-numbered years.  Beginning with 2016 reporting human resource data by race/ethnicity and gender is mandatory annually.
}


\section{Glossary}{
\describe{

\item{Annual contract}{An annually-renewable contract or employment agreement that is in effect for a stated annual period within one year of execution, and may be equal to a period of 365 days, or a standard academic year, or the equivalent. Does not include contracts for partial year periods such as a single semester, quarter, term, block, or course.}

\item{Faculty}{Persons identified by the institution as such and typically those whose initial assignments are made for the purpose of conducting instruction, research or public service as a principal activity (or activities). They may hold academic rank titles of professor, associate professor, assistant professor, instructor, lecturer or the equivalent of any of those academic ranks. Faculty may also include the chancellor/president, provost, vice provosts, deans, directors or the equivalent, as well as associate deans, assistant deans and executive officers of academic departments (chairpersons, heads or the equivalent) if their principal activity is instruction combined with research and/or public service. The designation as "faculty" is separate from the activities to which they may be currently assigned. For example, a newly appointed president of an institution may also be appointed as a faculty member. Graduate, instruction, and research assistants are not included in this category .}

\item{Full-time staff (employees) }{As defined by the institution. The type of appointment at the snapshot date determines whether an employee is full time or part time. The employee's term of contract is not considered in making the determination of full or part time.}

\item{Less-than-annual-contract}{ A contract or employment agreement that is in effect for a partial year period of less than 365 days, or less than a standard academic year or the equivalent. Includes contracts for partial year periods such as a single semester, quarter, term, block, or course.}

\item{Multi-year contract }{A contract or employment agreement that is in effect for more than one year (e.g., more than 365 days) where subsequent years of the contract may be contingent upon other factors such as the appropriation of funds. The renewal period of a multi-year contract is not on an annual basis (e.g., a 5-year multi-year contract is renewed every 5 years NOT annually).}

\item{New hires}{Persons who were hired for full-time permanent employment for the first time, or after a break in service, between July 1 and September 30 of the survey year. These do not include persons who have returned from sabbatical leave. }

\item{Not on tenure track}{Personnel positions that are considered non-tenure earning positions }

\item{Occupation [OCCUPCAT]}{This variable identifies postsecondary staff by IPEDS Human Resource occupation categories.

The IPEDS HR survey component uses the Standard Occupational Classification (SOC) system to assist institutions in classifying employees for IPEDS reporting purposes. The SOC is a system used by Federal statistical agencies to classify workers and jobs into occupational categories for the purpose of collecting, calculating, analyzing, or disseminating data. In 2010 and 2018, updated versions of the SOC system (http://www.bls.gov/soc/) was released. Consequently, the IPEDS HR survey was revised to incorporate the 2010 and then 2018 SOC information. The new IPEDS occupational categories are effective with  2012-13 and then 2018-19 Human Resources reporting.}

\item{Old new hire categories that are consitent with new codes [SGTYPE]}{Original function code on survey form}

\item{Primary occupational activity}{The principal activity of a staff member as determined by the institution. If an individual participates in two or more activities, the primary activity is normally determined by the amount of time spent in each activity.}

\item{Staff category [SNHCAT]}{This variable classifies Full-time new hires by occupational category.  Instructional staff  are also classified by faculty and tenure status. }

\item{Tenure}{Status of a personnel position, or a person occupying a position or occupation, with respect to permanence of the position.}

\item{Tenure track}{Personnel positions that lead to consideration for tenure.  }
  }
}

