\name{s_sis}
\alias{s_sis}

\title{Fall Staff: S_SIS}

\description{
This file contains the number of full-time instruction/research/public service staff on the payroll of the institution as of November 1,  by academic rank, faculty and tenure status.
}

\section{Data Layout}{
 This file has multiple records per institution.  Each record is uniquely defined by the variables IPEDS ID (UNITID), and the variable FACSTAT which is the combination of faculty status and tenure status (tenured, on tenure track, and not on-tenure track/no tenure system).
}

\section{Institutions}{
These data are applicable to degree-granting institutions with 15 or more full-time employees and related administrative offices.
}

\section{Amendments}{
Beginning with 2016, staff that are not on-tenure/no tenure system are also classified by the type of contract (multi-year, indefinite, annual, and less-than-annual).  From 2012 through 2015 the multi-year and indefinite contract lengths were combined as one category and often referred to as multi-year in past documentation.
}


\section{Glossary}{
\describe{

\item{All ranks [SISTOTL]}{A status designated by the institution according to the institution's policies. The IPEDS HR survey includes the ranks of Professor, Associate Professor, Assistant Professor, Instructor, and Lecturer.}

\item{Annual contract}{An annually-renewable contract or employment agreement that is in effect for a stated annual period within one year of execution, and may be equal to a period of 365 days, or a standard academic year, or the equivalent. Does not include contracts for partial year periods such as a single semester, quarter, term, block, or course.}

\item{Faculty}{Persons identified by the institution as such and typically those whose initial assignments are made for the purpose of conducting instruction, research or public service as a principal activity (or activities). They may hold academic rank titles of professor, associate professor, assistant professor, instructor, lecturer or the equivalent of any of those academic ranks. Faculty may also include the chancellor/president, provost, vice provosts, deans, directors or the equivalent, as well as associate deans, assistant deans and executive officers of academic departments (chairpersons, heads or the equivalent) if their principal activity is instruction combined with research and/or public service. The designation as "faculty" is separate from the activities to which they may be currently assigned. For example, a newly appointed president of an institution may also be appointed as a faculty member. Graduate, instruction, and research assistants are not included in this category .}

\item{Full-time staff (employees) }{As defined by the institution. The type of appointment at the snapshot date determines whether an employee is full time or part time. The employee's term of contract is not considered in making the determination of full or part time.}

\item{Instruction (only) category}{Persons whose specific assignments customarily are made for the purpose of providing instruction or teaching. Regardless of title, academic rank, or tenure status, these employees formally spend the majority of their time providing instruction or teaching.}

\item{Instruction combined with research and/or public service (IRPS)}{Persons for whom it is not possible to differentiate between instruction or teaching, research, and public service because each of these functions is an integral component of his/her regular assignment. Regardless, of title, academic rank, or tenure status, these employees formally spend the majority of their time providing instruction, research, and/or public service. }

\item{Instructional Staff}{An occupational category that is comprised of staff who are either:
\itemize{
  \item Primarily Instruction (See "Instruction (only) category")
  \item or Instruction combined with research and/or public service. (See "Instruction combined with research and/or public service (IRPS)")
}
 The intent of the Instructional Staff category is to include all individuals whose primary occupation includes instruction at the institution.  }

\item{Less-than-annual-contract}{ A contract or employment agreement that is in effect for a partial year period of less than 365 days, or less than a standard academic year or the equivalent. Includes contracts for partial year periods such as a single semester, quarter, term, block, or course.}

\item{Multi-year contract }{A contract or employment agreement that is in effect for more than one year (e.g., more than 365 days) where subsequent years of the contract may be contingent upon other factors such as the appropriation of funds. The renewal period of a multi-year contract is not on an annual basis (e.g., a 5-year multi-year contract is renewed every 5 years NOT annually).}

\item{Not on tenure track}{Personnel positions that are considered non-tenure earning positions }

\item{Tenure}{Status of a personnel position, or a person occupying a position or occupation, with respect to permanence of the position.}

\item{Tenure track}{Personnel positions that lead to consideration for tenure.  }
  }
}

