\name{sal_a}
\alias{sal_a}

\title{Instructional Staff/Salaries: SAL_A}

\description{
This file contains the number of full-time instructional staff and the total and average salary outlays for the full-time instructional staff, by length of contract and academic rank.
}

\section{Data Layout}{
This data file has multiple records for each institution. Each record is uniquely defined by the variables IPEDS ID (UNITID), and the variable contract length (CONTRACT) and academic rank (ARANK) of full-time instructional staff.
}


\section{Glossary}{
\describe{

\item{11/12-month salary contract/teaching period}{The contracted teaching period of faculty employed for the entire year, usually a period of 11 or 12 months. }

\item{9/10-month salary contract/teaching period}{The contracted teaching period of faculty employed for 2 semesters, 3 quarters, 2 trimesters,  4-month sessions, or the equivalent.
}

\item{Academic rank [ARANK]}{Assigned by the institution and includes Professors, Associate professors, Assistant professors, Instructors, Lecturers, and No academic rank.}

\item{Average Salary (equated 9-month contract)}{Salary outlays (equated 9-month contract) divided by the number of faculty (equated 9-month contract)}

\item{Full-time instructional facutly}{Instruction/research staff employed full time (as defined by the institution) whose major regular assignment is instruction, including those with released time for research. For the Faculty Salaries survey, this group includes faculty designated as "primarily instruction" and "instruction, combined with research and public service."}

\item{Less-than-9-month salary contract/teaching period}{The contracted teaching period of faculty employed for less than 2 semesters, 3 quarters, 2 trimesters, or 2 4-month sessions.}

\item{Number of faculty(equated 9-month contract)}{Number of faculty on 9/10 month salary contract+Number of faculty on 11/12 month contract}

\item{Salary outlays (equated 9-month contract)}{Salary outlays of faculty on 9/10 month contract + (.8182 * salary outlays of faculty on 11/12 month contract)}
  }
}

