\name{sal_a_lt9}
\alias{sal_a_lt9}

\title{Instructional Staff/Salaries: SAL_A_LT9}

\description{
This file contains the number of full-time instructional staff with less than 9-month contracts by academic rank.
}

\section{Data Layout}{
This data file has multiple records for each institution. Each record is uniquely defined by the variables IPEDS ID (UNITID),  and academic rank (ARANK) of full-time instructional staff.
}


\section{Glossary}{
\describe{

\item{Academic rank [ARANK]}{Assigned by the institution and includes Professors, Associate professors, Assistant professors, Instructors, Lecturers, and No academic rank.}

\item{Full-time instructional facutly}{Instruction/research staff employed full time (as defined by the institution) whose major regular assignment is instruction, including those with released time for research. For the Faculty Salaries survey, this group includes faculty designated as "primarily instruction" and "instruction, combined with research and public service."}
  }
}

