\name{sal_faculty}
\alias{sal_faculty}

\title{Instructional Staff/Salaries: SAL_FACULTY}

\description{
This file contains the number of full-time instructional staff by faculty status, length of contract and academic rank.
}

\section{Data Layout}{
This data file has multiple records for each institution. Each record is uniquely defined by the variables IPEDS ID (UNITID), and the variable length of contract (CONTRACT) and academic rank (ARANK) of full-time instructional staff.
}

\section{Faculty Status}{
Faculty status includes the following categories:
\itemize{
  \item tenured,
  \item on-tenure-track,
  \item not-on-tenure track/no tenure system,
  \item without faculty status.
}
}


\section{Glossary}{
\describe{

\item{11/12-month salary contract/teaching period}{The contracted teaching period of faculty employed for the entire year, usually a period of 11 or 12 months. }

\item{9/10-month salary contract/teaching period}{The contracted teaching period of faculty employed for 2 semesters, 3 quarters, 2 trimesters,  4-month sessions, or the equivalent.
}

\item{Academic rank [ARANK]}{Assigned by the institution and includes Professors, Associate professors, Assistant professors, Instructors, Lecturers, and No academic rank.}

\item{Faculty}{Persons identified by the institution as such and typically those whose initial assignments are made for the purpose of conducting instruction, research or public service as a principal activity (or activities). They may hold academic rank titles of professor, associate professor, assistant professor, instructor, lecturer or the equivalent of any of those academic ranks. Faculty may also include the chancellor/president, provost, vice provosts, deans, directors or the equivalent, as well as associate deans, assistant deans and executive officers of academic departments (chairpersons, heads or the equivalent) if their principal activity is instruction combined with research and/or public service. The designation as "faculty" is separate from the activities to which they may be currently assigned. For example, a newly appointed president of an institution may also be appointed as a faculty member. Graduate, instruction, and research assistants are not included in this category .}

\item{Full-time instructional facutly}{Instruction/research staff employed full time (as defined by the institution) whose major regular assignment is instruction, including those with released time for research. Includes faculty designated as "primarily instruction" and "instruction, combined with research and public service."}

\item{Not on tenure track/no tenure system}{Faculty who are in non-tenure earning positions.}

\item{On tenure track}{Instructional staff who are in positions that lead to consideration for tenure.}

\item{Tenured}{Faculty who are tenured according to the institution's criteria or requirements for either, notwithstanding the fact that the policy used by the institution may be different from that which meets or refers to a national set of principles.}
  }
}

