\name{sal_nis}
\alias{sal_nis}

\title{Instructional Staff/Salaries: SAL_NIS}

\description{
This file contains the number and salary outlays for full-time, nonmedical, non-instructional staff by occupational category.
}

\section{Data Layout}{
This data file has multiple records for each institution. Each record is uniquely defined by the variables IPEDS ID (UNITID), and the variable contract length (CONTRACT) and academic rank (ARANK) of full-time instructional staff.
}


\section{Glossary}{
\describe{

\item{Academic Affairs and Other Education Services Occupations}{An occupational category based on the following three minor groups in the 2010 Standard Occupational Classification (SOC) Manual:
\itemize{
  \item 1) Pre-school, Primary, Secondary, and Special Education School Teachers;
  \item 2) Other Teachers and Instructors;
  \item 3) Other Education, Training, and Library Occupations.
}
For more information visit: https://www.bls.gov/soc/2010/home.htm}

\item{Archivists, Curators, and Museum Technicians}{An occupational category based on the broad occupation in the 2010 Standard Occupational Classification (SOC) Manual called "Archivists, Curators, and Museum Technicians."  For detailed information, refer to the following website: https://www.bls.gov/soc/2010/home.htm. }

\item{Business and Financial Operations Occupations}{An occupational category based on the major group in the 2010 Standard Occupational Classification (SOC) Manual called “Business and Financial Operations Occupations.” For detailed information refer to the following website: https://www.bls.gov/soc/2010/home.htm .}

\item{Community, Social  Service, Legal, Arts, Design, Entertainment, Sports and Media  Occupations}{An occupational category based on the following three major groups in the 2010 Standard Occupational Classification (SOC) Manual:
\itemize{
  \item 1) Community and Social Service Occupations;
  \item 2) Legal Occupations;
  \item 3) Arts, Design, Entertainment, Sports, and Media Occupations.
}
For more information visit: https://www.bls.gov/soc/2010/home.htm}

\item{Computer, Engineering, and Science Occupations}{An occupational category based on the following three major groups in the 2010 Standard Occupational Classification (SOC) Manual:
\itemize{
  \item 1) Computer and Mathematical Occupations;
  \item  2) Architecture and Engineering Occupations;
  \item 3) Life, Physical, and Social Science Occupations.
}
For more information visit: https://www.bls.gov/soc/2010/home.htm}

\item{Healthcare Practitioners and Technical Occupations}{An occupational category based on the major group in the 2010 Standard Occupational Classification (SOC) Manual called “Healthcare Practitioners and Technical Occupations.” For detailed information refer to the following website:  https://www.bls.gov/soc/2010/home.htm}

\item{Librarians}{An occupational category based on the broad occupation in the 2010 Standard Occupational Classification (SOC) Manual called "Librarians."  For detailed information, refer to the following website: https://www.bls.gov/soc/2010/home.htm }

\item{Library Technicians}{An occupational category based on the broad occupation in the 2010 Standard Occupational Classification (SOC) Manual called "Library Technicians."  For detailed information, refer to the following website: https://www.bls.gov/soc/2010/home.htm .  }

\item{Management Occupations}{An occupational category based on the major group in the 2010 Standard Occupational Classification (SOC) Manual called “Management Occupations.” For detailed information refer to the following website: https://www.bls.gov/soc/2010/home.htm .}

\item{Medical staff}{Staff employed by or staff working in the medical school (Doctor of Medicine [M.D.] and/or Doctor of Osteopathic Medicine [D.O.]) component of a postsecondary institution or in a free standing medical school. Does not include staff employed by or employees working strictly in a hospital associated with a medical school or those who work in health or allied health schools or departments such as dentistry, veterinary medicine, nursing or dental hygiene unless the health or allied health schools or departments are affiliated with (housed in or under the authority of) the medical school.}

\item{Natural Resources, Construction, and Maintenance Occupations}{An occupational category based on the following three major groups in the 2010 Standard Occupational Classification (SOC) Manual:
\itemize{
  \item 1) Farming, Fishing, and Forestry Occupations
  \item 2) Construction and Extraction Occupations
  \item 3) Installation, Maintenance, and Repair Occupations
}
For more information visit: https://www.bls.gov/soc/2010/home.htm}

\item{Non-instructional staff}{Full-time, non-medical, non-instructional staff.}

\item{Office and Administrative Support Occupations}{An occupational category based on the major group in the 2010 Standard Occupational Classification (SOC) Manual called “Office and Administrative Support Occupations.” For detailed information refer to the following website: https://www.bls.gov/soc/2010/home.htm.}

\item{Production, Transportation, and Material Moving Occupations}{An occupational category based on the following two major groups in the 2010 Standard Occupational Classification (SOC) Manual:
\itemize{
  \item 1) Production Occupations
  \item 2) Transportation and Material Moving Occupations
}
For more information visit: https://www.bls.gov/soc/2010/home.htm
}

\item{Public service (staff)}{An occupational category used to classify persons whose specific assignments customarily are made for the purpose of carrying out public service activities such as agricultural extension services, clinical services, or continuing education. Regardless of title, academic rank, or tenure status, these employees formally spend the majority of their time carrying out public service activities. (This category includes employees with a public service assignment regardless of the location of the assignment (e.g., in the field rather than on campus).}

\item{Research (staff)}{An occupational category used to classify persons whose specific assignments customarily are made for the purpose of conducting research. Regardless of title, academic rank, or tenure status, these employees formally spend the majority of their time conducting research}

\item{Sales and Related Occupations}{An occupational category based on the major group in the 2010 Standard Occupational Classification (SOC) Manual called “Sales and Related Occupations.” For detailed information refer to the following website:  https://www.bls.gov/soc/2010/home.htm.}

\item{Service Occupations}{An occupational category based on the following five major groups in the 2010 Standard Occupational Classification (SOC) Manual:
\itemize{
  \item 1) Healthcare Support Occupations;
  \item 2) Protective Service Occupations;
  \item 3) Food Preparation and Serving Related Occupations;
  \item 4) Building and Grounds Cleaning and Maintenance Occupations;
  \item 5) Personal Care and Service Occupations.
}
For more information visit: https://www.bls.gov/soc/2010/home.htm}

  }
}

