\name{sfa}
\alias{sfa}

\title{Student Financial Aid and Net Price: SFA}

\description{
This file contains data on the number of full-time, first-time degree/certificate-seeking undergraduate students and all undergraduate students who receive different types of student financial aid, including grants and loans, from different sources at each institution.
}

\section{Sources and Types of Grants}{
Sources and types of aid reported for full-time, first-time degree/certificate-seeking undergraduate students include federal grants, Pell grants, other federal grants, state/local grants, grants from the institution, federal and nonfederal student loans. Types of aid reported for all undergraduates include total grants, Pell grants and federal loans. Included are the total and average dollar amount of aid received by these students.
}

\section{Characteristics}{
 This file also contains the average net price at each institution for the following two groups:
 \itemize{
  \item (1) full-time, first-time degree/certificate-seeking undergraduate students who receive grant and scholarship aid;
  \item (2) full-time, first-time degree/certificate-seeking undergraduate students who receive Title IV federal student aid.
}
For public institutions, group 1 and group 2 include only students who pay the in-district or in-state tuition rate.
}

\section{Duplicate Variable Titles}{
Some variables have the same variable titles. The distinction is:
\itemize{
  \item "GIST" starting variables apply to full-time, first-time degree/certificate-seeking undergraduates, paying the in-state or in-district tuition rate who were awarded grant or scholarship aid from federal, state or local governments, or the institution. Other sources of grant aid are excluded. Aid awarded anytime during the full aid year is included.

  Applicable to public institutions with standard calendar systems (semester, quarter, trimester, 4-1-4) that report financial aid data for students enrolled for the FALL and awarded aid anytime during the full aid year.
  \item "GRNT" starting variables apply to  full-time, first-time degree/certificate-seeking undergraduates who were awarded grant or scholarship aid from federal, state or local governments, or the institution.Students awarded only other sources of grant aid are excluded.

  Applicable to private not-for-profit and for-profit institutions with standard calendar systems (semester, quarter, trimester, 4-1-4) that report financial aid data for students enrolled for the FALL and awarded aid anytime during the full aid year.

Also, applicable to public or private institutions that measure courses primarily by contact hours, or offers primarily occupational programs measured in credit hours with calendar systems that are continuous or differs by programs. These institutions report financial aid data for  students enrolled between the 12-month period (September 1 through August 31) and awarded aid anytime during the full aid year.

  \item "NPIS" starting variables apply to full-time, first-time degree/certificate-seeking undergraduates paying the in-state or in-district tuition rate.

  Applicable to public institutions with standard calendar systems (semester, quarter, trimester, 4-1-4) that report financial aid data for students enrolled for the FALL and awarded aid anytime during the full aid year.


  Students whose living arrangements are unknown are excluded from the calculation.  For some institutions the # of students by living arrangement will be known, but dollar amounts will not be known.  In this case the # of students with no corresponding dollar amount will be excluded from the denominator.

  \item "NPT" starting variables apply to full-time, first-time degree/certificate-seeking undergraduates.

  Applicable to private not-for-profit and for-profit institutions with standard calendar systems (semester, quarter, trimester, 4-1-4) that report financial aid data for students enrolled for the FALL and awarded aid anytime during the full aid year. These institutions report cost of attendance for the full academic year.

Also, applicable to public or private institutions that measure courses primarily by contact hours, or offers primarily occupational programs measured in credit hours with calendar systems that are continuous, or differs by programs. These institutions report financial aid data for students enrolled between the 12-month period (September 1 through August 31) and awarded aid anytime during the full aid year. These institutions report cost of attendance for their largest program for the length of that program.

 For institutions that report cost of attendance for the full academic year an average yearly net price is generated by subtracting the average amount of federal, state or local government, or institutional grant and scholarship aid from the total cost of attendance.

 For institutions that report the cost of attendance for their largest program for the length of that program, the cost of attendance is generated as stated above, but is then divided by the number of months to complete the program to produce a monthly cost of attendance. The average amount of federal, state or local government, or institutional grant and scholarship aid is also divided by 12 to produce a monthly average. The average net price is generated by subtracting the average monthly grant and scholarship aid from the average monthly cost of attendance and the difference is multiplied by the number of months to complete the program.

  Students whose living arrangements are unknown are excluded from the calculation.  For some institutions the # of students by living arrangement will be known, but dollar amounts will not be known.  In this case the # of students with no corresponding dollar amount will be excluded from the denominator.

  \item "GIS4G" starting variables apply to full-time, first-time degree/certificate-seeking undergraduate students who paid the in-state or in-district tuition rate and were awarded any Title IV federal student aid and were awarded any grant or scholarship aid from the following sources: the federal government, state/local government, or the institution.

  This variable includes Title IV recipients who were awarded grants or scholarship aid from the federal government, state/local government, or a grant or scholarship from the institution. Students that awarded grant or scholarship aid from private or other sources, or loans or Federal work study are excluded.

  Applicable to public institutions with standard calendar systems (semester, quarter, trimester, 4-1-4) that report financial aid data for students enrolled for the FALL and awarded aid anytime during the full aid year.

  \item "GRN4G" starting variables apply to full-time, first-time degree/certificate-seeking undergraduate students who were awarded any Title IV federal student aid and were awarded any grant or scholarship aid from the following sources: the federal government, state/local government, or the institution.

This variable includes Title IV recipients who were awarded grants or scholarship aid from the federal government, state/local government, or a grant or scholarship from the institution. Students that awarded grant or scholarship aid from private or other sources, or loans or Federal work study are excluded.

Applicable to private not-for-profit and for-profit institutions with standard calendar systems (semester, quarter, trimester, 4-1-4) that report financial aid data for students enrolled for the FALL and awarded aid anytime during the full aid year.

Also, applicable to private institutions that measure courses primarily by contact hours, or offers primarily occupational programs measured in credit hours with calendar systems that are continuous or differs by programs. These institutions report financial aid data for  students enrolled between the 12-month period (September 1 through August 31) and awarded aid anytime during the full aid year.

  \item "GIS4N"/"GIS4U"/"GIS4O"/"GIS4W"/"GIS4T" starting variables apply to full-time, first-time degree/certificate-seeking undergraduates paying the in-state or in-district tuition rate who were awarded title IV federal student aid.  Title IV federal student aid, includes federal grants or federal student loans.

  Applicable to public institutions with standard calendar systems (semester, quarter, trimester, 4-1-4) that report financial aid data for students enrolled for the FALL and awarded aid anytime during the full aid year.

  \item "GRN4N"/"GRN4U"/"GRN4O"/"GRN4W"/"GRN4T" starting variables apply to full-time, first-time degree/certificate-seeking undergraduates who were awarded title IV federal student aid.  Title IV federal student aid, includes federal grants or federal student loans.

  Applicable to private not-for-profit and for-profit institutions with standard calendar systems (semester, quarter, trimester, 4-1-4) that report financial aid data for students enrolled for the FALL and awarded aid anytime during the full aid year.

Also, applicable to public or private institutions that measure courses primarily by contact hours, or offers primarily occupational programs measured in credit hours with calendar systems that are continuous or differs by programs. These institutions report financial aid data for  students enrolled between the 12-month period (September 1 through August 31) and awarded aid anytime during the full aid year.
}
}


\section{Glossary}{
\describe{

\item{Average net price}{Generated by subtracting the average amount of federal, state or local government, or institutional grant and scholarship aid from the total cost of attendance.}

\item{Degree/certificate-seeking students}{Students enrolled in courses for credit who are recognized by the institution as seeking a degree or other formal award. At the undergraduate level, this is intended to include students enrolled in vocational or occupational programs.}

\item{Federal grants (grants/educational assistance funds)}{Grants provided by federal agencies such as the U.S. Department of Education, including Title IV Pell Grants and Supplemental Educational Opportunity Grants (SEOG). Also includes need-based and merit-based educational assistance funds and training vouchers provided from other federal agencies and/or federally-sponsored educational benefits programs, including the Veteran's Administration, Department of Labor, and other federal agencies.}

\item{Financial aid}{Grants, loans, assistantships, scholarships, fellowships, tuition waivers, tuition discounts, veteran's benefits, employer aid (tuition reimbursement) and other monies (other than from relatives/friends) provided to students to meet expenses. This includes Title IV subsidized and unsubsidized loans made directly to students. }

\item{First-time student (undergraduate)}{A student attending any institution for the first time at the undergraduate level. Includes students enrolled in academic or occupational programs . Also includes students enrolled in the fall term who attended college for the first time in the prior summer term, and students who entered with advanced standing (college credits earned before graduation from high school). }

\item{Full-time student (Undergraduate)}{A student enrolled for 12 or more semester credits , or 12 or more quarter credits, or 24 or more contact hours a week each term.        }

\item{Full-time, first-time degree/certificate-seeking undergraduates}{A student enrolled in a 4- or 5-year bachelor's degree program, an associate's degree program, or a vocational or technical program below the baccalaureate level, who has no prior postsecondary experience, and is enrolled for 12 or more semester credits, or 12 or more quarter credits, or 24 or more contact hours a week each term.}

\item{Grant aid}{Includes any grant or scholarship aid awarded, from the federal government, a state or local government, the institution, and other sources known by the institution}

\item{Income level}{The income that was used by the institution's financial aid office to determine the student's Expected Family Contribution (EFC). For dependent students this will include the parents' adjusted gross income and the student's adjusted gross income. For independent students this will include the student's adjusted gross income.}

\item{In-district student}{A student who is a legal resident of the locality in which he/she attends school and thus is entitled to reduced tuition charges if offered by the institution.}

\item{In-district tuition}{The tuition charged by the institution to those students residing in the locality in which they attend schooland may be a lower rate than in-state tuition if offered by the institution.}

\item{In-state student}{A student who is a legal resident of the state in which he/she attends school. }

\item{In-state tuition}{The tuition charged by institutions to those students who meet the state's or institution's residency requirements. }

\item{Institutional grants}{Scholarships and fellowships granted and funded by the institution and/or individual departments within the institution, (i.e., instruction, research, public service) that may contribute indirectly to the enhancement of these programs . Includes scholarships targeted to certain individuals (e.g., based on state of residence, major field of study, athletic team participation) for which the institution designates the recipient. }

\item{Loans to students}{Any monies that must be repaid to the lending institution for which the student is the designated borrower. Includes all Title IV subsidized and unsubsidized loans. Does not include PLUS and other loans made directly to parents.}

\item{Number of full-time first-time undergraduates awarded any financial aid [ANYAIDN]}{Number of full-time, first-time degree/certificate-seeking undergraduate students who were awarded any financial aid.  Includes those who were awarded Federal Work Study, any loans to students or grant or scholarship aid from the federal government, state/local government, or the institution.  Plus loans to parents are not included. Grant or scholarship aid from private or other sources (e.g., Rotary Club Scholarship).}

\item{Number of students in fall cohort [SCFA1N]}{The number of full-time first-time degree/certificate-seeking undergraduates (fall cohort of previous academic year).}

\item{Number of students in full-year cohort [SCFY1N]}{Unduplicated count of full-time first-time degree\ certificate-seeking undergraduate students (current academic year, full-year cohort). (See "Unduplicated count")}

\item{Off-campus housing}{Housing facility that is occupied by students but is not owned or controlled by the educational institution.}

\item{On-campus housing}{Residence halls owned or controlled by an institution within the same reasonably contiguous geographic area and used by the institution in direct support of or in a manner related to, the institution's educational purposes.}

\item{Other federal grants}{Federal monies awarded to the institution under federal government student aid programs, such as Supplemental Educational Opportunity Grants (SEOG), DHHS training grants (aid portion only), State Student Incentive Grants (SSIG), and other federal student aid programs. Pell grants are not included in this classification.

Note: if the federal government selects the student recipients and simply transmits the funds to the institution for disbursement to the student, the amounts are not considered as revenues and subsequently there are no discounts & allowances or scholarships and fellowships expenses. If the funds are made available to the institution for selection of student recipients, then the amounts awarded are considered as nonoperating revenues and subsequently as discounts & allowances or scholarships and fellowships expenses }

\item{Other loans to students}{All nonfederal loans institutionally- and privately-sponsored loans. Does not include other loans made directly to parents.  }

\item{Out-of-state student}{A student who is not a legal resident of the state in which he/she attends school. }

\item{Pell Grant program}{(Higher Education Act of 1965, Title IV, Part A, Subpart I, as amended.)
Provides grant assistance to eligible undergraduate postsecondary students with demonstrated financial need to help meet education expenses.}

\item{Standard academic terms}{Consist the following calendar systems:semester, quarter, trimester, or 4-1-4.}

\item{State and local government grants}{State and local monies awarded to the institution under state and local student aid programs, including the state portion of State Student Incentives Grants (SSIG). }

\item{Students awarded Title IV federal student aid}{Include students who were awarded federal grant aid or who were awarded federal work study or federal student loans.}

\item{Title IV aid to students}{Includes grant aid, work study aid, and loan aid. These include: Federal Pell Grant, Federal Supplemental Educational Opportunity Grant (FSEOG), Academic Competitiveness Grant (ACG), National Science and Mathematics Access to Retain Talent Grant (National SMART Grant), Teacher Education Assistance for College and Higher Education (TEACH) Grant, Federal Work-Study, Federal Perkins Loan, Subsidized Direct or FFEL Stafford Loan, and Unsubsidized Direct or FFEL Stafford Loan.}

\item{Total cost of attendance}{The sum of published tuition and required fees (lower of in-district or in-state), books and supplies and the weighted average room and board and other expenses.}

\item{Total number of full-time first-time degree/certificate seeking undergraduates - financial aid cohort [SCUGFFN]}{Total number of full-time first-time degree/certificate-seeking undergraduate students - financial aid cohort.

The number of full-time first-time degree/certificate-seeking undergraduate students include all full-time first-time degree/certificate-seeking undergraduate students enrolled at an institution as of October 15 (or the institution's official fall reporting date) for institutions with standard academic terms (See "Standard academic terms").
For institutions that do not operate on standard academic terms (program reporters) the number of full-time first-time degree/certificate-seeking undergraduate students are based on a full year cohort (unduplicated counts) of students enrolled during the 12-month period September 1 through August 31.}

\item{Total number of undergraduates - financial aid cohort [SCUGRAD]}{Total number of undergraduate students - as reported on the student financial aid component.

The number of undergraduate students include all undergraduate students enrolled at an institution as of October 15 (or the institution's official fall reporting date) for institutions with standard academic terms (See "Standard academic terms").
For institutions that do not operate on standard academic terms (program reporters) the number of undergraduate students are based on a full year cohort (unduplicated counts) of students enrolled during the 12-month period September 1 through August 31.}

\item{Total number of undergraduates - full-year cohort [SCFY2]}{Total unduplicated count of all undergraduates enrolled. (See "Unduplicated count")}

\item{Undergraduate (student)}{A student enrolled in a 4- or 5-year bachelor's degree program, an associate's degree program, or a vocational or technical program below the baccalaureate.}

\item{Unduplicated count}{The sum of students enrolled for credit with each student counted only once during the reporting period, regardless of when the student enrolled.}

\item{Unknown residence}{Status used when the reporting institution is unable to determine from existing records the home state or residence of the student.       }

\item{Weighted average for room and board and other expenses}{Generated as follows:

  (amount for on-campus room, board and other expenses * # of students living on-campus.
  +  amount for off-campus (with family) room, board and other expenses * # of students living off-campus with family
  +  amount for off-campus (not with family) room, board and other expenses * # of students living off-campus not with family)

divided by the total # of students.}
  }
}

