\name{sfav}
\alias{sfav}

\title{Student Financial Aid and Net Price: SFAV}

\description{
This file contains the number of students, who received either or both Post-9/11 GI Benefits or the Department of Defense Tuition Assistance through the institution.
}

\section{Characteristics}{
Data includes:
\itemize{
  \item Eligible dependents or eligible spouse receiving such benefits.
  \item Students who received both benefits, are counted in both programs.
  \item Data are provided for undergraduate and graduate students.
  \item Total dollar amounts received by students.
}
Information reported is only what is known to the institution. Benefits that were not awarded/certified through the institution are not included.

For Post-9/11 GI Bill Benefits, matching institutional aid provided through the Yellow Ribbon Program is not included.

}

\section{Reporting}{
Data reported for the Post-9/11 GI Bill benefit are for the academic year July 1 of previous year to June 30 of current year.

Data reported for the Department of Defense Tuition Assistance are for the academic year October 1 of previous year to September 30 of current year.
}


\section{Glossary}{
\describe{
\item{Military Tuition Assistance Program (TA)}{A program that funds up to 100 percent of an eligible servicemember’s college tuition and course-specific fees. Available only to eligible servicemembers who are currently in active service as long as criteria limits are not exceeded and students are enrolled off-duty in an U.S. Department of Education accredited post-secondary institution. This military benefit is paid directly to the postsecondary institution by the individual’s Armed service.}

\item{Post 9/11 GI Bill}{A federal education benefit program for veterans, who served on active duty after September 10, 2001. This Department of Veteran Affairs benefit provides up to 36 months of education benefits at an approved institution for the following college costs: tuition and fees, books and supplies and housing. The tuition and fees payment, which is the cost for an in-state student attending a public institution, is made directly to the postsecondary institution whereas payments for books and supplies and housing are sent directly to the student.}

\item{Yellow Ribbon Program}{A voluntary program through which participating public and private institutions can provide veterans and eligible beneficiaries additional institutional aid to cover the costs of tuition and fees at their institutions. The Yellow Ribbon Program is a supplementary program to the Post 9/11 GI Bill coverage of in-state tuition and fees. The Department of Veterans Affairs matches the institutional aid provided beyond the in-state tuition and fees, but to certain limit each year.}

  }
}

