% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyipeds-package.R
\docType{package}
\name{tidyipeds-package}
\alias{tidyipeds}
\alias{tidyipeds-package}
\title{tidyipeds: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Sorenson-Impact/tidyipeds}
  \item Report bugs at \url{https://github.com/Sorenson-Impact/tidyipeds/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Zadra \email{jon.zadra@sorensonimpact.com}

Other contributors:
\itemize{
  \item Sorenson Impact Center [copyright holder]
}

}
\keyword{internal}
