% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data - scrape and download.R
\name{update_available_ipeds}
\alias{update_available_ipeds}
\title{Update local list of available data center files}
\usage{
update_available_ipeds(force = FALSE)
}
\arguments{
\item{force}{Logical indicating whether to force an update if local copy is less than 1 month old. Defaults to FALSE}
}
\description{
\lifecycle{experimental}
This function is primarily internal. It is used to update (or download if it does not exist) a local copy of the data files available from IPEDS data center.  If the local copy is less than 1 month old, it does not update as scraping the data center is time consuming and new files are added only a few times a year.  This behavior can be circumvented by using \code{`force = TRUE`}.  If the local copy is over 1 month old, it is updated.
}
\examples{
\dontrun{
update_available_ipeds(force = TRUE)
}
}
