\name{adm}
\alias{adm}

\title{Admissions and Test Scores: ADM}

\description{
This file contains information about the undergraduate selection process for entering first-time, degree/certificate-seeking students. This includes information about admission considerations,  applicants, applicants that were admitted, and admitted students who enrolled.
}

\section{SAT and ACT test}{
SAT and ACT test scores are included for institutions, that require test scores for admission. These data are applicable for institutions that do not have an open admissions policy for entering first-time students.
Beginning with 2016, writing scores for both SAT and ACT are no longer collected.
}

\section{Admission Consideration}{
The possible values for the admission consideration variables have changed from data prior to 2016.  "Do not know" is no longer an option and "considered but not required" was added.  See frequencies for complete listing.
}

\section{Data Releases}{
Preliminary release data have been edited but are subject to further NCES quality control procedures. Imputed data for nonresponding institutions are not included. These data are used for First Look (Preliminary Data) publications.

Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions. These data are used for First Look (Provisional Data) publications.

Final release data include revisions to the provisional release data that have been made by institutions during the subsequent data collection year. The final release data can be used when the most up to date data are required; however, these data may not match tables from the First Look reports based on preliminary and provisional data.
}

\section{Glossary}{
\describe{
  \item{SAT (Scholastic Aptitude Test (SAT)}{An examination administered by the Educational Testing Service and used to predict the facility with which an individual will progress in learning college-level academic subjects.}
  \item{American College Testing Program (ACT)}{The ACT assessment program measures educational development and readiness to pursue college-level coursework in English, mathematics, natural science, and social studies. Student performance does not reflect innate ability and is influenced by a student's educational preparedness.}
  \item{Test of English as a Foreign Language (TOEFL)}{- Standardized test designed to determine an applicant's ability to benefit from instruction in English.}
  \item{First-time student (Undergraduate)}{A student attending any institution for the first time at the undergraduate level. Includes students enrolled in academic or occupational programs. Also includes students enrolled in the fall term who attended college for the first time in the prior summer term, and students who entered with advanced standing (college credits earned before graduation from high school).}
  \item{Eearly decision}{A plan that allows students to apply and be notified of an admission decision (and financial aid offer, if applicable) well in advance of the regular notification date. Applicants agree to accept an offer of admission and, if admitted, to withdraw their applications from other colleges. There are three possible decision applications: admitted, denied, or not admitted but forwarded for consideration with the regular applicant pool, without prejudice.}
  \item{Early action}{An admission plan that allows students to apply and be notified of an admission decision well in advance of the regular notification dates. If admitted, the candidate is not committed to enroll (unlike early decision). Students may reply to the offer under the college's regular reply policy.}
  }
}
