\name{adm}
\alias{adm}

\title{Admissions and Test Scores: ADM}

\description{
This file contains information about the undergraduate selection process for entering first-time, degree/certificate-seeking students. This includes information about admission considerations,  applicants, applicants that were admitted, and admitted students who enrolled.
}

\section{SAT and ACT test}{
SAT and ACT test scores are included for institutions, that require test scores for admission.

These data are applicable for institutions that do not have an open admissions policy for entering first-time students.

Beginning with 2016, writing scores for both SAT and ACT are no longer collected.
}

\section{Admission Consideration}{
The possible values for the admission consideration variables have changed from data prior to 2016.  "Do not know" is no longer an option and "considered but not required" was added.  See frequencies for complete listing.
}


\section{Glossary}{
\describe{

\item{ACT (American college testing program)}{The ACT assessment program measures educational development and readiness to pursue college-level coursework in English, mathematics, natural science, and social studies. Student performance does not reflect innate ability and is influenced by a student's educational preparedness.}

\item{Admissions}{Applicants that have been granted an official offer to enroll in a college or university. }

\item{Admissions test scores}{Scores on standardized admissions tests or special admissions tests }

\item{Admissions total [ADMSSN]}{The number of first-time, degree/certificate-seeking undergraduate students who applied, were admitted, and enrolled (full or part time) at your institution for the most recent fall period available. Include early decision, early action, and students who began studies during the summer prior to that fall. }

\item{Applicant}{An individual who has fulfilled the institution’s requirements to be considered for admission (including payment or waiving of the application fee, if any) and who has been notified of one of the following actions: admission, nonadmission, placement on waiting list, or application withdrawn (by applicant or institution). }

\item{Applicants total [APPLCN]}{The number of first-time, degree/certificate-seeking undergraduate students who applied, were admitted, and enrolled (full or part time) at your institution for the most recent fall period available. Include early decision, early action, and students who began studies during the summer prior to that fall. }

\item{Degree/certificate-seeking students }{Students enrolled in courses for credit who are recognized by the institution as seeking a degree or formal award. At the undergraduate level, this is intended to include students enrolled in vocational or occupational programs. }

\item{Early action}{An admission plan that allows students to apply and be notified of an admission decision well in advance of the regular notification dates. If admitted, the candidate is not committed to enroll (unlike early decision). Students may reply to the offer under the college's regular reply policy.}

\item{Eearly decision}{A plan that allows students to apply and be notified of an admission decision (and financial aid offer, if applicable) well in advance of the regular notification date. Applicants agree to accept an offer of admission and, if admitted, to withdraw their applications from other colleges. There are three possible decision applications: admitted, denied, or not admitted but forwarded for consideration with the regular applicant pool, without prejudice. }

\item{Enrolled total [ENRLT]}{The number of first-time, degree/certificate-seeking undergraduate students who applied, were admitted, and enrolled (full or part time) at your institution for the most recent fall period available. Include early decision, early action, and students who began studies during the summer prior to that fall. }

\item{First-time student (Undergraduate)}{A student attending any institution for the first time at the undergraduate level. Includes students enrolled in academic or occupational programs. Also includes students enrolled in the fall term who attended college for the first time in the prior summer term, and students who entered with advanced standing (college credits earned before graduation from high school). }

\item{Full-time student Undergraduate}{A student enrolled for 12 or more semester credits, or 12 or more quarter credits, or 24 or more contact hours a week each term.}

\item{Part-time student Undergraduate}{A student enrolled for either 11 semester credits or less, or 11 quarter credits or less, or less than 24 contact hours a week each term.}

\item{SAT (Scholastic aptitude test)}{An examination administered by the Educational Testing Service and used to predict the facility with which an individual will progress in learning college-level academic subjects. }

\item{Test Of English as a Foreign Language (TOEFL)}{Standardized test designed to determine an applicant's ability to benefit from instruction in English.}
  }
}
