\name{al}
\alias{al}

\title{Academic Library: AL}

\description{
The Academic Library survey became part of the Integrated Postsecondary Education Data system in collection year 2014-15. Data include characteristics of the library, collections, expenditures, and services.
}

\section{Data Availability}{
These data are only applicable for degree-granting institutions.  Degree-granting institutions with total expenditures over $100,000 dollars will have expenditure data.
In 2016-17 the types of collections was expanded to collect information on physical and electronic serials.
}


\section{Glossary}{
  \describe{
\item{Branch and independent libraries}{Defined as auxiliary library service outlets with quarters separate from the central library of an institution, which have a basic collection of books and other materials, a regular staffing level, and an established schedule. Include virtual/digital-based branch and independent libraries.}

\item{Circulations}{The number of times librart units are checked out from the general and reserve collections. Includes initial transactions and renewals.}

\item{Database}{Collection of electronically stored data or unit records (facts, bibliographic data, and texts) with a common user interface and software for the retrieval and manipulation of the data. The data or records are usually collected with a particular intent and relate to a defined topic.}

\item{E-books}{Digital documents (including those digitized by the library), licensed or not, where searchable text is prevalent, and which can be seen in analogy to a printed book (monograph).}

\item{E-serial }{A periodical publication that is published in digital form to be displayed on a computer screen.}

\item{Fringe benefits}{Cash contributions in the form of supplementary or deferred compensation other than salary. Employee fringe benefits include retirement plans, social security taxes, medical/dental plans, unemployment compensation plans, group life insurance plans, worker's compensation plans, and other benefits in-kind with cash options.}

\item{Interlibrary loan}{The process by which a library requests material from, or supplies material to, another library" where "'material' includes books, audiovisual materials, and other returnable items as well as copies of journal articles, book chapters, excerpts, and other non-returnable items.}

\item{Library collections}{Comprise of documents held locally and remote resources for which permanentor temporary access rights have been acquired.  Access rights may be acquired by the library itself, by a consortium and/or through external funding.  Acquisition is to be understood as securing rights and including it on the OPAC (online public access catalog) or other databases of the library.}

\item{Number of digital/electronic books [LEBOOKS]}{The number of digital/electronic book units held locally and remote units for which permanent or temporary access rights have been acquired. E-books that are available as part of a database are excluded.  Include non-serial government documents.}

\item{Number of digital/electronic media [LEMEDIA]}{The number of units of downloadable media materials featuring video, graphics, or sound, including streaming media and graphic materials that the library has selected as part of its collection. Image databases (ARTStor, AP Photo Archives) are excluded.}

\item{Number of electronic serials [LESERIA]}{The number of e-serial titles that are accessible through the library’s catalog or discovery system.  Included are open access (OA) titles if the individual titles are searchable through the library’s catalog or discovery system, except for e-serial titles from HathiTrust, Center for Research Lirbraries, Internet Archive, and similar collections unless the library owns the digitized item and it is accessible under current copyright law.}

\item{Number of physical books [LPBOOKS]}{The number of volumes held in the library collection. }

\item{Number of physical media [LPMEDIA]}{The number of units of media materials  that are displayed by visual projection or magnification, or through sound reproduction, or both, including sound recordings, motion pictures and video recordings, and graphic materials. Microforms are excluded. Prior to fiscal year 2014, media materials were not categorized as physical or digital materials separately.}

\item{Number of physical serials [LPSERIA]}{The number of physical serial titles that are accessible through the library’s catalog or discovery system. Included are only serial titles, not subscriptions.}

\item{One-time purchases of books, serial backfiles, and other materials [LEXMSBB]}{The cost of one-time purchases of books, serial backfiles, and other materials. Includes expenditures for published materials in all formats including archives and special collections, except current subscriptions to serials.}

\item{Ongoing commitments to subscriptions [LEXMSCS]}{The expenditures for ongoing subscriptions to serials in all formats, including duplicates, for all outlets. These are publications issued in successive parts, usually at regular intervals, and, as a rule, intended to be continued indefinitely. Serial subscriptions include periodicals, newspapers, annuals (reports, yearbooks, etc.), memoirs, proceedings, and transactions of societies. Includes the costs of electronic serials bought in aggregations and serial packages.}

\item{Other materials/services expenditures [LEXMSOT]}{This field includes all other materials/services expenditures that are not one-time purchases of materials or ongoing subscriptions. Includes document/interlibrary loan services expenses and other expenditures for information resources.}

\item{Other operation and maintenance expenditures [LEXOMOT]}{This field includes all other operation and maintenance expenditures that are not preservation services. Includes computer hardware, software, and equipment expenses; national, regional, and local bibliographic utilities, networks and consortia fees; and other expenses not reported relating to maintenence.}

\item{Preservation services [LEXOMPS]}{The total expenditures associated with maintaining library and archival materials for use either in their original physical form or in some other usable way. This includes but is not limited to binding and rebinding, materials conservation, de-acidification, lamination, and restoration. Also, include preservation-related contracts for services (e.g. digitization).}

\item{Serial}{A publication in any medium issued in successive parts bearing numerical or chronological designations and intended to be continued indefinitely. This includes, in any physical format, periodicals, newspapers, and annuals (reports, yearbooks, etc.); the journals, memoirs, proceedings, transactions, etc. of societies; and numbered monographic series.}

\item{Volume}{A single physical unit of any printed, typewritten, handwritten, mimeographed, or processed work, distinguished from other units by a separate binding, encasement, portfolio, or other clear distinction, which has been cataloged, classified, and made ready for use, and which is typically the unit used to charge circulation transactions.}
    }
  }
