\name{al}
\alias{al}

\title{Academic Library: AL}

\description{
The Academic Library survey became part of the Integrated Postsecondary Education Data system in collection year 2014-15. Data include characteristics of the library, collections, expenditures, and services
}

\section{Data Availability}{
These data are only applicable for degree-granting institutions.  Degree-granting institutions with total expenditures over $100,000 dollars will have expenditure data.
In 2016-17 the types of collections was expanded to collect information on physical and electronic serials.
}

\section{Glossary}{
  \describe{
    \item{Total library circulations}{
    The number of times book and media units are "checked out" from the general and reserve collections. Includes intitial transactions and renewals.
    }
    \item{Ongoing commitments to subscriptions}{
    The expenditures for ongoing subscriptions to serials in all formats, including duplicates, for all outlets. These are publications issued in successive parts, usually at regular intervals, and, as a rule, intended to be continued indefinitely.
    Serial subscriptions include periodicals, newspapers, annuals (reports, yearbooks, etc.), memoirs, proceedings, and transactions of societies. Includes the costs of electronic serials bought in aggregations and serial packages.
    }
  }
}
