\name{c_a}
\alias{c_a}

\title{Completions: C_A}

\description{
This file contains the number of awards by type of program, level of award (certificate or degree), first or second major and by race/ethnicity and gender.
}

\section{Time Period}{
Data covers all awards granted between July 1 of previous year and June 30 of current year.
}

\section{Type of Instructional Program}{
Type of program is categorized according to the Classification of Instructional Programs (CIP), a detailed coding system for postsecondary instructional programs.
}

\section{Data Layout}{
This file contains multiple records per institution. Each record is uniquely defined by the variables IPEDS ID (UNITID), classification of instructional program (CIPCODE), first or second major (MAJORNUM) and award level (AWLEVEL).  Each record will contain the total awards,  awards for men and women, and the total awards and awards for men and women for all nine race/ethnicity categories.
}

\section{Data Releases}{
Preliminary release data have been edited but are subject to further NCES quality control procedures. Imputed data for nonresponding institutions are not included. These data are used for First Look (Preliminary Data) publications.

Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions. These data are used for First Look (Provisional Data) publications.

Final release data include revisions to the provisional release data that have been made by institutions during the subsequent data collection year. The final release data can be used when the most up to date data are required; however, these data may not match tables from the First Look reports based on preliminary and provisional data.
}

\section{Glossary}{
  \describe{
  \item{Classification of instrucional Program (CIP) code}{A six-digit code in the form xx.xxxx that identifies instructional  program specialties within educational institutions.}
  }
}
