\name{c_b}
\alias{c_b}

\title{Completions: C_B}

\description{
This file contains the number of students who completed any degree or certificate by race/ethnicity and gender.
}

\section{Time Period}{
Data covers all awards granted between July 1 of previous year and June 30 of current year.
}

\section{Data Layout}{
This file contains one record per institution. Each record is uniquely defined by the variables IPEDS ID (UNITID).  Each record will contain the total number of students,  men and women, and the total number of students for men and women for all nine race/ethnicity categories.
}

\section{Data Releases}{
Preliminary release data have been edited but are subject to further NCES quality control procedures. Imputed data for nonresponding institutions are not included. These data are used for First Look (Preliminary Data) publications.

Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions. These data are used for First Look (Provisional Data) publications.

Final release data include revisions to the provisional release data that have been made by institutions during the subsequent data collection year. The final release data can be used when the most up to date data are required; however, these data may not match tables from the First Look reports based on preliminary and provisional data.
}

