\name{c_c}
\alias{c_c}

\title{Completions: C_C}

\description{
This file contains the number of students receiving a degree or certificate by level of award and by race/ethnicity,  gender and age categories.
}

\section{Time Period}{
Data covers all awards granted between July 1 of previous year and June 30 of current year.
}

\section{Data Layout}{
This file contains multiple records per institution. Each record is uniquely defined by the variables IPEDS ID (UNITID) and award level (AWLEVELC).  Each record will contain the total number of students receiving awards;  number of  men and women receiving awards: number of students receiving awards for all nine race/ethnicity categories; number of students receiving awards are also available for the following age groups: under 18, 18-24, 25-39, 40 and above.

Numbers are not available for combinations of race, gender and age.
}
