\name{eap}
\alias{eap}

\title{Employees by Assigned Position: EAP}

\description{
This file contains the number of staff classified by full- and part-time status, medical school status, occupational category and faculty and tenure status for applicable occupations.
}

\section{Tenure}{
Faculty and tenure status categories include tenured, on-tenured-track, not-on-tenure track/no tenure system and without faculty status. Staff that are not on tenure track are further disaggregated by contract length (multi-year, indefinite, annual and less-than-annual).
}

\section{Contracts}{
The multi-year and indefinite contracts were new categories beginning in 2016. From 2012 through 2015 the multi-year and indefinite contract lengths were combined as one category and often referred to as multi-year in past documentation.
}

\section{Data Layout}{
This data file has multiple records for each institution.   Each record is uniquely defined by the variables IPEDS ID (UNITID), and the variable EAPCAT which defines occupational category and tenure status of staff.
}

\section{Data Releases}{
Preliminary release data have been edited but are subject to further NCES quality control procedures. Imputed data for nonresponding institutions are not included. These data are used for First Look (Preliminary Data) publications.

Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions. These data are used for First Look (Provisional Data) publications.

Final release data include revisions to the provisional release data that have been made by institutions during the subsequent data collection year. The final release data can be used when the most up to date data are required; however, these data may not match tables from the First Look reports based on preliminary and provisional data.
}

\section{Glossary}{
\describe{
  \item{Faculty Persons}{
  Faculty Persons identified by the institution as such and typically those whose initial assignments are made for the purpose of conducting instruction, research or public service as a principal activity (or activities). They may hold academic rank titles of professor, associate professor, assistant professor, instructor, lecturer or the equivalent of any of those academic ranks. Faculty may also include the chancellor/president, provost, vice provosts, deans, directors or the equivalent, as well as associate deans, assistant deans and executive officers of academic departments (chairpersons, heads or the equivalent) if their principal activity is instruction combined with research and/or public service. The designation as "faculty" is separate from the activities to which they may be currently assigned. For example, a newly appointed president of an institution may also be appointed as a faculty member. Graduate, instruction, and research assistants are not included in this category.
  }
  \item{Tenure Status}{
  \itemize{
    \item Tenure - Status of a personnel position with respect to permanence of the position.
    \item Tenure track - Personnel positions that lead to consideration for tenure.
    \item Not on tenure track - Personnel positions that are considered non-tenure earning positions.
    }
  }
  \item{Multi-year contract}{
  A contract or employment agreement that is in effect for more than one year (e.g., more than 365 days) where subsequent years of the contract may be contingent upon other factors such as the appropriation of funds. The renewal period of a multi-year contract is not on an annual basis (e.g., a 5-year multi-year contract is renewed every 5 years NOT annually).
  }
  \item{Annual contract}{
  An annually-renewable contract or employment agreement that is in effect for a stated annual period within one year of execution, and may be equal to a period of 365 days, or a standard academic year, or the equivalent. Does not include contracts for partial year periods such as a single semester, quarter, term, block, or course.
  }
  \item{Less-than-annual-contract}{
  A contract or employment agreement that is in effect for a partial year period of less than 365 days, or less than a standard academic year or the equivalent. Includes contracts for partial year periods such as a single semester, quarter, term, block, or course.
  }
  }
}

\section{More Info}{
\href{http://www.bls.gov/soc/}{Standard Occupational Classification (SOC) system}

The IPEDS HR survey component uses the Standard Occupational Classification (SOC) system to assist institutions in classifying employees for IPEDS reporting purposes.

The SOC is a system used by Federal statistical agencies to classify workers and jobs into occupational categories for the purpose of collecting, calculating, analyzing, or disseminating data. In 2018, an updated version of the SOC system (click the link above) was released. Consequently, the IPEDS HR survey was revised to incorporate the 2018 SOC information. The new IPEDS occupational categories are effective with 2018-19 Human Resources reporting.}

