\name{efc}
\alias{efc}

\title{Fall Enrollment: EFC}

\description{
This file contains the number of first-time freshmen by state of residence, along with data on the number who graduated from high school the previous year.
}

\section{Reporting}{
 Institutions with traditional academic year calendar systems (semester, quarter, trimester or  4-1-4) report their enrollment as of October 15 or the official fall reporting date of the institution.

 Institutions with calendar systems that differ by program or allow continuous enrollment report students that are enrolled at any time between August 1 and October 31.
}

\section{Data Layout}{
This file has multiple records per institution.  Records are uniquely defined by the variables IPEDS ID (UNITID), and the state of residence (EFCSTATE).
}

\section{Optional Data: Enrollment by Residency}{
Submission of enrollment of first-time undergraduate students by residency is mandatory in even-numbered years and optional in odd-numbered years. Do not use these data in odd-numbered years to produce national or state estimates as data will not be available from all institutions. Data in optional years are provided to allow users to compare institutional level data among reporting institutions.
}

\section{Data Releases}{
Preliminary release data have been edited but are subject to further NCES quality control procedures. Imputed data for nonresponding institutions are not included. These data are used for First Look (Preliminary Data) publications.

Provisional release data have undergone all NCES data quality control procedures. Data are imputed for nonresponding institutions. These data are used for First Look (Provisional Data) publications.

Final release data include revisions to the provisional release data that have been made by institutions during the subsequent data collection year. The final release data can be used when the most up to date data are required; however, these data may not match tables from the First Look reports based on preliminary and provisional data.
}

\section{Glossary}{
\describe{
  \item{First-time student (undergraduate)}{
  A student who has no prior postsecondary experience (except as noted below) attending any institution for the first time at the undergraduate level. This includes students enrolled in academic or occupational programs. It also includes students enrolled in the fall term who attended college for the first time in the prior summer term, and students who entered with advanced standing (college credits earned before graduation from high school).
  }
  }
}
